Create a tibble with averaged high-dimensional data

This function calculates the average values of high-dimensional data within each hexagonal bin.

Usage

avg_highd_data(highd_data, scaled_nldr_hexid)

Arguments

highd_data

A tibble that contains the high-dimensional data.

scaled_nldr_hexid

A tibble that contains the scaled embedding with hexagonal bin IDs.

Value

A tibble with the average values of the high-dimensional data within each hexagonal bin.

Examples

umap_with_hb_id <- scurve_model_obj$hb_obj$data_hb_id
avg_highd_data(highd_data = scurve, scaled_nldr_hexid = umap_with_hb_id)
#> # A tibble: 142 × 8
#>        h      x1    x2    x3         x4        x5       x6        x7
#>    <int>   <dbl> <dbl> <dbl>      <dbl>     <dbl>    <dbl>     <dbl>
#>  1    21 -0.992   1.91 1.11  -0.000427   0.000624  0.00749  0.00105 
#>  2    22 -0.906   1.93 1.41  -0.0000183  0.00331  -0.0204  -0.000363
#>  3    23 -0.680   1.93 1.72  -0.000810  -0.00259  -0.00449  0.00153 
#>  4    24 -0.272   1.93 1.96   0.00251    0.00668  -0.0460   0.00128 
#>  5    25  0.0760  1.93 2.00   0.00876    0.00447   0.00851 -0.00195 
#>  6    26  0.461   1.93 1.89  -0.00478    0.00492   0.00835  0.00172 
#>  7    27  0.719   1.99 1.70   0.0109    -0.00349  -0.0297  -0.00223 
#>  8    36 -0.985   1.75 0.853 -0.00202    0.000397  0.00331  0.000338
#>  9    37 -0.980   1.66 1.17  -0.000374  -0.00154   0.0165   0.000126
#> 10    38 -0.821   1.64 1.56  -0.000459   0.000538 -0.0123   0.000780
#> # ℹ 132 more rows