Compute Weights for Hexagonal Binning

This function computes weights for hexagonal binning based on the average values of each bin and the distances from these averages.

Usage

compute_weights(nldr_df_with_hex_id)

Arguments

nldr_df_with_hex_id

A data frame with 2D embeddings and hexagonal bin IDs.

Value

A data frame with weights calculated for each hexagonal bin.

Examples

num_bins_x <- calculate_effective_x_bins(nldr_df = s_curve_noise_umap_scaled,
x = "UMAP1", hex_size = NA, buffer_x = NA)
#> Hex size is set to 0.2.
#> Buffer along the x-axis is set to 0.519615242270663.
num_bins_y <- calculate_effective_y_bins(nldr_df = s_curve_noise_umap_scaled,
 y = "UMAP2", hex_size = NA, buffer_y = NA)
#> Hex size is set to 0.2.
#> Buffer along the y-axis is set to 0.45.
hex_bin_obj <- generate_hex_binning_info(nldr_df = s_curve_noise_umap_scaled,
x = "UMAP1", y = "UMAP2", num_bins_x = num_bins_x,
num_bins_y = num_bins_y, x_start = NA, y_start = NA, buffer_x = NA,
buffer_y = NA, hex_size = NA)
#> Hex size is set to 0.2.
#> x_start is set to 0.
#> y_start is set to 0.
#> Hex size is set to 0.2.
UMAP_data_with_hb_id <- as.data.frame(do.call(cbind, hex_bin_obj$nldr_data_with_hex_id))
compute_weights(nldr_df_with_hex_id = UMAP_data_with_hb_id)
#> $hb_id
#>   [1]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#>  [26] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#>  [51] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#>  [76]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [101] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [126] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [151]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [176] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [201] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [226]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [251] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [276] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [301]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [326] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [351] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [376]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [401] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [426] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [451]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [476] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [501] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [526]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [551] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [576] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [601]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [626] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [651] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [676]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [701] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [726] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [751]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [776] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [801] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> [826]  1  1  1  1  1  1  2  2  2  2  6  6  6  6  6  6  6  6  6  6  6  6  6  7  7
#> [851] 12 12 12 13 13 13 13 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 23 27
#> [876] 28 28 28 28 28 28 28 28 28 28 28 28 28 28 28 33 33 33 33 33 33 33 34 34 34
#> 
#> $UMAP1
#>   [1] 0.16719939 0.16150976 0.09405763 0.10117880 0.13459287 0.16301062
#>   [7] 0.29256853 0.23125328 0.25450458 0.29499615 0.08044271 0.26292782
#>  [13] 0.05182936 0.07223295 0.13518498 0.08900367 0.28575187 0.21053454
#>  [19] 0.03170275 0.00000000 0.04209521 0.05349536 0.33548077 0.38392618
#>  [25] 0.39315580 0.46806684 0.50932690 0.49582819 0.62693057 0.64241759
#>  [31] 0.64916340 0.60311661 0.57541142 0.62412077 0.64842247 0.68677026
#>  [37] 0.73857864 0.83992020 0.73364820 0.80998347 0.71326033 0.76738169
#>  [43] 0.82638316 0.84482412 0.79995094 0.77828853 0.72681720 0.72192354
#>  [49] 0.79296306 0.68029522 0.90342138 0.96228699 0.80783543 0.80668446
#>  [55] 0.76427076 0.71265285 0.69339205 0.76831122 0.99318862 0.96227596
#>  [61] 0.94650462 0.88588122 0.92537063 0.80964667 0.91053035 0.83820653
#>  [67] 0.77510396 0.82904448 0.81275032 0.75504949 0.69843891 0.74175661
#>  [73] 0.96450414 0.94245034 1.00000000 0.16719939 0.16150976 0.09405763
#>  [79] 0.10117880 0.13459287 0.16301062 0.29256853 0.23125328 0.25450458
#>  [85] 0.29499615 0.08044271 0.26292782 0.05182936 0.07223295 0.13518498
#>  [91] 0.08900367 0.28575187 0.21053454 0.03170275 0.00000000 0.04209521
#>  [97] 0.05349536 0.33548077 0.38392618 0.39315580 0.46806684 0.50932690
#> [103] 0.49582819 0.62693057 0.64241759 0.64916340 0.60311661 0.57541142
#> [109] 0.62412077 0.64842247 0.68677026 0.73857864 0.83992020 0.73364820
#> [115] 0.80998347 0.71326033 0.76738169 0.82638316 0.84482412 0.79995094
#> [121] 0.77828853 0.72681720 0.72192354 0.79296306 0.68029522 0.90342138
#> [127] 0.96228699 0.80783543 0.80668446 0.76427076 0.71265285 0.69339205
#> [133] 0.76831122 0.99318862 0.96227596 0.94650462 0.88588122 0.92537063
#> [139] 0.80964667 0.91053035 0.83820653 0.77510396 0.82904448 0.81275032
#> [145] 0.75504949 0.69843891 0.74175661 0.96450414 0.94245034 1.00000000
#> [151] 0.16719939 0.16150976 0.09405763 0.10117880 0.13459287 0.16301062
#> [157] 0.29256853 0.23125328 0.25450458 0.29499615 0.08044271 0.26292782
#> [163] 0.05182936 0.07223295 0.13518498 0.08900367 0.28575187 0.21053454
#> [169] 0.03170275 0.00000000 0.04209521 0.05349536 0.33548077 0.38392618
#> [175] 0.39315580 0.46806684 0.50932690 0.49582819 0.62693057 0.64241759
#> [181] 0.64916340 0.60311661 0.57541142 0.62412077 0.64842247 0.68677026
#> [187] 0.73857864 0.83992020 0.73364820 0.80998347 0.71326033 0.76738169
#> [193] 0.82638316 0.84482412 0.79995094 0.77828853 0.72681720 0.72192354
#> [199] 0.79296306 0.68029522 0.90342138 0.96228699 0.80783543 0.80668446
#> [205] 0.76427076 0.71265285 0.69339205 0.76831122 0.99318862 0.96227596
#> [211] 0.94650462 0.88588122 0.92537063 0.80964667 0.91053035 0.83820653
#> [217] 0.77510396 0.82904448 0.81275032 0.75504949 0.69843891 0.74175661
#> [223] 0.96450414 0.94245034 1.00000000 0.16719939 0.16150976 0.09405763
#> [229] 0.10117880 0.13459287 0.16301062 0.29256853 0.23125328 0.25450458
#> [235] 0.29499615 0.08044271 0.26292782 0.05182936 0.07223295 0.13518498
#> [241] 0.08900367 0.28575187 0.21053454 0.03170275 0.00000000 0.04209521
#> [247] 0.05349536 0.33548077 0.38392618 0.39315580 0.46806684 0.50932690
#> [253] 0.49582819 0.62693057 0.64241759 0.64916340 0.60311661 0.57541142
#> [259] 0.62412077 0.64842247 0.68677026 0.73857864 0.83992020 0.73364820
#> [265] 0.80998347 0.71326033 0.76738169 0.82638316 0.84482412 0.79995094
#> [271] 0.77828853 0.72681720 0.72192354 0.79296306 0.68029522 0.90342138
#> [277] 0.96228699 0.80783543 0.80668446 0.76427076 0.71265285 0.69339205
#> [283] 0.76831122 0.99318862 0.96227596 0.94650462 0.88588122 0.92537063
#> [289] 0.80964667 0.91053035 0.83820653 0.77510396 0.82904448 0.81275032
#> [295] 0.75504949 0.69843891 0.74175661 0.96450414 0.94245034 1.00000000
#> [301] 0.16719939 0.16150976 0.09405763 0.10117880 0.13459287 0.16301062
#> [307] 0.29256853 0.23125328 0.25450458 0.29499615 0.08044271 0.26292782
#> [313] 0.05182936 0.07223295 0.13518498 0.08900367 0.28575187 0.21053454
#> [319] 0.03170275 0.00000000 0.04209521 0.05349536 0.33548077 0.38392618
#> [325] 0.39315580 0.46806684 0.50932690 0.49582819 0.62693057 0.64241759
#> [331] 0.64916340 0.60311661 0.57541142 0.62412077 0.64842247 0.68677026
#> [337] 0.73857864 0.83992020 0.73364820 0.80998347 0.71326033 0.76738169
#> [343] 0.82638316 0.84482412 0.79995094 0.77828853 0.72681720 0.72192354
#> [349] 0.79296306 0.68029522 0.90342138 0.96228699 0.80783543 0.80668446
#> [355] 0.76427076 0.71265285 0.69339205 0.76831122 0.99318862 0.96227596
#> [361] 0.94650462 0.88588122 0.92537063 0.80964667 0.91053035 0.83820653
#> [367] 0.77510396 0.82904448 0.81275032 0.75504949 0.69843891 0.74175661
#> [373] 0.96450414 0.94245034 1.00000000 0.16719939 0.16150976 0.09405763
#> [379] 0.10117880 0.13459287 0.16301062 0.29256853 0.23125328 0.25450458
#> [385] 0.29499615 0.08044271 0.26292782 0.05182936 0.07223295 0.13518498
#> [391] 0.08900367 0.28575187 0.21053454 0.03170275 0.00000000 0.04209521
#> [397] 0.05349536 0.33548077 0.38392618 0.39315580 0.46806684 0.50932690
#> [403] 0.49582819 0.62693057 0.64241759 0.64916340 0.60311661 0.57541142
#> [409] 0.62412077 0.64842247 0.68677026 0.73857864 0.83992020 0.73364820
#> [415] 0.80998347 0.71326033 0.76738169 0.82638316 0.84482412 0.79995094
#> [421] 0.77828853 0.72681720 0.72192354 0.79296306 0.68029522 0.90342138
#> [427] 0.96228699 0.80783543 0.80668446 0.76427076 0.71265285 0.69339205
#> [433] 0.76831122 0.99318862 0.96227596 0.94650462 0.88588122 0.92537063
#> [439] 0.80964667 0.91053035 0.83820653 0.77510396 0.82904448 0.81275032
#> [445] 0.75504949 0.69843891 0.74175661 0.96450414 0.94245034 1.00000000
#> [451] 0.16719939 0.16150976 0.09405763 0.10117880 0.13459287 0.16301062
#> [457] 0.29256853 0.23125328 0.25450458 0.29499615 0.08044271 0.26292782
#> [463] 0.05182936 0.07223295 0.13518498 0.08900367 0.28575187 0.21053454
#> [469] 0.03170275 0.00000000 0.04209521 0.05349536 0.33548077 0.38392618
#> [475] 0.39315580 0.46806684 0.50932690 0.49582819 0.62693057 0.64241759
#> [481] 0.64916340 0.60311661 0.57541142 0.62412077 0.64842247 0.68677026
#> [487] 0.73857864 0.83992020 0.73364820 0.80998347 0.71326033 0.76738169
#> [493] 0.82638316 0.84482412 0.79995094 0.77828853 0.72681720 0.72192354
#> [499] 0.79296306 0.68029522 0.90342138 0.96228699 0.80783543 0.80668446
#> [505] 0.76427076 0.71265285 0.69339205 0.76831122 0.99318862 0.96227596
#> [511] 0.94650462 0.88588122 0.92537063 0.80964667 0.91053035 0.83820653
#> [517] 0.77510396 0.82904448 0.81275032 0.75504949 0.69843891 0.74175661
#> [523] 0.96450414 0.94245034 1.00000000 0.16719939 0.16150976 0.09405763
#> [529] 0.10117880 0.13459287 0.16301062 0.29256853 0.23125328 0.25450458
#> [535] 0.29499615 0.08044271 0.26292782 0.05182936 0.07223295 0.13518498
#> [541] 0.08900367 0.28575187 0.21053454 0.03170275 0.00000000 0.04209521
#> [547] 0.05349536 0.33548077 0.38392618 0.39315580 0.46806684 0.50932690
#> [553] 0.49582819 0.62693057 0.64241759 0.64916340 0.60311661 0.57541142
#> [559] 0.62412077 0.64842247 0.68677026 0.73857864 0.83992020 0.73364820
#> [565] 0.80998347 0.71326033 0.76738169 0.82638316 0.84482412 0.79995094
#> [571] 0.77828853 0.72681720 0.72192354 0.79296306 0.68029522 0.90342138
#> [577] 0.96228699 0.80783543 0.80668446 0.76427076 0.71265285 0.69339205
#> [583] 0.76831122 0.99318862 0.96227596 0.94650462 0.88588122 0.92537063
#> [589] 0.80964667 0.91053035 0.83820653 0.77510396 0.82904448 0.81275032
#> [595] 0.75504949 0.69843891 0.74175661 0.96450414 0.94245034 1.00000000
#> [601] 0.16719939 0.16150976 0.09405763 0.10117880 0.13459287 0.16301062
#> [607] 0.29256853 0.23125328 0.25450458 0.29499615 0.08044271 0.26292782
#> [613] 0.05182936 0.07223295 0.13518498 0.08900367 0.28575187 0.21053454
#> [619] 0.03170275 0.00000000 0.04209521 0.05349536 0.33548077 0.38392618
#> [625] 0.39315580 0.46806684 0.50932690 0.49582819 0.62693057 0.64241759
#> [631] 0.64916340 0.60311661 0.57541142 0.62412077 0.64842247 0.68677026
#> [637] 0.73857864 0.83992020 0.73364820 0.80998347 0.71326033 0.76738169
#> [643] 0.82638316 0.84482412 0.79995094 0.77828853 0.72681720 0.72192354
#> [649] 0.79296306 0.68029522 0.90342138 0.96228699 0.80783543 0.80668446
#> [655] 0.76427076 0.71265285 0.69339205 0.76831122 0.99318862 0.96227596
#> [661] 0.94650462 0.88588122 0.92537063 0.80964667 0.91053035 0.83820653
#> [667] 0.77510396 0.82904448 0.81275032 0.75504949 0.69843891 0.74175661
#> [673] 0.96450414 0.94245034 1.00000000 0.16719939 0.16150976 0.09405763
#> [679] 0.10117880 0.13459287 0.16301062 0.29256853 0.23125328 0.25450458
#> [685] 0.29499615 0.08044271 0.26292782 0.05182936 0.07223295 0.13518498
#> [691] 0.08900367 0.28575187 0.21053454 0.03170275 0.00000000 0.04209521
#> [697] 0.05349536 0.33548077 0.38392618 0.39315580 0.46806684 0.50932690
#> [703] 0.49582819 0.62693057 0.64241759 0.64916340 0.60311661 0.57541142
#> [709] 0.62412077 0.64842247 0.68677026 0.73857864 0.83992020 0.73364820
#> [715] 0.80998347 0.71326033 0.76738169 0.82638316 0.84482412 0.79995094
#> [721] 0.77828853 0.72681720 0.72192354 0.79296306 0.68029522 0.90342138
#> [727] 0.96228699 0.80783543 0.80668446 0.76427076 0.71265285 0.69339205
#> [733] 0.76831122 0.99318862 0.96227596 0.94650462 0.88588122 0.92537063
#> [739] 0.80964667 0.91053035 0.83820653 0.77510396 0.82904448 0.81275032
#> [745] 0.75504949 0.69843891 0.74175661 0.96450414 0.94245034 1.00000000
#> [751] 0.16719939 0.16150976 0.09405763 0.10117880 0.13459287 0.16301062
#> [757] 0.29256853 0.23125328 0.25450458 0.29499615 0.08044271 0.26292782
#> [763] 0.05182936 0.07223295 0.13518498 0.08900367 0.28575187 0.21053454
#> [769] 0.03170275 0.00000000 0.04209521 0.05349536 0.33548077 0.38392618
#> [775] 0.39315580 0.46806684 0.50932690 0.49582819 0.62693057 0.64241759
#> [781] 0.64916340 0.60311661 0.57541142 0.62412077 0.64842247 0.68677026
#> [787] 0.73857864 0.83992020 0.73364820 0.80998347 0.71326033 0.76738169
#> [793] 0.82638316 0.84482412 0.79995094 0.77828853 0.72681720 0.72192354
#> [799] 0.79296306 0.68029522 0.90342138 0.96228699 0.80783543 0.80668446
#> [805] 0.76427076 0.71265285 0.69339205 0.76831122 0.99318862 0.96227596
#> [811] 0.94650462 0.88588122 0.92537063 0.80964667 0.91053035 0.83820653
#> [817] 0.77510396 0.82904448 0.81275032 0.75504949 0.69843891 0.74175661
#> [823] 0.96450414 0.94245034 1.00000000 0.16719939 0.16150976 0.09405763
#> [829] 0.10117880 0.13459287 0.16301062 0.29256853 0.23125328 0.25450458
#> [835] 0.29499615 0.08044271 0.26292782 0.05182936 0.07223295 0.13518498
#> [841] 0.08900367 0.28575187 0.21053454 0.03170275 0.00000000 0.04209521
#> [847] 0.05349536 0.33548077 0.38392618 0.39315580 0.46806684 0.50932690
#> [853] 0.49582819 0.62693057 0.64241759 0.64916340 0.60311661 0.57541142
#> [859] 0.62412077 0.64842247 0.68677026 0.73857864 0.83992020 0.73364820
#> [865] 0.80998347 0.71326033 0.76738169 0.82638316 0.84482412 0.79995094
#> [871] 0.77828853 0.72681720 0.72192354 0.79296306 0.68029522 0.90342138
#> [877] 0.96228699 0.80783543 0.80668446 0.76427076 0.71265285 0.69339205
#> [883] 0.76831122 0.99318862 0.96227596 0.94650462 0.88588122 0.92537063
#> [889] 0.80964667 0.91053035 0.83820653 0.77510396 0.82904448 0.81275032
#> [895] 0.75504949 0.69843891 0.74175661 0.96450414 0.94245034 1.00000000
#> 
#> $UMAP2
#>   [1] 0.037066861 0.009905107 0.058760856 0.050909321 0.000000000 0.030266486
#>   [7] 0.138015108 0.061042472 0.063074584 0.117297635 0.274498327 0.341734583
#>  [13] 0.320865388 0.218331864 0.335559538 0.252258947 0.202379969 0.151933039
#>  [19] 0.303630086 0.271778584 0.257916294 0.187982579 0.278688714 0.390190144
#>  [25] 0.328235991 0.482713207 0.617404710 0.604985028 0.618409921 0.631646064
#>  [31] 0.593427000 0.597381477 0.708486670 0.712391079 0.661926514 0.778563498
#>  [37] 0.859778569 0.929412446 0.833935370 0.868291303 0.797282390 1.018811832
#>  [43] 1.011632238 0.983163036 0.964486583 0.839937252 0.931757885 0.952956061
#>  [49] 1.100876246 1.508729997 1.600664528 1.546732900 1.374772146 1.381019256
#>  [55] 1.552630670 1.539719959 1.448210796 1.497157149 1.521102615 1.546125959
#>  [61] 1.524113512 1.408512652 1.455380663 1.529334517 1.445460921 1.720961652
#>  [67] 1.730112607 1.722765733 1.732050808 1.670893354 1.621642749 1.681926517
#>  [73] 1.665355817 1.680924200 1.625538768 0.037066861 0.009905107 0.058760856
#>  [79] 0.050909321 0.000000000 0.030266486 0.138015108 0.061042472 0.063074584
#>  [85] 0.117297635 0.274498327 0.341734583 0.320865388 0.218331864 0.335559538
#>  [91] 0.252258947 0.202379969 0.151933039 0.303630086 0.271778584 0.257916294
#>  [97] 0.187982579 0.278688714 0.390190144 0.328235991 0.482713207 0.617404710
#> [103] 0.604985028 0.618409921 0.631646064 0.593427000 0.597381477 0.708486670
#> [109] 0.712391079 0.661926514 0.778563498 0.859778569 0.929412446 0.833935370
#> [115] 0.868291303 0.797282390 1.018811832 1.011632238 0.983163036 0.964486583
#> [121] 0.839937252 0.931757885 0.952956061 1.100876246 1.508729997 1.600664528
#> [127] 1.546732900 1.374772146 1.381019256 1.552630670 1.539719959 1.448210796
#> [133] 1.497157149 1.521102615 1.546125959 1.524113512 1.408512652 1.455380663
#> [139] 1.529334517 1.445460921 1.720961652 1.730112607 1.722765733 1.732050808
#> [145] 1.670893354 1.621642749 1.681926517 1.665355817 1.680924200 1.625538768
#> [151] 0.037066861 0.009905107 0.058760856 0.050909321 0.000000000 0.030266486
#> [157] 0.138015108 0.061042472 0.063074584 0.117297635 0.274498327 0.341734583
#> [163] 0.320865388 0.218331864 0.335559538 0.252258947 0.202379969 0.151933039
#> [169] 0.303630086 0.271778584 0.257916294 0.187982579 0.278688714 0.390190144
#> [175] 0.328235991 0.482713207 0.617404710 0.604985028 0.618409921 0.631646064
#> [181] 0.593427000 0.597381477 0.708486670 0.712391079 0.661926514 0.778563498
#> [187] 0.859778569 0.929412446 0.833935370 0.868291303 0.797282390 1.018811832
#> [193] 1.011632238 0.983163036 0.964486583 0.839937252 0.931757885 0.952956061
#> [199] 1.100876246 1.508729997 1.600664528 1.546732900 1.374772146 1.381019256
#> [205] 1.552630670 1.539719959 1.448210796 1.497157149 1.521102615 1.546125959
#> [211] 1.524113512 1.408512652 1.455380663 1.529334517 1.445460921 1.720961652
#> [217] 1.730112607 1.722765733 1.732050808 1.670893354 1.621642749 1.681926517
#> [223] 1.665355817 1.680924200 1.625538768 0.037066861 0.009905107 0.058760856
#> [229] 0.050909321 0.000000000 0.030266486 0.138015108 0.061042472 0.063074584
#> [235] 0.117297635 0.274498327 0.341734583 0.320865388 0.218331864 0.335559538
#> [241] 0.252258947 0.202379969 0.151933039 0.303630086 0.271778584 0.257916294
#> [247] 0.187982579 0.278688714 0.390190144 0.328235991 0.482713207 0.617404710
#> [253] 0.604985028 0.618409921 0.631646064 0.593427000 0.597381477 0.708486670
#> [259] 0.712391079 0.661926514 0.778563498 0.859778569 0.929412446 0.833935370
#> [265] 0.868291303 0.797282390 1.018811832 1.011632238 0.983163036 0.964486583
#> [271] 0.839937252 0.931757885 0.952956061 1.100876246 1.508729997 1.600664528
#> [277] 1.546732900 1.374772146 1.381019256 1.552630670 1.539719959 1.448210796
#> [283] 1.497157149 1.521102615 1.546125959 1.524113512 1.408512652 1.455380663
#> [289] 1.529334517 1.445460921 1.720961652 1.730112607 1.722765733 1.732050808
#> [295] 1.670893354 1.621642749 1.681926517 1.665355817 1.680924200 1.625538768
#> [301] 0.037066861 0.009905107 0.058760856 0.050909321 0.000000000 0.030266486
#> [307] 0.138015108 0.061042472 0.063074584 0.117297635 0.274498327 0.341734583
#> [313] 0.320865388 0.218331864 0.335559538 0.252258947 0.202379969 0.151933039
#> [319] 0.303630086 0.271778584 0.257916294 0.187982579 0.278688714 0.390190144
#> [325] 0.328235991 0.482713207 0.617404710 0.604985028 0.618409921 0.631646064
#> [331] 0.593427000 0.597381477 0.708486670 0.712391079 0.661926514 0.778563498
#> [337] 0.859778569 0.929412446 0.833935370 0.868291303 0.797282390 1.018811832
#> [343] 1.011632238 0.983163036 0.964486583 0.839937252 0.931757885 0.952956061
#> [349] 1.100876246 1.508729997 1.600664528 1.546732900 1.374772146 1.381019256
#> [355] 1.552630670 1.539719959 1.448210796 1.497157149 1.521102615 1.546125959
#> [361] 1.524113512 1.408512652 1.455380663 1.529334517 1.445460921 1.720961652
#> [367] 1.730112607 1.722765733 1.732050808 1.670893354 1.621642749 1.681926517
#> [373] 1.665355817 1.680924200 1.625538768 0.037066861 0.009905107 0.058760856
#> [379] 0.050909321 0.000000000 0.030266486 0.138015108 0.061042472 0.063074584
#> [385] 0.117297635 0.274498327 0.341734583 0.320865388 0.218331864 0.335559538
#> [391] 0.252258947 0.202379969 0.151933039 0.303630086 0.271778584 0.257916294
#> [397] 0.187982579 0.278688714 0.390190144 0.328235991 0.482713207 0.617404710
#> [403] 0.604985028 0.618409921 0.631646064 0.593427000 0.597381477 0.708486670
#> [409] 0.712391079 0.661926514 0.778563498 0.859778569 0.929412446 0.833935370
#> [415] 0.868291303 0.797282390 1.018811832 1.011632238 0.983163036 0.964486583
#> [421] 0.839937252 0.931757885 0.952956061 1.100876246 1.508729997 1.600664528
#> [427] 1.546732900 1.374772146 1.381019256 1.552630670 1.539719959 1.448210796
#> [433] 1.497157149 1.521102615 1.546125959 1.524113512 1.408512652 1.455380663
#> [439] 1.529334517 1.445460921 1.720961652 1.730112607 1.722765733 1.732050808
#> [445] 1.670893354 1.621642749 1.681926517 1.665355817 1.680924200 1.625538768
#> [451] 0.037066861 0.009905107 0.058760856 0.050909321 0.000000000 0.030266486
#> [457] 0.138015108 0.061042472 0.063074584 0.117297635 0.274498327 0.341734583
#> [463] 0.320865388 0.218331864 0.335559538 0.252258947 0.202379969 0.151933039
#> [469] 0.303630086 0.271778584 0.257916294 0.187982579 0.278688714 0.390190144
#> [475] 0.328235991 0.482713207 0.617404710 0.604985028 0.618409921 0.631646064
#> [481] 0.593427000 0.597381477 0.708486670 0.712391079 0.661926514 0.778563498
#> [487] 0.859778569 0.929412446 0.833935370 0.868291303 0.797282390 1.018811832
#> [493] 1.011632238 0.983163036 0.964486583 0.839937252 0.931757885 0.952956061
#> [499] 1.100876246 1.508729997 1.600664528 1.546732900 1.374772146 1.381019256
#> [505] 1.552630670 1.539719959 1.448210796 1.497157149 1.521102615 1.546125959
#> [511] 1.524113512 1.408512652 1.455380663 1.529334517 1.445460921 1.720961652
#> [517] 1.730112607 1.722765733 1.732050808 1.670893354 1.621642749 1.681926517
#> [523] 1.665355817 1.680924200 1.625538768 0.037066861 0.009905107 0.058760856
#> [529] 0.050909321 0.000000000 0.030266486 0.138015108 0.061042472 0.063074584
#> [535] 0.117297635 0.274498327 0.341734583 0.320865388 0.218331864 0.335559538
#> [541] 0.252258947 0.202379969 0.151933039 0.303630086 0.271778584 0.257916294
#> [547] 0.187982579 0.278688714 0.390190144 0.328235991 0.482713207 0.617404710
#> [553] 0.604985028 0.618409921 0.631646064 0.593427000 0.597381477 0.708486670
#> [559] 0.712391079 0.661926514 0.778563498 0.859778569 0.929412446 0.833935370
#> [565] 0.868291303 0.797282390 1.018811832 1.011632238 0.983163036 0.964486583
#> [571] 0.839937252 0.931757885 0.952956061 1.100876246 1.508729997 1.600664528
#> [577] 1.546732900 1.374772146 1.381019256 1.552630670 1.539719959 1.448210796
#> [583] 1.497157149 1.521102615 1.546125959 1.524113512 1.408512652 1.455380663
#> [589] 1.529334517 1.445460921 1.720961652 1.730112607 1.722765733 1.732050808
#> [595] 1.670893354 1.621642749 1.681926517 1.665355817 1.680924200 1.625538768
#> [601] 0.037066861 0.009905107 0.058760856 0.050909321 0.000000000 0.030266486
#> [607] 0.138015108 0.061042472 0.063074584 0.117297635 0.274498327 0.341734583
#> [613] 0.320865388 0.218331864 0.335559538 0.252258947 0.202379969 0.151933039
#> [619] 0.303630086 0.271778584 0.257916294 0.187982579 0.278688714 0.390190144
#> [625] 0.328235991 0.482713207 0.617404710 0.604985028 0.618409921 0.631646064
#> [631] 0.593427000 0.597381477 0.708486670 0.712391079 0.661926514 0.778563498
#> [637] 0.859778569 0.929412446 0.833935370 0.868291303 0.797282390 1.018811832
#> [643] 1.011632238 0.983163036 0.964486583 0.839937252 0.931757885 0.952956061
#> [649] 1.100876246 1.508729997 1.600664528 1.546732900 1.374772146 1.381019256
#> [655] 1.552630670 1.539719959 1.448210796 1.497157149 1.521102615 1.546125959
#> [661] 1.524113512 1.408512652 1.455380663 1.529334517 1.445460921 1.720961652
#> [667] 1.730112607 1.722765733 1.732050808 1.670893354 1.621642749 1.681926517
#> [673] 1.665355817 1.680924200 1.625538768 0.037066861 0.009905107 0.058760856
#> [679] 0.050909321 0.000000000 0.030266486 0.138015108 0.061042472 0.063074584
#> [685] 0.117297635 0.274498327 0.341734583 0.320865388 0.218331864 0.335559538
#> [691] 0.252258947 0.202379969 0.151933039 0.303630086 0.271778584 0.257916294
#> [697] 0.187982579 0.278688714 0.390190144 0.328235991 0.482713207 0.617404710
#> [703] 0.604985028 0.618409921 0.631646064 0.593427000 0.597381477 0.708486670
#> [709] 0.712391079 0.661926514 0.778563498 0.859778569 0.929412446 0.833935370
#> [715] 0.868291303 0.797282390 1.018811832 1.011632238 0.983163036 0.964486583
#> [721] 0.839937252 0.931757885 0.952956061 1.100876246 1.508729997 1.600664528
#> [727] 1.546732900 1.374772146 1.381019256 1.552630670 1.539719959 1.448210796
#> [733] 1.497157149 1.521102615 1.546125959 1.524113512 1.408512652 1.455380663
#> [739] 1.529334517 1.445460921 1.720961652 1.730112607 1.722765733 1.732050808
#> [745] 1.670893354 1.621642749 1.681926517 1.665355817 1.680924200 1.625538768
#> [751] 0.037066861 0.009905107 0.058760856 0.050909321 0.000000000 0.030266486
#> [757] 0.138015108 0.061042472 0.063074584 0.117297635 0.274498327 0.341734583
#> [763] 0.320865388 0.218331864 0.335559538 0.252258947 0.202379969 0.151933039
#> [769] 0.303630086 0.271778584 0.257916294 0.187982579 0.278688714 0.390190144
#> [775] 0.328235991 0.482713207 0.617404710 0.604985028 0.618409921 0.631646064
#> [781] 0.593427000 0.597381477 0.708486670 0.712391079 0.661926514 0.778563498
#> [787] 0.859778569 0.929412446 0.833935370 0.868291303 0.797282390 1.018811832
#> [793] 1.011632238 0.983163036 0.964486583 0.839937252 0.931757885 0.952956061
#> [799] 1.100876246 1.508729997 1.600664528 1.546732900 1.374772146 1.381019256
#> [805] 1.552630670 1.539719959 1.448210796 1.497157149 1.521102615 1.546125959
#> [811] 1.524113512 1.408512652 1.455380663 1.529334517 1.445460921 1.720961652
#> [817] 1.730112607 1.722765733 1.732050808 1.670893354 1.621642749 1.681926517
#> [823] 1.665355817 1.680924200 1.625538768 0.037066861 0.009905107 0.058760856
#> [829] 0.050909321 0.000000000 0.030266486 0.138015108 0.061042472 0.063074584
#> [835] 0.117297635 0.274498327 0.341734583 0.320865388 0.218331864 0.335559538
#> [841] 0.252258947 0.202379969 0.151933039 0.303630086 0.271778584 0.257916294
#> [847] 0.187982579 0.278688714 0.390190144 0.328235991 0.482713207 0.617404710
#> [853] 0.604985028 0.618409921 0.631646064 0.593427000 0.597381477 0.708486670
#> [859] 0.712391079 0.661926514 0.778563498 0.859778569 0.929412446 0.833935370
#> [865] 0.868291303 0.797282390 1.018811832 1.011632238 0.983163036 0.964486583
#> [871] 0.839937252 0.931757885 0.952956061 1.100876246 1.508729997 1.600664528
#> [877] 1.546732900 1.374772146 1.381019256 1.552630670 1.539719959 1.448210796
#> [883] 1.497157149 1.521102615 1.546125959 1.524113512 1.408512652 1.455380663
#> [889] 1.529334517 1.445460921 1.720961652 1.730112607 1.722765733 1.732050808
#> [895] 1.670893354 1.621642749 1.681926517 1.665355817 1.680924200 1.625538768
#> 
#> $weights
#>   [1] 12.369036 12.122174  9.902068 11.007998 12.309419 13.140469 10.050446
#>   [8]  9.981883 11.811964 11.785327 10.167293  4.809112  6.855934  8.359596
#>  [15]  8.021714 11.230263  4.558431  5.328370  6.483709  5.636883  7.410164
#>  [22]  6.500370  3.857655 12.297259 12.297259  7.210297  9.772877 11.504207
#>  [29] 10.563811 12.075006  8.234947  8.229794  8.144866  9.975980 15.060127
#>  [36]  5.615884  8.547554  8.604203  7.047757  9.167929  5.440462  6.529727
#>  [43]  6.306557  6.809333  9.566230  7.931209  9.922158  8.713297 20.000000
#>  [50] 20.000000  5.924440  5.913467  5.665973  5.840162  6.212313  4.949328
#>  [57]  4.565493  7.210170  5.275101  5.923645  6.874316  7.255238  7.846925
#>  [64]  9.046988  8.286622  8.760349 12.032731  9.386415 10.123158 11.720920
#>  [71]  6.243418 11.123340 16.879992 11.689726 10.596207 12.369036 12.122174
#>  [78]  9.902068 11.007998 12.309419 13.140469 10.050446  9.981883 11.811964
#>  [85] 11.785327 10.167293  4.809112  6.855934  8.359596  8.021714 11.230263
#>  [92]  4.558431  5.328370  6.483709  5.636883  7.410164  6.500370  3.857655
#>  [99] 12.297259 12.297259  7.210297  9.772877 11.504207 10.563811 12.075006
#> [106]  8.234947  8.229794  8.144866  9.975980 15.060127  5.615884  8.547554
#> [113]  8.604203  7.047757  9.167929  5.440462  6.529727  6.306557  6.809333
#> [120]  9.566230  7.931209  9.922158  8.713297 20.000000 20.000000  5.924440
#> [127]  5.913467  5.665973  5.840162  6.212313  4.949328  4.565493  7.210170
#> [134]  5.275101  5.923645  6.874316  7.255238  7.846925  9.046988  8.286622
#> [141]  8.760349 12.032731  9.386415 10.123158 11.720920  6.243418 11.123340
#> [148] 16.879992 11.689726 10.596207 12.369036 12.122174  9.902068 11.007998
#> [155] 12.309419 13.140469 10.050446  9.981883 11.811964 11.785327 10.167293
#> [162]  4.809112  6.855934  8.359596  8.021714 11.230263  4.558431  5.328370
#> [169]  6.483709  5.636883  7.410164  6.500370  3.857655 12.297259 12.297259
#> [176]  7.210297  9.772877 11.504207 10.563811 12.075006  8.234947  8.229794
#> [183]  8.144866  9.975980 15.060127  5.615884  8.547554  8.604203  7.047757
#> [190]  9.167929  5.440462  6.529727  6.306557  6.809333  9.566230  7.931209
#> [197]  9.922158  8.713297 20.000000 20.000000  5.924440  5.913467  5.665973
#> [204]  5.840162  6.212313  4.949328  4.565493  7.210170  5.275101  5.923645
#> [211]  6.874316  7.255238  7.846925  9.046988  8.286622  8.760349 12.032731
#> [218]  9.386415 10.123158 11.720920  6.243418 11.123340 16.879992 11.689726
#> [225] 10.596207 12.369036 12.122174  9.902068 11.007998 12.309419 13.140469
#> [232] 10.050446  9.981883 11.811964 11.785327 10.167293  4.809112  6.855934
#> [239]  8.359596  8.021714 11.230263  4.558431  5.328370  6.483709  5.636883
#> [246]  7.410164  6.500370  3.857655 12.297259 12.297259  7.210297  9.772877
#> [253] 11.504207 10.563811 12.075006  8.234947  8.229794  8.144866  9.975980
#> [260] 15.060127  5.615884  8.547554  8.604203  7.047757  9.167929  5.440462
#> [267]  6.529727  6.306557  6.809333  9.566230  7.931209  9.922158  8.713297
#> [274] 20.000000 20.000000  5.924440  5.913467  5.665973  5.840162  6.212313
#> [281]  4.949328  4.565493  7.210170  5.275101  5.923645  6.874316  7.255238
#> [288]  7.846925  9.046988  8.286622  8.760349 12.032731  9.386415 10.123158
#> [295] 11.720920  6.243418 11.123340 16.879992 11.689726 10.596207 12.369036
#> [302] 12.122174  9.902068 11.007998 12.309419 13.140469 10.050446  9.981883
#> [309] 11.811964 11.785327 10.167293  4.809112  6.855934  8.359596  8.021714
#> [316] 11.230263  4.558431  5.328370  6.483709  5.636883  7.410164  6.500370
#> [323]  3.857655 12.297259 12.297259  7.210297  9.772877 11.504207 10.563811
#> [330] 12.075006  8.234947  8.229794  8.144866  9.975980 15.060127  5.615884
#> [337]  8.547554  8.604203  7.047757  9.167929  5.440462  6.529727  6.306557
#> [344]  6.809333  9.566230  7.931209  9.922158  8.713297 20.000000 20.000000
#> [351]  5.924440  5.913467  5.665973  5.840162  6.212313  4.949328  4.565493
#> [358]  7.210170  5.275101  5.923645  6.874316  7.255238  7.846925  9.046988
#> [365]  8.286622  8.760349 12.032731  9.386415 10.123158 11.720920  6.243418
#> [372] 11.123340 16.879992 11.689726 10.596207 12.369036 12.122174  9.902068
#> [379] 11.007998 12.309419 13.140469 10.050446  9.981883 11.811964 11.785327
#> [386] 10.167293  4.809112  6.855934  8.359596  8.021714 11.230263  4.558431
#> [393]  5.328370  6.483709  5.636883  7.410164  6.500370  3.857655 12.297259
#> [400] 12.297259  7.210297  9.772877 11.504207 10.563811 12.075006  8.234947
#> [407]  8.229794  8.144866  9.975980 15.060127  5.615884  8.547554  8.604203
#> [414]  7.047757  9.167929  5.440462  6.529727  6.306557  6.809333  9.566230
#> [421]  7.931209  9.922158  8.713297 20.000000 20.000000  5.924440  5.913467
#> [428]  5.665973  5.840162  6.212313  4.949328  4.565493  7.210170  5.275101
#> [435]  5.923645  6.874316  7.255238  7.846925  9.046988  8.286622  8.760349
#> [442] 12.032731  9.386415 10.123158 11.720920  6.243418 11.123340 16.879992
#> [449] 11.689726 10.596207 12.369036 12.122174  9.902068 11.007998 12.309419
#> [456] 13.140469 10.050446  9.981883 11.811964 11.785327 10.167293  4.809112
#> [463]  6.855934  8.359596  8.021714 11.230263  4.558431  5.328370  6.483709
#> [470]  5.636883  7.410164  6.500370  3.857655 12.297259 12.297259  7.210297
#> [477]  9.772877 11.504207 10.563811 12.075006  8.234947  8.229794  8.144866
#> [484]  9.975980 15.060127  5.615884  8.547554  8.604203  7.047757  9.167929
#> [491]  5.440462  6.529727  6.306557  6.809333  9.566230  7.931209  9.922158
#> [498]  8.713297 20.000000 20.000000  5.924440  5.913467  5.665973  5.840162
#> [505]  6.212313  4.949328  4.565493  7.210170  5.275101  5.923645  6.874316
#> [512]  7.255238  7.846925  9.046988  8.286622  8.760349 12.032731  9.386415
#> [519] 10.123158 11.720920  6.243418 11.123340 16.879992 11.689726 10.596207
#> [526] 12.369036 12.122174  9.902068 11.007998 12.309419 13.140469 10.050446
#> [533]  9.981883 11.811964 11.785327 10.167293  4.809112  6.855934  8.359596
#> [540]  8.021714 11.230263  4.558431  5.328370  6.483709  5.636883  7.410164
#> [547]  6.500370  3.857655 12.297259 12.297259  7.210297  9.772877 11.504207
#> [554] 10.563811 12.075006  8.234947  8.229794  8.144866  9.975980 15.060127
#> [561]  5.615884  8.547554  8.604203  7.047757  9.167929  5.440462  6.529727
#> [568]  6.306557  6.809333  9.566230  7.931209  9.922158  8.713297 20.000000
#> [575] 20.000000  5.924440  5.913467  5.665973  5.840162  6.212313  4.949328
#> [582]  4.565493  7.210170  5.275101  5.923645  6.874316  7.255238  7.846925
#> [589]  9.046988  8.286622  8.760349 12.032731  9.386415 10.123158 11.720920
#> [596]  6.243418 11.123340 16.879992 11.689726 10.596207 12.369036 12.122174
#> [603]  9.902068 11.007998 12.309419 13.140469 10.050446  9.981883 11.811964
#> [610] 11.785327 10.167293  4.809112  6.855934  8.359596  8.021714 11.230263
#> [617]  4.558431  5.328370  6.483709  5.636883  7.410164  6.500370  3.857655
#> [624] 12.297259 12.297259  7.210297  9.772877 11.504207 10.563811 12.075006
#> [631]  8.234947  8.229794  8.144866  9.975980 15.060127  5.615884  8.547554
#> [638]  8.604203  7.047757  9.167929  5.440462  6.529727  6.306557  6.809333
#> [645]  9.566230  7.931209  9.922158  8.713297 20.000000 20.000000  5.924440
#> [652]  5.913467  5.665973  5.840162  6.212313  4.949328  4.565493  7.210170
#> [659]  5.275101  5.923645  6.874316  7.255238  7.846925  9.046988  8.286622
#> [666]  8.760349 12.032731  9.386415 10.123158 11.720920  6.243418 11.123340
#> [673] 16.879992 11.689726 10.596207 12.369036 12.122174  9.902068 11.007998
#> [680] 12.309419 13.140469 10.050446  9.981883 11.811964 11.785327 10.167293
#> [687]  4.809112  6.855934  8.359596  8.021714 11.230263  4.558431  5.328370
#> [694]  6.483709  5.636883  7.410164  6.500370  3.857655 12.297259 12.297259
#> [701]  7.210297  9.772877 11.504207 10.563811 12.075006  8.234947  8.229794
#> [708]  8.144866  9.975980 15.060127  5.615884  8.547554  8.604203  7.047757
#> [715]  9.167929  5.440462  6.529727  6.306557  6.809333  9.566230  7.931209
#> [722]  9.922158  8.713297 20.000000 20.000000  5.924440  5.913467  5.665973
#> [729]  5.840162  6.212313  4.949328  4.565493  7.210170  5.275101  5.923645
#> [736]  6.874316  7.255238  7.846925  9.046988  8.286622  8.760349 12.032731
#> [743]  9.386415 10.123158 11.720920  6.243418 11.123340 16.879992 11.689726
#> [750] 10.596207 12.369036 12.122174  9.902068 11.007998 12.309419 13.140469
#> [757] 10.050446  9.981883 11.811964 11.785327 10.167293  4.809112  6.855934
#> [764]  8.359596  8.021714 11.230263  4.558431  5.328370  6.483709  5.636883
#> [771]  7.410164  6.500370  3.857655 12.297259 12.297259  7.210297  9.772877
#> [778] 11.504207 10.563811 12.075006  8.234947  8.229794  8.144866  9.975980
#> [785] 15.060127  5.615884  8.547554  8.604203  7.047757  9.167929  5.440462
#> [792]  6.529727  6.306557  6.809333  9.566230  7.931209  9.922158  8.713297
#> [799] 20.000000 20.000000  5.924440  5.913467  5.665973  5.840162  6.212313
#> [806]  4.949328  4.565493  7.210170  5.275101  5.923645  6.874316  7.255238
#> [813]  7.846925  9.046988  8.286622  8.760349 12.032731  9.386415 10.123158
#> [820] 11.720920  6.243418 11.123340 16.879992 11.689726 10.596207 12.369036
#> [827] 12.122174  9.902068 11.007998 12.309419 13.140469 10.050446  9.981883
#> [834] 11.811964 11.785327 10.167293  4.809112  6.855934  8.359596  8.021714
#> [841] 11.230263  4.558431  5.328370  6.483709  5.636883  7.410164  6.500370
#> [848]  3.857655 12.297259 12.297259  7.210297  9.772877 11.504207 10.563811
#> [855] 12.075006  8.234947  8.229794  8.144866  9.975980 15.060127  5.615884
#> [862]  8.547554  8.604203  7.047757  9.167929  5.440462  6.529727  6.306557
#> [869]  6.809333  9.566230  7.931209  9.922158  8.713297 20.000000 20.000000
#> [876]  5.924440  5.913467  5.665973  5.840162  6.212313  4.949328  4.565493
#> [883]  7.210170  5.275101  5.923645  6.874316  7.255238  7.846925  9.046988
#> [890]  8.286622  8.760349 12.032731  9.386415 10.123158 11.720920  6.243418
#> [897] 11.123340 16.879992 11.689726 10.596207
#>