Extract hexagonal bin centroids coordinates and the corresponding standardise counts.

Extract hexagonal bin centroids coordinates and the corresponding standardise counts.

Usage

extract_hexbin_centroids(centroids_data, counts_data)

Arguments

centroids_data

A tibble that contains all hexagonal bin centroid coordinates with hexagon IDs.

counts_data

A tibble that contains hexagon IDs with the standardise number of points within each hexagon.

Value

A tibble contains hexagon ID (h), centroid coordinates (c_x, c_y), bin counts (n_h), and standardise counts (w_h).

Examples

all_centroids_df <- scurve_model_obj$hb_obj$centroids
counts_data <- scurve_model_obj$hb_obj$std_cts
extract_hexbin_centroids(centroids_data = all_centroids_df,
counts_data = counts_data)
#>       h         c_x         c_y n_h    w_h
#> 1     1 -0.10000000 -0.08923607   0 0.0000
#> 2     2 -0.01673864 -0.08923607   0 0.0000
#> 3     3  0.06652271 -0.08923607   0 0.0000
#> 4     4  0.14978407 -0.08923607   0 0.0000
#> 5     5  0.23304543 -0.08923607   0 0.0000
#> 6     6  0.31630679 -0.08923607   0 0.0000
#> 7     7  0.39956814 -0.08923607   0 0.0000
#> 8     8  0.48282950 -0.08923607   0 0.0000
#> 9     9  0.56609086 -0.08923607   0 0.0000
#> 10   10  0.64935221 -0.08923607   0 0.0000
#> 11   11  0.73261357 -0.08923607   0 0.0000
#> 12   12  0.81587493 -0.08923607   0 0.0000
#> 13   13  0.89913629 -0.08923607   0 0.0000
#> 14   14  0.98239764 -0.08923607   0 0.0000
#> 15   15  1.06565900 -0.08923607   0 0.0000
#> 16   16 -0.05836932 -0.01712962   0 0.0000
#> 17   17  0.02489204 -0.01712962   0 0.0000
#> 18   18  0.10815339 -0.01712962   0 0.0000
#> 19   19  0.19141475 -0.01712962   0 0.0000
#> 20   20  0.27467611 -0.01712962   0 0.0000
#> 21   21  0.35793746 -0.01712962  12 0.0024
#> 22   22  0.44119882 -0.01712962  17 0.0034
#> 23   23  0.52446018 -0.01712962  22 0.0044
#> 24   24  0.60772154 -0.01712962  10 0.0020
#> 25   25  0.69098289 -0.01712962   7 0.0014
#> 26   26  0.77424425 -0.01712962  12 0.0024
#> 27   27  0.85750561 -0.01712962   1 0.0002
#> 28   28  0.94076697 -0.01712962   0 0.0000
#> 29   29  1.02402832 -0.01712962   0 0.0000
#> 30   30  1.10728968 -0.01712962   0 0.0000
#> 31   31 -0.10000000  0.05497683   0 0.0000
#> 32   32 -0.01673864  0.05497683   0 0.0000
#> 33   33  0.06652271  0.05497683   0 0.0000
#> 34   34  0.14978407  0.05497683   0 0.0000
#> 35   35  0.23304543  0.05497683   0 0.0000
#> 36   36  0.31630679  0.05497683  42 0.0084
#> 37   37  0.39956814  0.05497683  42 0.0084
#> 38   38  0.48282950  0.05497683  44 0.0088
#> 39   39  0.56609086  0.05497683  39 0.0078
#> 40   40  0.64935221  0.05497683  52 0.0104
#> 41   41  0.73261357  0.05497683  38 0.0076
#> 42   42  0.81587493  0.05497683  52 0.0104
#> 43   43  0.89913629  0.05497683  36 0.0072
#> 44   44  0.98239764  0.05497683  27 0.0054
#> 45   45  1.06565900  0.05497683   0 0.0000
#> 46   46 -0.05836932  0.12708328   0 0.0000
#> 47   47  0.02489204  0.12708328   0 0.0000
#> 48   48  0.10815339  0.12708328   0 0.0000
#> 49   49  0.19141475  0.12708328   2 0.0004
#> 50   50  0.27467611  0.12708328  65 0.0130
#> 51   51  0.35793746  0.12708328  47 0.0094
#> 52   52  0.44119882  0.12708328  45 0.0090
#> 53   53  0.52446018  0.12708328  35 0.0070
#> 54   54  0.60772154  0.12708328  31 0.0062
#> 55   55  0.69098289  0.12708328  40 0.0080
#> 56   56  0.77424425  0.12708328  43 0.0086
#> 57   57  0.85750561  0.12708328  36 0.0072
#> 58   58  0.94076697  0.12708328  43 0.0086
#> 59   59  1.02402832  0.12708328   0 0.0000
#> 60   60  1.10728968  0.12708328   0 0.0000
#> 61   61 -0.10000000  0.19918973   0 0.0000
#> 62   62 -0.01673864  0.19918973   0 0.0000
#> 63   63  0.06652271  0.19918973   3 0.0006
#> 64   64  0.14978407  0.19918973  39 0.0078
#> 65   65  0.23304543  0.19918973  67 0.0134
#> 66   66  0.31630679  0.19918973  56 0.0112
#> 67   67  0.39956814  0.19918973  38 0.0076
#> 68   68  0.48282950  0.19918973  42 0.0084
#> 69   69  0.56609086  0.19918973  42 0.0084
#> 70   70  0.64935221  0.19918973  30 0.0060
#> 71   71  0.73261357  0.19918973  41 0.0082
#> 72   72  0.81587493  0.19918973  33 0.0066
#> 73   73  0.89913629  0.19918973  38 0.0076
#> 74   74  0.98239764  0.19918973  10 0.0020
#> 75   75  1.06565900  0.19918973   0 0.0000
#> 76   76 -0.05836932  0.27129618   0 0.0000
#> 77   77  0.02489204  0.27129618   1 0.0002
#> 78   78  0.10815339  0.27129618  62 0.0124
#> 79   79  0.19141475  0.27129618  62 0.0124
#> 80   80  0.27467611  0.27129618  68 0.0136
#> 81   81  0.35793746  0.27129618  50 0.0100
#> 82   82  0.44119882  0.27129618  32 0.0064
#> 83   83  0.52446018  0.27129618  44 0.0088
#> 84   84  0.60772154  0.27129618  34 0.0068
#> 85   85  0.69098289  0.27129618  40 0.0080
#> 86   86  0.77424425  0.27129618  36 0.0072
#> 87   87  0.85750561  0.27129618  43 0.0086
#> 88   88  0.94076697  0.27129618  40 0.0080
#> 89   89  1.02402832  0.27129618   0 0.0000
#> 90   90  1.10728968  0.27129618   0 0.0000
#> 91   91 -0.10000000  0.34340263   0 0.0000
#> 92   92 -0.01673864  0.34340263  13 0.0026
#> 93   93  0.06652271  0.34340263  45 0.0090
#> 94   94  0.14978407  0.34340263  34 0.0068
#> 95   95  0.23304543  0.34340263  58 0.0116
#> 96   96  0.31630679  0.34340263  32 0.0064
#> 97   97  0.39956814  0.34340263  25 0.0050
#> 98   98  0.48282950  0.34340263  38 0.0076
#> 99   99  0.56609086  0.34340263  40 0.0080
#> 100 100  0.64935221  0.34340263  43 0.0086
#> 101 101  0.73261357  0.34340263  37 0.0074
#> 102 102  0.81587493  0.34340263  42 0.0084
#> 103 103  0.89913629  0.34340263  50 0.0100
#> 104 104  0.98239764  0.34340263  14 0.0028
#> 105 105  1.06565900  0.34340263   0 0.0000
#> 106 106 -0.05836932  0.41550908   0 0.0000
#> 107 107  0.02489204  0.41550908  41 0.0082
#> 108 108  0.10815339  0.41550908  34 0.0068
#> 109 109  0.19141475  0.41550908  45 0.0090
#> 110 110  0.27467611  0.41550908  60 0.0120
#> 111 111  0.35793746  0.41550908   4 0.0008
#> 112 112  0.44119882  0.41550908   0 0.0000
#> 113 113  0.52446018  0.41550908   0 0.0000
#> 114 114  0.60772154  0.41550908   1 0.0002
#> 115 115  0.69098289  0.41550908   8 0.0016
#> 116 116  0.77424425  0.41550908  22 0.0044
#> 117 117  0.85750561  0.41550908  24 0.0048
#> 118 118  0.94076697  0.41550908  19 0.0038
#> 119 119  1.02402832  0.41550908   0 0.0000
#> 120 120  1.10728968  0.41550908   0 0.0000
#> 121 121 -0.10000000  0.48761554   0 0.0000
#> 122 122 -0.01673864  0.48761554  10 0.0020
#> 123 123  0.06652271  0.48761554  31 0.0062
#> 124 124  0.14978407  0.48761554  36 0.0072
#> 125 125  0.23304543  0.48761554  39 0.0078
#> 126 126  0.31630679  0.48761554  54 0.0108
#> 127 127  0.39956814  0.48761554  34 0.0068
#> 128 128  0.48282950  0.48761554   7 0.0014
#> 129 129  0.56609086  0.48761554   0 0.0000
#> 130 130  0.64935221  0.48761554   0 0.0000
#> 131 131  0.73261357  0.48761554   0 0.0000
#> 132 132  0.81587493  0.48761554   3 0.0006
#> 133 133  0.89913629  0.48761554   5 0.0010
#> 134 134  0.98239764  0.48761554   0 0.0000
#> 135 135  1.06565900  0.48761554   0 0.0000
#> 136 136 -0.05836932  0.55972199   0 0.0000
#> 137 137  0.02489204  0.55972199  32 0.0064
#> 138 138  0.10815339  0.55972199  32 0.0064
#> 139 139  0.19141475  0.55972199  52 0.0104
#> 140 140  0.27467611  0.55972199  48 0.0096
#> 141 141  0.35793746  0.55972199  48 0.0096
#> 142 142  0.44119882  0.55972199  49 0.0098
#> 143 143  0.52446018  0.55972199  36 0.0072
#> 144 144  0.60772154  0.55972199   1 0.0002
#> 145 145  0.69098289  0.55972199  28 0.0056
#> 146 146  0.77424425  0.55972199  46 0.0092
#> 147 147  0.85750561  0.55972199  53 0.0106
#> 148 148  0.94076697  0.55972199  42 0.0084
#> 149 149  1.02402832  0.55972199   0 0.0000
#> 150 150  1.10728968  0.55972199   0 0.0000
#> 151 151 -0.10000000  0.63182844   0 0.0000
#> 152 152 -0.01673864  0.63182844   0 0.0000
#> 153 153  0.06652271  0.63182844  47 0.0094
#> 154 154  0.14978407  0.63182844  35 0.0070
#> 155 155  0.23304543  0.63182844  37 0.0074
#> 156 156  0.31630679  0.63182844  41 0.0082
#> 157 157  0.39956814  0.63182844  46 0.0092
#> 158 158  0.48282950  0.63182844  42 0.0084
#> 159 159  0.56609086  0.63182844  46 0.0092
#> 160 160  0.64935221  0.63182844  56 0.0112
#> 161 161  0.73261357  0.63182844  35 0.0070
#> 162 162  0.81587493  0.63182844  31 0.0062
#> 163 163  0.89913629  0.63182844  44 0.0088
#> 164 164  0.98239764  0.63182844  26 0.0052
#> 165 165  1.06565900  0.63182844   0 0.0000
#> 166 166 -0.05836932  0.70393489   0 0.0000
#> 167 167  0.02489204  0.70393489   0 0.0000
#> 168 168  0.10815339  0.70393489  28 0.0056
#> 169 169  0.19141475  0.70393489  34 0.0068
#> 170 170  0.27467611  0.70393489  43 0.0086
#> 171 171  0.35793746  0.70393489  49 0.0098
#> 172 172  0.44119882  0.70393489  43 0.0086
#> 173 173  0.52446018  0.70393489  56 0.0112
#> 174 174  0.60772154  0.70393489  75 0.0150
#> 175 175  0.69098289  0.70393489  49 0.0098
#> 176 176  0.77424425  0.70393489  38 0.0076
#> 177 177  0.85750561  0.70393489  42 0.0084
#> 178 178  0.94076697  0.70393489  50 0.0100
#> 179 179  1.02402832  0.70393489   4 0.0008
#> 180 180  1.10728968  0.70393489   0 0.0000
#> 181 181 -0.10000000  0.77604134   0 0.0000
#> 182 182 -0.01673864  0.77604134   0 0.0000
#> 183 183  0.06652271  0.77604134   0 0.0000
#> 184 184  0.14978407  0.77604134  17 0.0034
#> 185 185  0.23304543  0.77604134  42 0.0084
#> 186 186  0.31630679  0.77604134  43 0.0086
#> 187 187  0.39956814  0.77604134  45 0.0090
#> 188 188  0.48282950  0.77604134  49 0.0098
#> 189 189  0.56609086  0.77604134  63 0.0126
#> 190 190  0.64935221  0.77604134  58 0.0116
#> 191 191  0.73261357  0.77604134  45 0.0090
#> 192 192  0.81587493  0.77604134  42 0.0084
#> 193 193  0.89913629  0.77604134  41 0.0082
#> 194 194  0.98239764  0.77604134  39 0.0078
#> 195 195  1.06565900  0.77604134   0 0.0000
#> 196 196 -0.05836932  0.84814779   0 0.0000
#> 197 197  0.02489204  0.84814779   0 0.0000
#> 198 198  0.10815339  0.84814779   0 0.0000
#> 199 199  0.19141475  0.84814779   0 0.0000
#> 200 200  0.27467611  0.84814779  11 0.0022
#> 201 201  0.35793746  0.84814779  39 0.0078
#> 202 202  0.44119882  0.84814779  36 0.0072
#> 203 203  0.52446018  0.84814779   8 0.0016
#> 204 204  0.60772154  0.84814779  17 0.0034
#> 205 205  0.69098289  0.84814779  44 0.0088
#> 206 206  0.77424425  0.84814779  47 0.0094
#> 207 207  0.85750561  0.84814779  47 0.0094
#> 208 208  0.94076697  0.84814779  41 0.0082
#> 209 209  1.02402832  0.84814779  11 0.0022
#> 210 210  1.10728968  0.84814779   0 0.0000
#> 211 211 -0.10000000  0.92025424   0 0.0000
#> 212 212 -0.01673864  0.92025424   0 0.0000
#> 213 213  0.06652271  0.92025424   0 0.0000
#> 214 214  0.14978407  0.92025424   0 0.0000
#> 215 215  0.23304543  0.92025424   0 0.0000
#> 216 216  0.31630679  0.92025424   0 0.0000
#> 217 217  0.39956814  0.92025424   0 0.0000
#> 218 218  0.48282950  0.92025424   0 0.0000
#> 219 219  0.56609086  0.92025424   0 0.0000
#> 220 220  0.64935221  0.92025424   0 0.0000
#> 221 221  0.73261357  0.92025424   3 0.0006
#> 222 222  0.81587493  0.92025424   8 0.0016
#> 223 223  0.89913629  0.92025424   1 0.0002
#> 224 224  0.98239764  0.92025424   0 0.0000
#> 225 225  1.06565900  0.92025424   0 0.0000
#> 226 226 -0.05836932  0.99236069   0 0.0000
#> 227 227  0.02489204  0.99236069   0 0.0000
#> 228 228  0.10815339  0.99236069   0 0.0000
#> 229 229  0.19141475  0.99236069   0 0.0000
#> 230 230  0.27467611  0.99236069   0 0.0000
#> 231 231  0.35793746  0.99236069   0 0.0000
#> 232 232  0.44119882  0.99236069   0 0.0000
#> 233 233  0.52446018  0.99236069   0 0.0000
#> 234 234  0.60772154  0.99236069   0 0.0000
#> 235 235  0.69098289  0.99236069   0 0.0000
#> 236 236  0.77424425  0.99236069   0 0.0000
#> 237 237  0.85750561  0.99236069   0 0.0000
#> 238 238  0.94076697  0.99236069   0 0.0000
#> 239 239  1.02402832  0.99236069   0 0.0000
#> 240 240  1.10728968  0.99236069   0 0.0000