Find low-density Hexagons

This function identifies hexagons with low density based on the mean density of their neighboring hexagons.

Usage

find_low_dens_hex(model_2d, b1 = 4, benchmark_mean_dens = 0.05)

Arguments

model_2d

The tibble that contains all hexagonal bin centroids.

b1

Number of bins along the x-axis for hexagon binning.

benchmark_mean_dens

A numeric value that contains

Value

A vector containing the IDs of hexagons to be removed after investigating their neighboring bins.

Examples

find_low_dens_hex(model_2d = scurve_model_obj$model_2d, b1 = 4,
benchmark_mean_dens = 0.05)
#>   [1]  21  22  23  26  36  37  38  39  40  41  42  43  44  50  51  52  53  54
#>  [19]  55  56  57  58  64  65  66  67  68  69  70  71  72  73  78  79  80  81
#>  [37]  82  83  84  85  86  87  88  92  93  94  95  96  97  98  99 100 101 102
#>  [55] 103 104 107 108 109 110 116 117 118 123 124 125 126 127 137 138 139 140
#>  [73] 141 142 143 145 146 147 148 153 154 155 156 157 158 159 160 161 162 163
#>  [91] 164 168 169 170 171 172 173 174 175 176 177 178 184 185 186 187 188 189
#> [109] 190 191 192 193 194 200 201 202 204 205 206 207 208 209