Find points in hexagonal bins

This function maps points to their corresponding hexagonal bins.

Usage

find_pts(scaled_nldr_hexid)

Arguments

scaled_nldr_hexid

A tibble that contains the scaled embedding with hexagonal bin IDs.

Value

A tibble with hexagonal bin IDs (h) and the corresponding points.

Examples

umap_with_hb_id <- scurve_model_obj$hb_obj$data_hb_id
find_pts(scaled_nldr_hexid = umap_with_hb_id)
#> # A tibble: 142 × 2
#>        h pts_list  
#>    <int> <list>    
#>  1    21 <int [12]>
#>  2    22 <int [17]>
#>  3    23 <int [22]>
#>  4    24 <int [10]>
#>  5    25 <int [7]> 
#>  6    26 <int [12]>
#>  7    27 <int [1]> 
#>  8    36 <int [42]>
#>  9    37 <int [42]>
#> 10    38 <int [44]>
#> # ℹ 132 more rows