Construct the 2D model and lift into high-D

This function fits a high-dimensional model using hexagonal bins and provides options to customize the modeling process, including the choice of bin centroids or bin means, removal of low-density hexagons, and averaging of high-dimensional data.

Usage

fit_high_d_model(
  training_data,
  nldr_df_with_id,
  x = "UMAP1",
  y = "UMAP2",
  num_bins_x = NA,
  num_bins_y = NA,
  x_start = NA,
  y_start = NA,
  buffer_x = NA,
  buffer_y = NA,
  hex_size = NA,
  is_bin_centroid = TRUE,
  is_rm_lwd_hex = FALSE,
  benchmark_to_rm_lwd_hex = NA,
  is_avg_high_d = TRUE,
  column_start_text = "x"
)

Arguments

training_data

A data frame containing the training high-dimensional data.

nldr_df_with_id

A data frame containing 2D embeddings with a unique identifier.

x

The name of the column that contains first 2D embeddings component.

y

The name of the column that contains second 2D embeddings component.

num_bins_x

Number of bins along the x-axis.

num_bins_y

Number of bins along the y-axis.

x_start

Starting point along the x-axis for hexagonal binning.

y_start

Starting point along the y-axis for hexagonal binning.

buffer_x

The buffer size along the x-axis.

buffer_y

The buffer size along the y-axis.

hex_size

A numeric value that initializes the radius of the outer circle surrounding the hexagon.

is_bin_centroid

Logical, indicating whether to use bin centroids (default is TRUE).

is_rm_lwd_hex

Logical, indicating whether to remove low-density hexagons (default is FALSE).

benchmark_to_rm_lwd_hex

The benchmark value to remove low-density hexagons.

is_avg_high_d

Logical, indicating whether to average the high-dimensional data within bins (default is TRUE).

column_start_text

The text prefix for columns in the high-dimensional data.

Value

A list containing the data frame with high-dimensional coordinates for 2D bin centroids (df_bin) and the data frame containing information about hexagonal bin centroids (df_bin_centroids) in 2D.

Examples

fit_high_d_model(training_data = s_curve_noise_training,
nldr_df_with_id = s_curve_noise_umap_scaled)
#> Hex size is set to 0.2.
#> Buffer along the x-axis is set to 0.519615242270663.
#> Hex size is set to 0.2.
#> Buffer along the y-axis is set to 0.45.
#> Hex size is set to 0.2.
#> x_start is set to 0.
#> y_start is set to 0.
#> Hex size is set to 0.2.
#> $df_bin
#> # A tibble: 12 × 8
#>    hb_id      x1    x2     x3        x4        x5       x6        x7
#>    <dbl>   <dbl> <dbl>  <dbl>     <dbl>     <dbl>    <dbl>     <dbl>
#>  1     1 -0.753  1.62  -1.56   0.00687   0.00277  -0.0176  -0.000562
#>  2     2  0.213  1.67  -1.91   0.00714  -0.00672  -0.00213 -0.000491
#>  3     6 -0.0291 0.403 -1.87   0.00370  -0.000454 -0.0209  -0.00177 
#>  4     7  0.943  0.757 -1.32  -0.00720  -0.00319   0.0572   0.00230 
#>  5    12  0.799  0.527 -0.476  0.00328   0.00490  -0.00246 -0.000786
#>  6    13  0.660  1.51  -0.331 -0.000441  0.000757  0.0274   0.00201 
#>  7    18 -0.452  1.07   0.246  0.00355   0.00165   0.00457  0.00121 
#>  8    23 -0.988  1.64   1.16   0.00754  -0.0104   -0.00129  0.00489 
#>  9    27 -0.557  1.52   1.83   0.00998   0.00198  -0.0936  -0.00417 
#> 10    28  0.0233 0.749  1.84  -0.00276   0.00397  -0.00751 -0.00132 
#> 11    33  0.689  1.74   1.64   0.00210   0.00370  -0.00532 -0.000634
#> 12    34  0.984  0.809  1.15   0.00437  -0.00191   0.0210   0.000373
#> 
#> $df_bin_centroids
#>    hexID       c_x c_y std_counts
#> 1      1 0.0000000 0.0 0.40000000
#> 2      2 0.3464102 0.0 0.26666667
#> 3      6 0.1732051 0.3 0.86666667
#> 4      7 0.5196152 0.3 0.13333333
#> 5     12 0.3464102 0.6 0.20000000
#> 6     13 0.6928203 0.6 0.53333333
#> 7     18 0.8660254 0.9 0.80000000
#> 8     23 0.6928203 1.2 0.06666667
#> 9     27 0.5196152 1.5 0.06666667
#> 10    28 0.8660254 1.5 1.00000000
#> 11    33 0.6928203 1.8 0.46666667
#> 12    34 1.0392305 1.8 0.20000000
#>