Scaling the data

This function scales the x and y coordinates.

Usage

gen_scaled_data(data, x, y, hex_ratio = NA)

Arguments

data

A tibble or data frame.

x

The name of the column that contains values along the x-axis.

y

The name of the column that contains values along the y-axis.

hex_ratio

Numeric value representing the ratio of the hexagon size.

Value

A list contains scaled x and y coordinates.

Examples

gen_scaled_data(data = s_curve_noise_umap, x = "UMAP1", y = "UMAP2")
#> $scaled_UMAP1
#>  [1] 0.08044271 0.73857864 0.83992020 0.16719939 0.26292782 0.83820653
#>  [7] 0.73364820 0.62693057 0.80998347 0.90342138 0.64241759 0.05182936
#> [13] 0.96228699 0.07223295 0.64916340 0.80783543 0.13518498 0.80668446
#> [19] 0.60311661 0.16150976 0.71326033 0.09405763 0.08900367 0.77510396
#> [25] 0.79296306 0.28575187 0.46806684 0.57541142 0.21053454 0.62412077
#> [31] 0.76738169 0.76427076 0.64842247 0.03170275 0.96450414 0.10117880
#> [37] 0.71265285 0.82904448 0.81275032 0.00000000 0.38392618 0.82638316
#> [43] 0.29256853 0.84482412 0.79995094 0.69339205 0.13459287 0.04209521
#> [49] 0.94245034 0.05349536 0.68677026 0.77828853 0.75504949 0.23125328
#> [55] 0.76831122 0.69843891 0.72681720 0.99318862 0.96227596 0.39315580
#> [61] 0.74175661 0.94650462 0.50932690 1.00000000 0.33548077 0.25450458
#> [67] 0.88588122 0.72192354 0.68029522 0.92537063 0.80964667 0.49582819
#> [73] 0.16301062 0.29499615 0.91053035
#> 
#> $scaled_UMAP2
#>  [1] 0.36599777 1.14637143 1.23921659 0.04942248 0.45564611 2.29461554
#>  [7] 1.11191383 0.82454656 1.15772174 2.13421937 0.84219475 0.42782052
#> [13] 2.06231053 0.29110915 0.79123600 1.83302953 0.44741272 1.84135901
#> [19] 0.79650864 0.01320681 1.06304319 0.07834781 0.33634526 2.30681681
#> [25] 1.46783500 0.26983996 0.64361761 0.94464889 0.20257739 0.94985477
#> [31] 1.35841578 2.07017423 0.88256869 0.40484011 2.22047442 0.06787909
#> [37] 2.05295995 2.29702098 2.30940108 0.36237145 0.52025353 1.34884298
#> [43] 0.18402014 1.31088405 1.28598211 1.93094773 0.00000000 0.34388839
#> [49] 2.24123227 0.25064344 1.03808466 1.11991634 2.22785780 0.08138996
#> [55] 1.99620953 2.16219033 1.24234385 2.02813682 2.06150128 0.43764799
#> [61] 2.24256869 2.03215135 0.82320628 2.16738502 0.37158495 0.08409944
#> [67] 1.87801687 1.27060808 2.01164000 1.94050755 2.03911269 0.80664670
#> [73] 0.04035531 0.15639685 1.92728123
#>