Scaling the 2D embeddings

This function scales the 2D embeddings.

Usage

generate_scaled_data(nldr_df, x = "UMAP1", y = "UMAP2")

Arguments

nldr_df

A data frame containing 2D embeddings.

x

The name of the column that contains first 2D embeddings component.

y

The name of the column that contains second 2D embeddings component.

Value

A list contains scaled 2D embeddings.

Examples

generate_scaled_data(nldr_df = s_curve_noise_umap, x = "UMAP1", y = "UMAP2")
#> $scaled_UMAP1
#>  [1] 0.08044271 0.73857864 0.83992020 0.16719939 0.26292782 0.83820653
#>  [7] 0.73364820 0.62693057 0.80998347 0.90342138 0.64241759 0.05182936
#> [13] 0.96228699 0.07223295 0.64916340 0.80783543 0.13518498 0.80668446
#> [19] 0.60311661 0.16150976 0.71326033 0.09405763 0.08900367 0.77510396
#> [25] 0.79296306 0.28575187 0.46806684 0.57541142 0.21053454 0.62412077
#> [31] 0.76738169 0.76427076 0.64842247 0.03170275 0.96450414 0.10117880
#> [37] 0.71265285 0.82904448 0.81275032 0.00000000 0.38392618 0.82638316
#> [43] 0.29256853 0.84482412 0.79995094 0.69339205 0.13459287 0.04209521
#> [49] 0.94245034 0.05349536 0.68677026 0.77828853 0.75504949 0.23125328
#> [55] 0.76831122 0.69843891 0.72681720 0.99318862 0.96227596 0.39315580
#> [61] 0.74175661 0.94650462 0.50932690 1.00000000 0.33548077 0.25450458
#> [67] 0.88588122 0.72192354 0.68029522 0.92537063 0.80964667 0.49582819
#> [73] 0.16301062 0.29499615 0.91053035
#> 
#> $scaled_UMAP2
#>  [1] 0.274498327 0.859778569 0.929412446 0.037066861 0.341734583 1.720961652
#>  [7] 0.833935370 0.618409921 0.868291303 1.600664528 0.631646064 0.320865388
#> [13] 1.546732900 0.218331864 0.593427000 1.374772146 0.335559538 1.381019256
#> [19] 0.597381477 0.009905107 0.797282390 0.058760856 0.252258947 1.730112607
#> [25] 1.100876246 0.202379969 0.482713207 0.708486670 0.151933039 0.712391079
#> [31] 1.018811832 1.552630670 0.661926514 0.303630086 1.665355817 0.050909321
#> [37] 1.539719959 1.722765733 1.732050808 0.271778584 0.390190144 1.011632238
#> [43] 0.138015108 0.983163036 0.964486583 1.448210796 0.000000000 0.257916294
#> [49] 1.680924200 0.187982579 0.778563498 0.839937252 1.670893354 0.061042472
#> [55] 1.497157149 1.621642749 0.931757885 1.521102615 1.546125959 0.328235991
#> [61] 1.681926517 1.524113512 0.617404710 1.625538768 0.278688714 0.063074584
#> [67] 1.408512652 0.952956061 1.508729997 1.455380663 1.529334517 0.604985028
#> [73] 0.030266486 0.117297635 1.445460921
#>