Extract hexagonal bin centroids coordinates and the corresponding standardise counts.

Extract hexagonal bin centroids coordinates and the corresponding standardise counts.

Usage

merge_hexbin_centroids(centroids_data, counts_data)

Arguments

centroids_data

A tibble that contains all hexagonal bin centroid coordinates with hexagon IDs.

counts_data

A tibble that contains hexagon IDs with the standardise number of points within each hexagon.

Value

A tibble contains hexagon ID (h), centroid coordinates (c_x, c_y), bin counts (n_h), and standardise counts (w_h).

Examples

all_centroids_df <- scurve_model_obj$hb_obj$centroids
counts_data <- scurve_model_obj$hb_obj$std_cts
merge_hexbin_centroids(centroids_data = all_centroids_df,
counts_data = counts_data)
#>       h        c_x         c_y n_h   w_h
#> 1     1 -0.1000000 -0.11568014   0 0.000
#> 2     2 -0.0165892 -0.11568014   0 0.000
#> 3     3  0.0668216 -0.11568014   0 0.000
#> 4     4  0.1502324 -0.11568014   0 0.000
#> 5     5  0.2336432 -0.11568014   0 0.000
#> 6     6  0.3170540 -0.11568014   0 0.000
#> 7     7  0.4004648 -0.11568014   0 0.000
#> 8     8  0.4838756 -0.11568014   0 0.000
#> 9     9  0.5672864 -0.11568014   0 0.000
#> 10   10  0.6506972 -0.11568014   0 0.000
#> 11   11  0.7341080 -0.11568014   0 0.000
#> 12   12  0.8175188 -0.11568014   0 0.000
#> 13   13  0.9009296 -0.11568014   0 0.000
#> 14   14  0.9843404 -0.11568014   0 0.000
#> 15   15  1.0677512 -0.11568014   0 0.000
#> 16   16 -0.0582946 -0.04344427   0 0.000
#> 17   17  0.0251162 -0.04344427   0 0.000
#> 18   18  0.1085270 -0.04344427   0 0.000
#> 19   19  0.1919378 -0.04344427   0 0.000
#> 20   20  0.2753486 -0.04344427   0 0.000
#> 21   21  0.3587594 -0.04344427   0 0.000
#> 22   22  0.4421702 -0.04344427   0 0.000
#> 23   23  0.5255810 -0.04344427   0 0.000
#> 24   24  0.6089918 -0.04344427   0 0.000
#> 25   25  0.6924026 -0.04344427   0 0.000
#> 26   26  0.7758134 -0.04344427   0 0.000
#> 27   27  0.8592242 -0.04344427   0 0.000
#> 28   28  0.9426350 -0.04344427   0 0.000
#> 29   29  1.0260458 -0.04344427   0 0.000
#> 30   30  1.1094566 -0.04344427   0 0.000
#> 31   31 -0.1000000  0.02879160   0 0.000
#> 32   32 -0.0165892  0.02879160   0 0.000
#> 33   33  0.0668216  0.02879160   0 0.000
#> 34   34  0.1502324  0.02879160   1 0.001
#> 35   35  0.2336432  0.02879160  10 0.010
#> 36   36  0.3170540  0.02879160   8 0.008
#> 37   37  0.4004648  0.02879160   8 0.008
#> 38   38  0.4838756  0.02879160   2 0.002
#> 39   39  0.5672864  0.02879160   0 0.000
#> 40   40  0.6506972  0.02879160   0 0.000
#> 41   41  0.7341080  0.02879160   2 0.002
#> 42   42  0.8175188  0.02879160  12 0.012
#> 43   43  0.9009296  0.02879160   0 0.000
#> 44   44  0.9843404  0.02879160   0 0.000
#> 45   45  1.0677512  0.02879160   0 0.000
#> 46   46 -0.0582946  0.10102747   0 0.000
#> 47   47  0.0251162  0.10102747   0 0.000
#> 48   48  0.1085270  0.10102747   0 0.000
#> 49   49  0.1919378  0.10102747  11 0.011
#> 50   50  0.2753486  0.10102747   9 0.009
#> 51   51  0.3587594  0.10102747   4 0.004
#> 52   52  0.4421702  0.10102747  11 0.011
#> 53   53  0.5255810  0.10102747  12 0.012
#> 54   54  0.6089918  0.10102747   1 0.001
#> 55   55  0.6924026  0.10102747   6 0.006
#> 56   56  0.7758134  0.10102747  12 0.012
#> 57   57  0.8592242  0.10102747  12 0.012
#> 58   58  0.9426350  0.10102747   0 0.000
#> 59   59  1.0260458  0.10102747   0 0.000
#> 60   60  1.1094566  0.10102747   0 0.000
#> 61   61 -0.1000000  0.17326335   0 0.000
#> 62   62 -0.0165892  0.17326335   0 0.000
#> 63   63  0.0668216  0.17326335   0 0.000
#> 64   64  0.1502324  0.17326335   5 0.005
#> 65   65  0.2336432  0.17326335  10 0.010
#> 66   66  0.3170540  0.17326335  10 0.010
#> 67   67  0.4004648  0.17326335  11 0.011
#> 68   68  0.4838756  0.17326335   5 0.005
#> 69   69  0.5672864  0.17326335  14 0.014
#> 70   70  0.6506972  0.17326335  11 0.011
#> 71   71  0.7341080  0.17326335   9 0.009
#> 72   72  0.8175188  0.17326335   6 0.006
#> 73   73  0.9009296  0.17326335   6 0.006
#> 74   74  0.9843404  0.17326335   0 0.000
#> 75   75  1.0677512  0.17326335   0 0.000
#> 76   76 -0.0582946  0.24549922   0 0.000
#> 77   77  0.0251162  0.24549922   0 0.000
#> 78   78  0.1085270  0.24549922   1 0.001
#> 79   79  0.1919378  0.24549922  10 0.010
#> 80   80  0.2753486  0.24549922   5 0.005
#> 81   81  0.3587594  0.24549922   8 0.008
#> 82   82  0.4421702  0.24549922   8 0.008
#> 83   83  0.5255810  0.24549922   1 0.001
#> 84   84  0.6089918  0.24549922   9 0.009
#> 85   85  0.6924026  0.24549922   5 0.005
#> 86   86  0.7758134  0.24549922  11 0.011
#> 87   87  0.8592242  0.24549922   7 0.007
#> 88   88  0.9426350  0.24549922   9 0.009
#> 89   89  1.0260458  0.24549922   0 0.000
#> 90   90  1.1094566  0.24549922   0 0.000
#> 91   91 -0.1000000  0.31773509   0 0.000
#> 92   92 -0.0165892  0.31773509   0 0.000
#> 93   93  0.0668216  0.31773509   0 0.000
#> 94   94  0.1502324  0.31773509  11 0.011
#> 95   95  0.2336432  0.31773509   5 0.005
#> 96   96  0.3170540  0.31773509  10 0.010
#> 97   97  0.4004648  0.31773509   8 0.008
#> 98   98  0.4838756  0.31773509   0 0.000
#> 99   99  0.5672864  0.31773509   0 0.000
#> 100 100  0.6506972  0.31773509  10 0.010
#> 101 101  0.7341080  0.31773509  12 0.012
#> 102 102  0.8175188  0.31773509   2 0.002
#> 103 103  0.9009296  0.31773509  13 0.013
#> 104 104  0.9843404  0.31773509   1 0.001
#> 105 105  1.0677512  0.31773509   0 0.000
#> 106 106 -0.0582946  0.38997096   0 0.000
#> 107 107  0.0251162  0.38997096   0 0.000
#> 108 108  0.1085270  0.38997096   0 0.000
#> 109 109  0.1919378  0.38997096   9 0.009
#> 110 110  0.2753486  0.38997096   3 0.003
#> 111 111  0.3587594  0.38997096   2 0.002
#> 112 112  0.4421702  0.38997096   0 0.000
#> 113 113  0.5255810  0.38997096   0 0.000
#> 114 114  0.6089918  0.38997096   0 0.000
#> 115 115  0.6924026  0.38997096   6 0.006
#> 116 116  0.7758134  0.38997096   9 0.009
#> 117 117  0.8592242  0.38997096  10 0.010
#> 118 118  0.9426350  0.38997096   7 0.007
#> 119 119  1.0260458  0.38997096   0 0.000
#> 120 120  1.1094566  0.38997096   0 0.000
#> 121 121 -0.1000000  0.46220683   0 0.000
#> 122 122 -0.0165892  0.46220683   0 0.000
#> 123 123  0.0668216  0.46220683   0 0.000
#> 124 124  0.1502324  0.46220683   0 0.000
#> 125 125  0.2336432  0.46220683   0 0.000
#> 126 126  0.3170540  0.46220683   0 0.000
#> 127 127  0.4004648  0.46220683   0 0.000
#> 128 128  0.4838756  0.46220683   0 0.000
#> 129 129  0.5672864  0.46220683   0 0.000
#> 130 130  0.6506972  0.46220683   5 0.005
#> 131 131  0.7341080  0.46220683  12 0.012
#> 132 132  0.8175188  0.46220683   6 0.006
#> 133 133  0.9009296  0.46220683   6 0.006
#> 134 134  0.9843404  0.46220683   9 0.009
#> 135 135  1.0677512  0.46220683   0 0.000
#> 136 136 -0.0582946  0.53444270   0 0.000
#> 137 137  0.0251162  0.53444270   0 0.000
#> 138 138  0.1085270  0.53444270   0 0.000
#> 139 139  0.1919378  0.53444270   0 0.000
#> 140 140  0.2753486  0.53444270   0 0.000
#> 141 141  0.3587594  0.53444270   0 0.000
#> 142 142  0.4421702  0.53444270   0 0.000
#> 143 143  0.5255810  0.53444270   0 0.000
#> 144 144  0.6089918  0.53444270   0 0.000
#> 145 145  0.6924026  0.53444270   9 0.009
#> 146 146  0.7758134  0.53444270   9 0.009
#> 147 147  0.8592242  0.53444270   6 0.006
#> 148 148  0.9426350  0.53444270   6 0.006
#> 149 149  1.0260458  0.53444270   1 0.001
#> 150 150  1.1094566  0.53444270   0 0.000
#> 151 151 -0.1000000  0.60667858   0 0.000
#> 152 152 -0.0165892  0.60667858   0 0.000
#> 153 153  0.0668216  0.60667858   0 0.000
#> 154 154  0.1502324  0.60667858   0 0.000
#> 155 155  0.2336432  0.60667858   0 0.000
#> 156 156  0.3170540  0.60667858   0 0.000
#> 157 157  0.4004648  0.60667858   0 0.000
#> 158 158  0.4838756  0.60667858   0 0.000
#> 159 159  0.5672864  0.60667858   0 0.000
#> 160 160  0.6506972  0.60667858  10 0.010
#> 161 161  0.7341080  0.60667858   9 0.009
#> 162 162  0.8175188  0.60667858   5 0.005
#> 163 163  0.9009296  0.60667858   5 0.005
#> 164 164  0.9843404  0.60667858   8 0.008
#> 165 165  1.0677512  0.60667858   0 0.000
#> 166 166 -0.0582946  0.67891445   0 0.000
#> 167 167  0.0251162  0.67891445   0 0.000
#> 168 168  0.1085270  0.67891445   0 0.000
#> 169 169  0.1919378  0.67891445   0 0.000
#> 170 170  0.2753486  0.67891445   0 0.000
#> 171 171  0.3587594  0.67891445   0 0.000
#> 172 172  0.4421702  0.67891445   0 0.000
#> 173 173  0.5255810  0.67891445   1 0.001
#> 174 174  0.6089918  0.67891445  10 0.010
#> 175 175  0.6924026  0.67891445   9 0.009
#> 176 176  0.7758134  0.67891445   6 0.006
#> 177 177  0.8592242  0.67891445  11 0.011
#> 178 178  0.9426350  0.67891445   5 0.005
#> 179 179  1.0260458  0.67891445   0 0.000
#> 180 180  1.1094566  0.67891445   0 0.000
#> 181 181 -0.1000000  0.75115032   0 0.000
#> 182 182 -0.0165892  0.75115032   0 0.000
#> 183 183  0.0668216  0.75115032   0 0.000
#> 184 184  0.1502324  0.75115032   3 0.003
#> 185 185  0.2336432  0.75115032   5 0.005
#> 186 186  0.3170540  0.75115032   3 0.003
#> 187 187  0.4004648  0.75115032  11 0.011
#> 188 188  0.4838756  0.75115032   3 0.003
#> 189 189  0.5672864  0.75115032  12 0.012
#> 190 190  0.6506972  0.75115032   6 0.006
#> 191 191  0.7341080  0.75115032   5 0.005
#> 192 192  0.8175188  0.75115032   8 0.008
#> 193 193  0.9009296  0.75115032   9 0.009
#> 194 194  0.9843404  0.75115032   0 0.000
#> 195 195  1.0677512  0.75115032   0 0.000
#> 196 196 -0.0582946  0.82338619   0 0.000
#> 197 197  0.0251162  0.82338619  12 0.012
#> 198 198  0.1085270  0.82338619   9 0.009
#> 199 199  0.1919378  0.82338619  10 0.010
#> 200 200  0.2753486  0.82338619  11 0.011
#> 201 201  0.3587594  0.82338619   9 0.009
#> 202 202  0.4421702  0.82338619  12 0.012
#> 203 203  0.5255810  0.82338619   7 0.007
#> 204 204  0.6089918  0.82338619   8 0.008
#> 205 205  0.6924026  0.82338619  11 0.011
#> 206 206  0.7758134  0.82338619   5 0.005
#> 207 207  0.8592242  0.82338619   7 0.007
#> 208 208  0.9426350  0.82338619   0 0.000
#> 209 209  1.0260458  0.82338619   0 0.000
#> 210 210  1.1094566  0.82338619   0 0.000
#> 211 211 -0.1000000  0.89562206   0 0.000
#> 212 212 -0.0165892  0.89562206   4 0.004
#> 213 213  0.0668216  0.89562206   8 0.008
#> 214 214  0.1502324  0.89562206   5 0.005
#> 215 215  0.2336432  0.89562206   8 0.008
#> 216 216  0.3170540  0.89562206   8 0.008
#> 217 217  0.4004648  0.89562206   4 0.004
#> 218 218  0.4838756  0.89562206   9 0.009
#> 219 219  0.5672864  0.89562206   7 0.007
#> 220 220  0.6506972  0.89562206  10 0.010
#> 221 221  0.7341080  0.89562206   7 0.007
#> 222 222  0.8175188  0.89562206  15 0.015
#> 223 223  0.9009296  0.89562206   0 0.000
#> 224 224  0.9843404  0.89562206   0 0.000
#> 225 225  1.0677512  0.89562206   0 0.000
#> 226 226 -0.0582946  0.96785794   0 0.000
#> 227 227  0.0251162  0.96785794  12 0.012
#> 228 228  0.1085270  0.96785794   6 0.006
#> 229 229  0.1919378  0.96785794   5 0.005
#> 230 230  0.2753486  0.96785794   5 0.005
#> 231 231  0.3587594  0.96785794   7 0.007
#> 232 232  0.4421702  0.96785794  14 0.014
#> 233 233  0.5255810  0.96785794   3 0.003
#> 234 234  0.6089918  0.96785794   3 0.003
#> 235 235  0.6924026  0.96785794   8 0.008
#> 236 236  0.7758134  0.96785794   7 0.007
#> 237 237  0.8592242  0.96785794   0 0.000
#> 238 238  0.9426350  0.96785794   0 0.000
#> 239 239  1.0260458  0.96785794   0 0.000
#> 240 240  1.1094566  0.96785794   0 0.000
#> 241 241 -0.1000000  1.04009381   0 0.000
#> 242 242 -0.0165892  1.04009381   4 0.004
#> 243 243  0.0668216  1.04009381   9 0.009
#> 244 244  0.1502324  1.04009381   9 0.009
#> 245 245  0.2336432  1.04009381   5 0.005
#> 246 246  0.3170540  1.04009381   8 0.008
#> 247 247  0.4004648  1.04009381   8 0.008
#> 248 248  0.4838756  1.04009381   7 0.007
#> 249 249  0.5672864  1.04009381   5 0.005
#> 250 250  0.6506972  1.04009381   4 0.004
#> 251 251  0.7341080  1.04009381   0 0.000
#> 252 252  0.8175188  1.04009381   0 0.000
#> 253 253  0.9009296  1.04009381   0 0.000
#> 254 254  0.9843404  1.04009381   0 0.000
#> 255 255  1.0677512  1.04009381   0 0.000
#> 256 256 -0.0582946  1.11232968   0 0.000
#> 257 257  0.0251162  1.11232968   4 0.004
#> 258 258  0.1085270  1.11232968  13 0.013
#> 259 259  0.1919378  1.11232968  14 0.014
#> 260 260  0.2753486  1.11232968   6 0.006
#> 261 261  0.3587594  1.11232968   7 0.007
#> 262 262  0.4421702  1.11232968  10 0.010
#> 263 263  0.5255810  1.11232968  10 0.010
#> 264 264  0.6089918  1.11232968   0 0.000
#> 265 265  0.6924026  1.11232968   0 0.000
#> 266 266  0.7758134  1.11232968   0 0.000
#> 267 267  0.8592242  1.11232968   0 0.000
#> 268 268  0.9426350  1.11232968   0 0.000
#> 269 269  1.0260458  1.11232968   0 0.000
#> 270 270  1.1094566  1.11232968   0 0.000
#> 271 271 -0.1000000  1.18456555   0 0.000
#> 272 272 -0.0165892  1.18456555   0 0.000
#> 273 273  0.0668216  1.18456555   3 0.003
#> 274 274  0.1502324  1.18456555   1 0.001
#> 275 275  0.2336432  1.18456555   2 0.002
#> 276 276  0.3170540  1.18456555   1 0.001
#> 277 277  0.4004648  1.18456555   1 0.001
#> 278 278  0.4838756  1.18456555   0 0.000
#> 279 279  0.5672864  1.18456555   0 0.000
#> 280 280  0.6506972  1.18456555   0 0.000
#> 281 281  0.7341080  1.18456555   0 0.000
#> 282 282  0.8175188  1.18456555   0 0.000
#> 283 283  0.9009296  1.18456555   0 0.000
#> 284 284  0.9843404  1.18456555   0 0.000
#> 285 285  1.0677512  1.18456555   0 0.000
#> 286 286 -0.0582946  1.25680142   0 0.000
#> 287 287  0.0251162  1.25680142   0 0.000
#> 288 288  0.1085270  1.25680142   0 0.000
#> 289 289  0.1919378  1.25680142   0 0.000
#> 290 290  0.2753486  1.25680142   0 0.000
#> 291 291  0.3587594  1.25680142   0 0.000
#> 292 292  0.4421702  1.25680142   0 0.000
#> 293 293  0.5255810  1.25680142   0 0.000
#> 294 294  0.6089918  1.25680142   0 0.000
#> 295 295  0.6924026  1.25680142   0 0.000
#> 296 296  0.7758134  1.25680142   0 0.000
#> 297 297  0.8592242  1.25680142   0 0.000
#> 298 298  0.9426350  1.25680142   0 0.000
#> 299 299  1.0260458  1.25680142   0 0.000
#> 300 300  1.1094566  1.25680142   0 0.000