Plot Projected Data with Axes and Circles

Plot Projected Data with Axes and Circles

Usage

plot_proj(
  projected_df,
  model_df,
  axes,
  circle,
  point_param = c(1.5, 0.5, "#000000"),
  line_param = c(0.5, 0.5, "#000000"),
  plot_limits,
  cex = 2,
  position = c(0.92, 0.92),
  axis_text_size = 3,
  is_category = FALSE
)

Arguments

projected_df

A data frame containing the projected data.

model_df

A data frame containing the model reference data.

axes

A data frame or list specifying the axes details.

circle

A list defining circle parameters.

point_param

A vector specifying point size, alpha, and color (default: c(1.5, 0.5, "#000000")).

line_param

A vector specifying line width, alpha, and color (default: c(0.5, 0.5, "#000000")).

plot_limits

Limits for the plot axes.

cex

Scaling factor for point size (default: 2).

position

Position of elements within the plot (default: c(0.92, 0.92)).

axis_text_size

Size of axis text (default: 3).

is_category

Logical indicating if the data is categorical (default: FALSE).

Value

A ggplot object.

Examples

projection_df <- cbind(
c(-0.17353,-0.02906,0.19857,0.00037,0.00131,-0.05019,0.03371),
c(-0.10551,0.14829,-0.02063,0.02658,-0.03150,0.19698,0.00044))

df_bin <- s_curve_obj$s_curve_umap_model_obj$df_bin
edge_data <- s_curve_obj$s_curve_umap_model_tr_from_to_df

proj_obj1 <- get_projection(projection = projection_df, proj_scale = 1,
highd_data = s_curve_noise_training, model_highd = df_bin,
tr_from_to_df = edge_data,
axis_param = list(limits = 1, axis_scaled = 1, axis_pos_x = -0.72,
axis_pos_y = -0.72,threshold = 0))
#> New names:
#> • `` -> `...1`
#> • `` -> `...2`
#> New names:
#> • `` -> `...1`
#> • `` -> `...2`
#> Error in dplyr::left_join(model_df, projected_model_df, by = c(to = "ID")): Input columns in `x` must be unique.
#> ✖ Problem with `proj1_from`, `proj2_from`, `proj1_from`, and `proj2_from`.

projected_df_n <- proj_obj1$projected_df
#> Error: object 'proj_obj1' not found
model_df <- proj_obj1$model_df
#> Error: object 'proj_obj1' not found
axes <- proj_obj1$axes
#> Error: object 'proj_obj1' not found
circle <- proj_obj1$circle
#> Error: object 'proj_obj1' not found

plot_proj(projected_df = projected_df_n, model_df = model_df,
axes = axes, circle = circle, plot_limits = c(-1, 1))
#> Error: object 'projected_df_n' not found