Visualize the model overlaid on high-dimensional data along with 2D wireframe model.

This function generates a LangeviTour visualization based on different conditions and input parameters with 2D wireframe.

Usage

show_link_plots(
  df_all,
  df_b,
  df_b_with_center_data,
  benchmark_value,
  distance_df,
  distance_col,
  use_default_benchmark_val = FALSE,
  col_start = "x",
  type_nldr,
  r2
)

Arguments

df_all

A tibble that contains the high-dimensional data and embedding data.

df_b

A tibble that contains the high-dimensional coordinates of bin centroids.

df_b_with_center_data

The dataset with hexagonal bin centroids.

benchmark_value

The benchmark value used to remove long edges (optional).

distance_df

The tibble with distance.

distance_col

The name of the distance column.

use_default_benchmark_val

Logical, indicating whether to use default benchmark value to remove long edges (default is FALSE).

col_start

The text that begin the column name of the high-dimensional data.

type_nldr

The type of non-linear dimensionality reduction (NLDR) used.

r2

The ratio of the ranges of the original embedding components.

Value

A browsable HTML element.

Examples

df_all <- dplyr::bind_cols(s_curve_noise_training |> dplyr::select(-ID),
umap_data_with_hb_id)
#> Error: object 'umap_data_with_hb_id' not found
df_bin_centroids <- s_curve_obj$s_curve_umap_model_distance_df$df_bin_centroids
#> Warning: Unknown or uninitialised column: `df_bin_centroids`.
df_bin <- s_curve_obj$s_curve_umap_model_distance_df$df_bin
#> Warning: Unknown or uninitialised column: `df_bin`.
distance_df <- s_curve_obj$distance_df
show_link_plots(df_all = df_all, df_b = df_bin, df_b_with_center_data = df_bin_centroids,
benchmark_value = 1.16, distance = distance_df, distance_col = "distance",
use_default_benchmark_val = FALSE, col_start = "x", type_nldr = "UMAP", r2 = r2)
#> Error: object 'df_all' not found