Triangulate bin centroids

This function triangulates the bin centroids using the x and y coordinates provided in the input data frame and returns the triangular object.

Usage

tri_bin_centroids(hex_df, x, y)

Arguments

hex_df

The tibble containing the bin centroids.

x

The name of the column that contains x coordinates of bin centroids.

y

The name of the column that contains y coordinates of bin centroids.

Value

A triangular object representing the triangulated bin centroids.

Examples

df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
tri_bin_centroids(hex_df = df_bin_centroids, x = "c_x", y = "c_y")
#> Warning: triangle collapsed!
#> Warning: three points coincide or are collinear!
#> $trimesh_object
#> Delaunay triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,2,6), [3,2,0]
#> 2: (5,1,6), [1,4,0]
#> 3: (7,6,2), [1,8,6]
#> 4: (13,5,6), [2,5,22]
#> 5: (6,14,13), [12,4,6]
#> 6: (6,7,14), [9,5,3]
#> 7: (8,7,3), [8,14,10]
#> 8: (3,7,2), [3,0,7]
#> 9: (7,15,14), [15,6,10]
#> 10: (8,15,7), [9,7,17]
#> 11: (13,23,34), [23,22,12]
#> 12: (23,13,14), [5,13,11]
#> 13: (14,24,23), [24,12,15]
#> 14: (4,8,3), [7,0,21]
#> 15: (14,15,24), [19,13,9]
#> 16: (15,16,25), [28,19,17]
#> 17: (16,15,8), [10,18,16]
#> 18: (8,9,16), [20,17,21]
#> 19: (15,25,24), [25,15,16]
#> 20: (16,9,17), [36,27,18]
#> 21: (8,4,9), [38,18,14]
#> 22: (13,34,5), [89,4,11]
#> 23: (35,34,23), [11,24,70]
#> 24: (35,23,24), [13,26,23]
#> 25: (24,25,36), [29,26,19]
#> 26: (24,36,35), [55,24,25]
#> 27: (16,17,26), [32,28,20]
#> 28: (16,26,25), [30,16,27]
#> 29: (25,37,36), [58,25,30]
#> 30: (25,26,37), [33,29,28]
#> 31: (26,27,38), [39,33,32]
#> 32: (27,26,17), [27,35,31]
#> 33: (26,38,37), [77,30,31]
#> 34: (27,18,28), [40,41,35]
#> 35: (18,27,17), [32,36,34]
#> 36: (18,17,9), [20,37,35]
#> 37: (9,10,18), [43,36,38]
#> 38: (9,4,10), [45,37,21]
#> 39: (27,39,38), [59,31,41]
#> 40: (18,19,28), [42,34,43]
#> 41: (27,28,39), [46,39,34]
#> 42: (28,19,29), [49,47,40]
#> 43: (18,10,19), [44,40,37]
#> 44: (11,19,10), [43,45,48]
#> 45: (11,10,4), [38,52,44]
#> 46: (28,40,39), [60,41,47]
#> 47: (28,29,40), [53,46,42]
#> 48: (19,11,20), [51,49,44]
#> 49: (19,20,29), [54,42,48]
#> 50: (21,20,12), [51,76,65]
#> 51: (12,20,11), [48,52,50]
#> 52: (12,11,4), [45,0,51]
#> 53: (41,40,29), [47,63,61]
#> 54: (29,20,30), [65,63,49]
#> 55: (62,35,36), [26,71,56]
#> 56: (62,61,35), [70,55,85]
#> 57: (37,64,63), [94,58,77]
#> 58: (36,37,63), [57,71,29]
#> 59: (66,38,39), [39,96,72]
#> 60: (46,39,40), [46,61,96]
#> 61: (46,40,41), [53,62,60]
#> 62: (47,46,41), [61,66,81]
#> 63: (29,30,41), [67,53,54]
#> 64: (31,30,21), [65,75,68]
#> 65: (21,30,20), [54,50,64]
#> 66: (41,42,47), [79,62,67]
#> 67: (41,30,42), [68,66,63]
#> 68: (31,42,30), [67,64,80]
#> 69: (34,61,72), [90,89,70]
#> 70: (35,61,34), [69,23,56]
#> 71: (36,63,62), [86,55,58]
#> 72: (66,65,38), [73,59,104]
#> 73: (65,64,38), [77,72,97]
#> 74: (32,31,22), [75,114,93]
#> 75: (22,31,21), [64,76,74]
#> 76: (22,21,12), [50,0,75]
#> 77: (37,38,64), [73,57,33]
#> 78: (47,48,52), [83,81,79]
#> 79: (48,47,42), [66,84,78]
#> 80: (42,31,43), [93,84,68]
#> 81: (52,46,47), [62,78,102]
#> 82: (46,67,66), [101,96,102]
#> 83: (53,52,48), [78,99,127]
#> 84: (42,43,48), [100,79,80]
#> 85: (73,61,62), [56,87,90]
#> 86: (74,62,63), [71,88,87]
#> 87: (74,73,62), [85,86,106]
#> 88: (75,74,63), [86,94,107]
#> 89: (34,72,5), [0,22,69]
#> 90: (72,61,73), [85,105,69]
#> 91: (76,75,64), [94,97,111]
#> 92: (44,43,32), [93,113,103]
#> 93: (32,43,31), [80,74,92]
#> 94: (75,63,64), [57,91,88]
#> 95: (77,76,65), [97,104,121]
#> 96: (66,39,46), [60,82,59]
#> 97: (76,64,65), [73,95,91]
#> 98: (52,59,68), [143,119,124]
#> 99: (48,49,53), [122,83,100]
#> 100: (48,43,49), [103,99,84]
#> 101: (78,66,67), [82,139,117]
#> 102: (52,67,46), [82,81,119]
#> 103: (44,49,43), [100,92,123]
#> 104: (77,65,66), [72,117,95]
#> 105: (83,72,73), [90,108,135]
#> 106: (84,73,74), [87,109,108]
#> 107: (85,74,75), [88,115,109]
#> 108: (84,83,73), [105,106,129]
#> 109: (85,84,74), [106,107,130]
#> 110: (76,87,86), [145,111,121]
#> 111: (86,75,76), [91,110,115]
#> 112: (45,44,33), [113,0,134]
#> 113: (33,44,32), [92,114,112]
#> 114: (33,32,22), [74,0,113]
#> 115: (86,85,75), [107,111,141]
#> 116: (77,88,87), [142,121,137]
#> 117: (78,77,66), [104,101,137]
#> 118: (79,67,68), [119,149,139]
#> 119: (68,67,52), [102,98,118]
#> 120: (57,53,54), [122,152,127]
#> 121: (76,77,87), [116,110,95]
#> 122: (53,49,54), [126,120,99]
#> 123: (49,44,50), [134,126,103]
#> 124: (57,59,52), [98,127,151]
#> 125: (55,58,57), [151,152,178]
#> 126: (49,50,54), [153,122,123]
#> 127: (57,52,53), [83,120,124]
#> 128: (83,94,93), [158,135,129]
#> 129: (94,83,84), [108,131,128]
#> 130: (95,84,85), [109,132,131]
#> 131: (95,94,84), [129,130,159]
#> 132: (96,95,85), [130,141,160]
#> 133: (51,50,45), [134,180,155]
#> 134: (45,50,44), [123,112,133]
#> 135: (83,93,72), [0,105,128]
#> 136: (97,96,86), [141,145,170]
#> 137: (77,78,88), [147,116,117]
#> 138: (89,78,79), [139,168,147]
#> 139: (79,78,67), [101,118,138]
#> 140: (80,68,69), [143,172,149]
#> 141: (96,85,86), [115,136,132]
#> 142: (87,88,98), [146,154,116]
#> 143: (68,59,69), [164,140,98]
#> 144: (70,59,60), [150,177,164]
#> 145: (97,86,87), [110,154,136]
#> 146: (88,99,98), [171,142,166]
#> 147: (89,88,78), [137,138,166]
#> 148: (90,79,80), [149,181,168]
#> 149: (80,79,68), [118,140,148]
#> 150: (59,58,60), [157,144,151]
#> 151: (59,57,58), [125,150,124]
#> 152: (55,57,54), [120,153,125]
#> 153: (54,50,55), [155,152,126]
#> 154: (87,98,97), [163,145,142]
#> 155: (51,55,50), [153,133,179]
#> 156: (81,69,70), [164,183,172]
#> 157: (71,60,58), [150,176,177]
#> 158: (103,93,94), [128,161,0]
#> 159: (104,94,95), [131,162,161]
#> 160: (105,95,96), [132,165,162]
#> 161: (104,103,94), [158,159,0]
#> 162: (105,104,95), [159,160,0]
#> 163: (97,98,106), [171,170,154]
#> 164: (70,69,59), [143,144,156]
#> 165: (106,105,96), [160,170,0]
#> 166: (88,89,99), [174,146,147]
#> 167: (100,89,90), [168,187,174]
#> 168: (90,89,79), [138,148,167]
#> 169: (91,80,81), [172,189,181]
#> 170: (106,96,97), [136,163,165]
#> 171: (98,99,106), [173,163,146]
#> 172: (81,80,69), [140,156,169]
#> 173: (107,106,99), [171,190,0]
#> 174: (100,99,89), [166,167,190]
#> 175: (101,90,91), [181,192,187]
#> 176: (71,58,56), [178,185,157]
#> 177: (71,70,60), [144,157,182]
#> 178: (58,55,56), [179,176,125]
#> 179: (56,55,51), [155,180,178]
#> 180: (56,45,51), [133,179,0]
#> 181: (91,90,80), [148,169,175]
#> 182: (82,70,71), [177,185,183]
#> 183: (82,81,70), [156,182,184]
#> 184: (92,81,82), [183,0,189]
#> 185: (82,56,71), [176,182,0]
#> 186: (100,101,107), [191,190,187]
#> 187: (101,100,90), [167,175,186]
#> 188: (102,91,92), [189,0,192]
#> 189: (92,91,81), [169,184,188]
#> 190: (107,99,100), [174,186,173]
#> 191: (101,102,107), [0,186,192]
#> 192: (102,101,91), [175,188,191]
#> boundary nodes:  1 2 3 4 12 22 33 45 56 45 56 45 56 45 56 45 56 45 56 45 
#> 
#> $bin_counts
#>   [1] 0.05747126 0.28735632 0.17241379 0.37931034 0.29885057 0.62068966
#>   [7] 0.55172414 0.83908046 0.02298851 0.18390805 0.36781609 0.14942529
#>  [13] 0.59770115 0.50574713 0.43678161 0.45977011 0.19540230 0.29885057
#>  [19] 0.63218391 0.51724138 0.35632184 0.04597701 0.68965517 0.37931034
#>  [25] 0.45977011 0.42528736 0.34482759 0.67816092 0.24137931 0.67816092
#>  [31] 0.45977011 0.41379310 0.02298851 0.57471264 0.45977011 0.18390805
#>  [37] 0.27586207 0.03448276 0.37931034 0.44827586 0.29885057 0.13793103
#>  [43] 0.22988506 0.49425287 0.44827586 0.29885057 0.78160920 0.29885057
#>  [49] 0.21839080 0.44827586 0.56321839 0.25287356 0.52873563 0.47126437
#>  [55] 0.54022989 0.41379310 0.58620690 0.26436782 0.35632184 1.00000000
#>  [61] 0.47126437 0.74712644 0.14942529 0.94252874 0.58620690 0.29885057
#>  [67] 0.47126437 0.47126437 0.73563218 0.35632184 0.35632184 0.28735632
#>  [73] 0.47126437 0.39080460 0.45977011 0.16091954 0.08045977 0.58620690
#>  [79] 0.41379310 0.39080460 0.36781609 0.51724138 0.54022989 0.55172414
#>  [85] 0.36781609 0.21839080 0.63218391 0.31034483 0.17241379 0.35632184
#>  [91] 0.50574713 0.48275862 0.17241379 0.60919540 0.59770115 0.51724138
#>  [97] 0.59770115 0.40229885 0.63218391 0.51724138 0.60919540 0.39080460
#> [103] 0.19540230 0.13793103 0.04597701 0.02298851 0.04597701
#>