Compute Weights for Hexagonal Binning

This function computes weights for hexagonal binning based on the average values of each bin and the distances from these averages.

Usage

compute_weights(nldr_df_with_hex_id)

Arguments

nldr_df_with_hex_id

A data frame with 2D embeddings and hexagonal bin IDs.

Value

A data frame with weights calculated for each hexagonal bin.

Examples

num_bins_x <- calculate_effective_x_bins(nldr_df = s_curve_noise_umap_scaled,
x = "UMAP1", hex_size = NA, buffer_x = NA)
#> Hex size is set to 0.2.
#> Buffer along the x-axis is set to 0.519615242270663.
num_bins_y <- calculate_effective_y_bins(nldr_df = s_curve_noise_umap_scaled,
 y = "UMAP2", hex_size = NA, buffer_y = NA)
#> Hex size is set to 0.2.
#> Buffer along the y-axis is set to 0.45.
hex_bin_obj <- generate_hex_binning_info(nldr_df = s_curve_noise_umap_scaled,
x = "UMAP1", y = "UMAP2", num_bins_x = num_bins_x,
num_bins_y = num_bins_y, x_start = NA, y_start = NA, buffer_x = NA,
buffer_y = NA, hex_size = NA)
#> Hex size is set to 0.2.
#> x_start is set to -0.173205080756888.
#> y_start is set to -0.15.
#> Hex size is set to 0.2.
UMAP_data_with_hb_id <- as.data.frame(do.call(cbind, hex_bin_obj$nldr_data_with_hex_id))
compute_weights(nldr_df_with_hex_id = UMAP_data_with_hb_id)
#> $hb_id
#>   [1]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#>  [26] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#>  [51] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#>  [76]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [101] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [126] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> [151]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [176] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [201] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> [226]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [251] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [276] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> [301]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [326] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [351] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> [376]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [401] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [426] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> [451]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [476] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [501] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> [526]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [551] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [576] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> [601]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [626] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [651] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> [676]  2  2  2  2  6  6  6  6  6  6  6  6  6  7  7  7  7  7  7  7 12 12 12 13 13
#> [701] 13 13 13 13 18 18 18 18 18 18 18 18 18 18 18 18 24 24 24 24 24 24 24 24 28
#> [726] 28 28 28 28 29 29 29 29 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34 34
#> 
#> $UMAP1
#>   [1] 0.16719939 0.16150976 0.13459287 0.16301062 0.08044271 0.07223295
#>   [7] 0.09405763 0.08900367 0.03170275 0.10117880 0.00000000 0.04209521
#>  [13] 0.05349536 0.28575187 0.21053454 0.29256853 0.23125328 0.33548077
#>  [19] 0.25450458 0.29499615 0.26292782 0.05182936 0.13518498 0.60311661
#>  [25] 0.46806684 0.38392618 0.39315580 0.50932690 0.49582819 0.73857864
#>  [31] 0.73364820 0.62693057 0.80998347 0.64241759 0.64916340 0.71326033
#>  [37] 0.57541142 0.62412077 0.64842247 0.68677026 0.77828853 0.83992020
#>  [43] 0.79296306 0.76738169 0.82638316 0.84482412 0.79995094 0.72681720
#>  [49] 0.72192354 0.80783543 0.80668446 0.69339205 0.76831122 0.68029522
#>  [55] 0.99318862 0.88588122 0.92537063 0.91053035 0.83820653 0.90342138
#>  [61] 0.96228699 0.77510396 0.76427076 0.96450414 0.71265285 0.82904448
#>  [67] 0.81275032 0.94245034 0.75504949 0.69843891 0.96227596 0.74175661
#>  [73] 0.94650462 1.00000000 0.80964667 0.16719939 0.16150976 0.13459287
#>  [79] 0.16301062 0.08044271 0.07223295 0.09405763 0.08900367 0.03170275
#>  [85] 0.10117880 0.00000000 0.04209521 0.05349536 0.28575187 0.21053454
#>  [91] 0.29256853 0.23125328 0.33548077 0.25450458 0.29499615 0.26292782
#>  [97] 0.05182936 0.13518498 0.60311661 0.46806684 0.38392618 0.39315580
#> [103] 0.50932690 0.49582819 0.73857864 0.73364820 0.62693057 0.80998347
#> [109] 0.64241759 0.64916340 0.71326033 0.57541142 0.62412077 0.64842247
#> [115] 0.68677026 0.77828853 0.83992020 0.79296306 0.76738169 0.82638316
#> [121] 0.84482412 0.79995094 0.72681720 0.72192354 0.80783543 0.80668446
#> [127] 0.69339205 0.76831122 0.68029522 0.99318862 0.88588122 0.92537063
#> [133] 0.91053035 0.83820653 0.90342138 0.96228699 0.77510396 0.76427076
#> [139] 0.96450414 0.71265285 0.82904448 0.81275032 0.94245034 0.75504949
#> [145] 0.69843891 0.96227596 0.74175661 0.94650462 1.00000000 0.80964667
#> [151] 0.16719939 0.16150976 0.13459287 0.16301062 0.08044271 0.07223295
#> [157] 0.09405763 0.08900367 0.03170275 0.10117880 0.00000000 0.04209521
#> [163] 0.05349536 0.28575187 0.21053454 0.29256853 0.23125328 0.33548077
#> [169] 0.25450458 0.29499615 0.26292782 0.05182936 0.13518498 0.60311661
#> [175] 0.46806684 0.38392618 0.39315580 0.50932690 0.49582819 0.73857864
#> [181] 0.73364820 0.62693057 0.80998347 0.64241759 0.64916340 0.71326033
#> [187] 0.57541142 0.62412077 0.64842247 0.68677026 0.77828853 0.83992020
#> [193] 0.79296306 0.76738169 0.82638316 0.84482412 0.79995094 0.72681720
#> [199] 0.72192354 0.80783543 0.80668446 0.69339205 0.76831122 0.68029522
#> [205] 0.99318862 0.88588122 0.92537063 0.91053035 0.83820653 0.90342138
#> [211] 0.96228699 0.77510396 0.76427076 0.96450414 0.71265285 0.82904448
#> [217] 0.81275032 0.94245034 0.75504949 0.69843891 0.96227596 0.74175661
#> [223] 0.94650462 1.00000000 0.80964667 0.16719939 0.16150976 0.13459287
#> [229] 0.16301062 0.08044271 0.07223295 0.09405763 0.08900367 0.03170275
#> [235] 0.10117880 0.00000000 0.04209521 0.05349536 0.28575187 0.21053454
#> [241] 0.29256853 0.23125328 0.33548077 0.25450458 0.29499615 0.26292782
#> [247] 0.05182936 0.13518498 0.60311661 0.46806684 0.38392618 0.39315580
#> [253] 0.50932690 0.49582819 0.73857864 0.73364820 0.62693057 0.80998347
#> [259] 0.64241759 0.64916340 0.71326033 0.57541142 0.62412077 0.64842247
#> [265] 0.68677026 0.77828853 0.83992020 0.79296306 0.76738169 0.82638316
#> [271] 0.84482412 0.79995094 0.72681720 0.72192354 0.80783543 0.80668446
#> [277] 0.69339205 0.76831122 0.68029522 0.99318862 0.88588122 0.92537063
#> [283] 0.91053035 0.83820653 0.90342138 0.96228699 0.77510396 0.76427076
#> [289] 0.96450414 0.71265285 0.82904448 0.81275032 0.94245034 0.75504949
#> [295] 0.69843891 0.96227596 0.74175661 0.94650462 1.00000000 0.80964667
#> [301] 0.16719939 0.16150976 0.13459287 0.16301062 0.08044271 0.07223295
#> [307] 0.09405763 0.08900367 0.03170275 0.10117880 0.00000000 0.04209521
#> [313] 0.05349536 0.28575187 0.21053454 0.29256853 0.23125328 0.33548077
#> [319] 0.25450458 0.29499615 0.26292782 0.05182936 0.13518498 0.60311661
#> [325] 0.46806684 0.38392618 0.39315580 0.50932690 0.49582819 0.73857864
#> [331] 0.73364820 0.62693057 0.80998347 0.64241759 0.64916340 0.71326033
#> [337] 0.57541142 0.62412077 0.64842247 0.68677026 0.77828853 0.83992020
#> [343] 0.79296306 0.76738169 0.82638316 0.84482412 0.79995094 0.72681720
#> [349] 0.72192354 0.80783543 0.80668446 0.69339205 0.76831122 0.68029522
#> [355] 0.99318862 0.88588122 0.92537063 0.91053035 0.83820653 0.90342138
#> [361] 0.96228699 0.77510396 0.76427076 0.96450414 0.71265285 0.82904448
#> [367] 0.81275032 0.94245034 0.75504949 0.69843891 0.96227596 0.74175661
#> [373] 0.94650462 1.00000000 0.80964667 0.16719939 0.16150976 0.13459287
#> [379] 0.16301062 0.08044271 0.07223295 0.09405763 0.08900367 0.03170275
#> [385] 0.10117880 0.00000000 0.04209521 0.05349536 0.28575187 0.21053454
#> [391] 0.29256853 0.23125328 0.33548077 0.25450458 0.29499615 0.26292782
#> [397] 0.05182936 0.13518498 0.60311661 0.46806684 0.38392618 0.39315580
#> [403] 0.50932690 0.49582819 0.73857864 0.73364820 0.62693057 0.80998347
#> [409] 0.64241759 0.64916340 0.71326033 0.57541142 0.62412077 0.64842247
#> [415] 0.68677026 0.77828853 0.83992020 0.79296306 0.76738169 0.82638316
#> [421] 0.84482412 0.79995094 0.72681720 0.72192354 0.80783543 0.80668446
#> [427] 0.69339205 0.76831122 0.68029522 0.99318862 0.88588122 0.92537063
#> [433] 0.91053035 0.83820653 0.90342138 0.96228699 0.77510396 0.76427076
#> [439] 0.96450414 0.71265285 0.82904448 0.81275032 0.94245034 0.75504949
#> [445] 0.69843891 0.96227596 0.74175661 0.94650462 1.00000000 0.80964667
#> [451] 0.16719939 0.16150976 0.13459287 0.16301062 0.08044271 0.07223295
#> [457] 0.09405763 0.08900367 0.03170275 0.10117880 0.00000000 0.04209521
#> [463] 0.05349536 0.28575187 0.21053454 0.29256853 0.23125328 0.33548077
#> [469] 0.25450458 0.29499615 0.26292782 0.05182936 0.13518498 0.60311661
#> [475] 0.46806684 0.38392618 0.39315580 0.50932690 0.49582819 0.73857864
#> [481] 0.73364820 0.62693057 0.80998347 0.64241759 0.64916340 0.71326033
#> [487] 0.57541142 0.62412077 0.64842247 0.68677026 0.77828853 0.83992020
#> [493] 0.79296306 0.76738169 0.82638316 0.84482412 0.79995094 0.72681720
#> [499] 0.72192354 0.80783543 0.80668446 0.69339205 0.76831122 0.68029522
#> [505] 0.99318862 0.88588122 0.92537063 0.91053035 0.83820653 0.90342138
#> [511] 0.96228699 0.77510396 0.76427076 0.96450414 0.71265285 0.82904448
#> [517] 0.81275032 0.94245034 0.75504949 0.69843891 0.96227596 0.74175661
#> [523] 0.94650462 1.00000000 0.80964667 0.16719939 0.16150976 0.13459287
#> [529] 0.16301062 0.08044271 0.07223295 0.09405763 0.08900367 0.03170275
#> [535] 0.10117880 0.00000000 0.04209521 0.05349536 0.28575187 0.21053454
#> [541] 0.29256853 0.23125328 0.33548077 0.25450458 0.29499615 0.26292782
#> [547] 0.05182936 0.13518498 0.60311661 0.46806684 0.38392618 0.39315580
#> [553] 0.50932690 0.49582819 0.73857864 0.73364820 0.62693057 0.80998347
#> [559] 0.64241759 0.64916340 0.71326033 0.57541142 0.62412077 0.64842247
#> [565] 0.68677026 0.77828853 0.83992020 0.79296306 0.76738169 0.82638316
#> [571] 0.84482412 0.79995094 0.72681720 0.72192354 0.80783543 0.80668446
#> [577] 0.69339205 0.76831122 0.68029522 0.99318862 0.88588122 0.92537063
#> [583] 0.91053035 0.83820653 0.90342138 0.96228699 0.77510396 0.76427076
#> [589] 0.96450414 0.71265285 0.82904448 0.81275032 0.94245034 0.75504949
#> [595] 0.69843891 0.96227596 0.74175661 0.94650462 1.00000000 0.80964667
#> [601] 0.16719939 0.16150976 0.13459287 0.16301062 0.08044271 0.07223295
#> [607] 0.09405763 0.08900367 0.03170275 0.10117880 0.00000000 0.04209521
#> [613] 0.05349536 0.28575187 0.21053454 0.29256853 0.23125328 0.33548077
#> [619] 0.25450458 0.29499615 0.26292782 0.05182936 0.13518498 0.60311661
#> [625] 0.46806684 0.38392618 0.39315580 0.50932690 0.49582819 0.73857864
#> [631] 0.73364820 0.62693057 0.80998347 0.64241759 0.64916340 0.71326033
#> [637] 0.57541142 0.62412077 0.64842247 0.68677026 0.77828853 0.83992020
#> [643] 0.79296306 0.76738169 0.82638316 0.84482412 0.79995094 0.72681720
#> [649] 0.72192354 0.80783543 0.80668446 0.69339205 0.76831122 0.68029522
#> [655] 0.99318862 0.88588122 0.92537063 0.91053035 0.83820653 0.90342138
#> [661] 0.96228699 0.77510396 0.76427076 0.96450414 0.71265285 0.82904448
#> [667] 0.81275032 0.94245034 0.75504949 0.69843891 0.96227596 0.74175661
#> [673] 0.94650462 1.00000000 0.80964667 0.16719939 0.16150976 0.13459287
#> [679] 0.16301062 0.08044271 0.07223295 0.09405763 0.08900367 0.03170275
#> [685] 0.10117880 0.00000000 0.04209521 0.05349536 0.28575187 0.21053454
#> [691] 0.29256853 0.23125328 0.33548077 0.25450458 0.29499615 0.26292782
#> [697] 0.05182936 0.13518498 0.60311661 0.46806684 0.38392618 0.39315580
#> [703] 0.50932690 0.49582819 0.73857864 0.73364820 0.62693057 0.80998347
#> [709] 0.64241759 0.64916340 0.71326033 0.57541142 0.62412077 0.64842247
#> [715] 0.68677026 0.77828853 0.83992020 0.79296306 0.76738169 0.82638316
#> [721] 0.84482412 0.79995094 0.72681720 0.72192354 0.80783543 0.80668446
#> [727] 0.69339205 0.76831122 0.68029522 0.99318862 0.88588122 0.92537063
#> [733] 0.91053035 0.83820653 0.90342138 0.96228699 0.77510396 0.76427076
#> [739] 0.96450414 0.71265285 0.82904448 0.81275032 0.94245034 0.75504949
#> [745] 0.69843891 0.96227596 0.74175661 0.94650462 1.00000000 0.80964667
#> 
#> $UMAP2
#>   [1] 0.037066861 0.009905107 0.000000000 0.030266486 0.274498327 0.218331864
#>   [7] 0.058760856 0.252258947 0.303630086 0.050909321 0.271778584 0.257916294
#>  [13] 0.187982579 0.202379969 0.151933039 0.138015108 0.061042472 0.278688714
#>  [19] 0.063074584 0.117297635 0.341734583 0.320865388 0.335559538 0.597381477
#>  [25] 0.482713207 0.390190144 0.328235991 0.617404710 0.604985028 0.859778569
#>  [31] 0.833935370 0.618409921 0.868291303 0.631646064 0.593427000 0.797282390
#>  [37] 0.708486670 0.712391079 0.661926514 0.778563498 0.839937252 0.929412446
#>  [43] 1.100876246 1.018811832 1.011632238 0.983163036 0.964486583 0.931757885
#>  [49] 0.952956061 1.374772146 1.381019256 1.448210796 1.497157149 1.508729997
#>  [55] 1.521102615 1.408512652 1.455380663 1.445460921 1.720961652 1.600664528
#>  [61] 1.546732900 1.730112607 1.552630670 1.665355817 1.539719959 1.722765733
#>  [67] 1.732050808 1.680924200 1.670893354 1.621642749 1.546125959 1.681926517
#>  [73] 1.524113512 1.625538768 1.529334517 0.037066861 0.009905107 0.000000000
#>  [79] 0.030266486 0.274498327 0.218331864 0.058760856 0.252258947 0.303630086
#>  [85] 0.050909321 0.271778584 0.257916294 0.187982579 0.202379969 0.151933039
#>  [91] 0.138015108 0.061042472 0.278688714 0.063074584 0.117297635 0.341734583
#>  [97] 0.320865388 0.335559538 0.597381477 0.482713207 0.390190144 0.328235991
#> [103] 0.617404710 0.604985028 0.859778569 0.833935370 0.618409921 0.868291303
#> [109] 0.631646064 0.593427000 0.797282390 0.708486670 0.712391079 0.661926514
#> [115] 0.778563498 0.839937252 0.929412446 1.100876246 1.018811832 1.011632238
#> [121] 0.983163036 0.964486583 0.931757885 0.952956061 1.374772146 1.381019256
#> [127] 1.448210796 1.497157149 1.508729997 1.521102615 1.408512652 1.455380663
#> [133] 1.445460921 1.720961652 1.600664528 1.546732900 1.730112607 1.552630670
#> [139] 1.665355817 1.539719959 1.722765733 1.732050808 1.680924200 1.670893354
#> [145] 1.621642749 1.546125959 1.681926517 1.524113512 1.625538768 1.529334517
#> [151] 0.037066861 0.009905107 0.000000000 0.030266486 0.274498327 0.218331864
#> [157] 0.058760856 0.252258947 0.303630086 0.050909321 0.271778584 0.257916294
#> [163] 0.187982579 0.202379969 0.151933039 0.138015108 0.061042472 0.278688714
#> [169] 0.063074584 0.117297635 0.341734583 0.320865388 0.335559538 0.597381477
#> [175] 0.482713207 0.390190144 0.328235991 0.617404710 0.604985028 0.859778569
#> [181] 0.833935370 0.618409921 0.868291303 0.631646064 0.593427000 0.797282390
#> [187] 0.708486670 0.712391079 0.661926514 0.778563498 0.839937252 0.929412446
#> [193] 1.100876246 1.018811832 1.011632238 0.983163036 0.964486583 0.931757885
#> [199] 0.952956061 1.374772146 1.381019256 1.448210796 1.497157149 1.508729997
#> [205] 1.521102615 1.408512652 1.455380663 1.445460921 1.720961652 1.600664528
#> [211] 1.546732900 1.730112607 1.552630670 1.665355817 1.539719959 1.722765733
#> [217] 1.732050808 1.680924200 1.670893354 1.621642749 1.546125959 1.681926517
#> [223] 1.524113512 1.625538768 1.529334517 0.037066861 0.009905107 0.000000000
#> [229] 0.030266486 0.274498327 0.218331864 0.058760856 0.252258947 0.303630086
#> [235] 0.050909321 0.271778584 0.257916294 0.187982579 0.202379969 0.151933039
#> [241] 0.138015108 0.061042472 0.278688714 0.063074584 0.117297635 0.341734583
#> [247] 0.320865388 0.335559538 0.597381477 0.482713207 0.390190144 0.328235991
#> [253] 0.617404710 0.604985028 0.859778569 0.833935370 0.618409921 0.868291303
#> [259] 0.631646064 0.593427000 0.797282390 0.708486670 0.712391079 0.661926514
#> [265] 0.778563498 0.839937252 0.929412446 1.100876246 1.018811832 1.011632238
#> [271] 0.983163036 0.964486583 0.931757885 0.952956061 1.374772146 1.381019256
#> [277] 1.448210796 1.497157149 1.508729997 1.521102615 1.408512652 1.455380663
#> [283] 1.445460921 1.720961652 1.600664528 1.546732900 1.730112607 1.552630670
#> [289] 1.665355817 1.539719959 1.722765733 1.732050808 1.680924200 1.670893354
#> [295] 1.621642749 1.546125959 1.681926517 1.524113512 1.625538768 1.529334517
#> [301] 0.037066861 0.009905107 0.000000000 0.030266486 0.274498327 0.218331864
#> [307] 0.058760856 0.252258947 0.303630086 0.050909321 0.271778584 0.257916294
#> [313] 0.187982579 0.202379969 0.151933039 0.138015108 0.061042472 0.278688714
#> [319] 0.063074584 0.117297635 0.341734583 0.320865388 0.335559538 0.597381477
#> [325] 0.482713207 0.390190144 0.328235991 0.617404710 0.604985028 0.859778569
#> [331] 0.833935370 0.618409921 0.868291303 0.631646064 0.593427000 0.797282390
#> [337] 0.708486670 0.712391079 0.661926514 0.778563498 0.839937252 0.929412446
#> [343] 1.100876246 1.018811832 1.011632238 0.983163036 0.964486583 0.931757885
#> [349] 0.952956061 1.374772146 1.381019256 1.448210796 1.497157149 1.508729997
#> [355] 1.521102615 1.408512652 1.455380663 1.445460921 1.720961652 1.600664528
#> [361] 1.546732900 1.730112607 1.552630670 1.665355817 1.539719959 1.722765733
#> [367] 1.732050808 1.680924200 1.670893354 1.621642749 1.546125959 1.681926517
#> [373] 1.524113512 1.625538768 1.529334517 0.037066861 0.009905107 0.000000000
#> [379] 0.030266486 0.274498327 0.218331864 0.058760856 0.252258947 0.303630086
#> [385] 0.050909321 0.271778584 0.257916294 0.187982579 0.202379969 0.151933039
#> [391] 0.138015108 0.061042472 0.278688714 0.063074584 0.117297635 0.341734583
#> [397] 0.320865388 0.335559538 0.597381477 0.482713207 0.390190144 0.328235991
#> [403] 0.617404710 0.604985028 0.859778569 0.833935370 0.618409921 0.868291303
#> [409] 0.631646064 0.593427000 0.797282390 0.708486670 0.712391079 0.661926514
#> [415] 0.778563498 0.839937252 0.929412446 1.100876246 1.018811832 1.011632238
#> [421] 0.983163036 0.964486583 0.931757885 0.952956061 1.374772146 1.381019256
#> [427] 1.448210796 1.497157149 1.508729997 1.521102615 1.408512652 1.455380663
#> [433] 1.445460921 1.720961652 1.600664528 1.546732900 1.730112607 1.552630670
#> [439] 1.665355817 1.539719959 1.722765733 1.732050808 1.680924200 1.670893354
#> [445] 1.621642749 1.546125959 1.681926517 1.524113512 1.625538768 1.529334517
#> [451] 0.037066861 0.009905107 0.000000000 0.030266486 0.274498327 0.218331864
#> [457] 0.058760856 0.252258947 0.303630086 0.050909321 0.271778584 0.257916294
#> [463] 0.187982579 0.202379969 0.151933039 0.138015108 0.061042472 0.278688714
#> [469] 0.063074584 0.117297635 0.341734583 0.320865388 0.335559538 0.597381477
#> [475] 0.482713207 0.390190144 0.328235991 0.617404710 0.604985028 0.859778569
#> [481] 0.833935370 0.618409921 0.868291303 0.631646064 0.593427000 0.797282390
#> [487] 0.708486670 0.712391079 0.661926514 0.778563498 0.839937252 0.929412446
#> [493] 1.100876246 1.018811832 1.011632238 0.983163036 0.964486583 0.931757885
#> [499] 0.952956061 1.374772146 1.381019256 1.448210796 1.497157149 1.508729997
#> [505] 1.521102615 1.408512652 1.455380663 1.445460921 1.720961652 1.600664528
#> [511] 1.546732900 1.730112607 1.552630670 1.665355817 1.539719959 1.722765733
#> [517] 1.732050808 1.680924200 1.670893354 1.621642749 1.546125959 1.681926517
#> [523] 1.524113512 1.625538768 1.529334517 0.037066861 0.009905107 0.000000000
#> [529] 0.030266486 0.274498327 0.218331864 0.058760856 0.252258947 0.303630086
#> [535] 0.050909321 0.271778584 0.257916294 0.187982579 0.202379969 0.151933039
#> [541] 0.138015108 0.061042472 0.278688714 0.063074584 0.117297635 0.341734583
#> [547] 0.320865388 0.335559538 0.597381477 0.482713207 0.390190144 0.328235991
#> [553] 0.617404710 0.604985028 0.859778569 0.833935370 0.618409921 0.868291303
#> [559] 0.631646064 0.593427000 0.797282390 0.708486670 0.712391079 0.661926514
#> [565] 0.778563498 0.839937252 0.929412446 1.100876246 1.018811832 1.011632238
#> [571] 0.983163036 0.964486583 0.931757885 0.952956061 1.374772146 1.381019256
#> [577] 1.448210796 1.497157149 1.508729997 1.521102615 1.408512652 1.455380663
#> [583] 1.445460921 1.720961652 1.600664528 1.546732900 1.730112607 1.552630670
#> [589] 1.665355817 1.539719959 1.722765733 1.732050808 1.680924200 1.670893354
#> [595] 1.621642749 1.546125959 1.681926517 1.524113512 1.625538768 1.529334517
#> [601] 0.037066861 0.009905107 0.000000000 0.030266486 0.274498327 0.218331864
#> [607] 0.058760856 0.252258947 0.303630086 0.050909321 0.271778584 0.257916294
#> [613] 0.187982579 0.202379969 0.151933039 0.138015108 0.061042472 0.278688714
#> [619] 0.063074584 0.117297635 0.341734583 0.320865388 0.335559538 0.597381477
#> [625] 0.482713207 0.390190144 0.328235991 0.617404710 0.604985028 0.859778569
#> [631] 0.833935370 0.618409921 0.868291303 0.631646064 0.593427000 0.797282390
#> [637] 0.708486670 0.712391079 0.661926514 0.778563498 0.839937252 0.929412446
#> [643] 1.100876246 1.018811832 1.011632238 0.983163036 0.964486583 0.931757885
#> [649] 0.952956061 1.374772146 1.381019256 1.448210796 1.497157149 1.508729997
#> [655] 1.521102615 1.408512652 1.455380663 1.445460921 1.720961652 1.600664528
#> [661] 1.546732900 1.730112607 1.552630670 1.665355817 1.539719959 1.722765733
#> [667] 1.732050808 1.680924200 1.670893354 1.621642749 1.546125959 1.681926517
#> [673] 1.524113512 1.625538768 1.529334517 0.037066861 0.009905107 0.000000000
#> [679] 0.030266486 0.274498327 0.218331864 0.058760856 0.252258947 0.303630086
#> [685] 0.050909321 0.271778584 0.257916294 0.187982579 0.202379969 0.151933039
#> [691] 0.138015108 0.061042472 0.278688714 0.063074584 0.117297635 0.341734583
#> [697] 0.320865388 0.335559538 0.597381477 0.482713207 0.390190144 0.328235991
#> [703] 0.617404710 0.604985028 0.859778569 0.833935370 0.618409921 0.868291303
#> [709] 0.631646064 0.593427000 0.797282390 0.708486670 0.712391079 0.661926514
#> [715] 0.778563498 0.839937252 0.929412446 1.100876246 1.018811832 1.011632238
#> [721] 0.983163036 0.964486583 0.931757885 0.952956061 1.374772146 1.381019256
#> [727] 1.448210796 1.497157149 1.508729997 1.521102615 1.408512652 1.455380663
#> [733] 1.445460921 1.720961652 1.600664528 1.546732900 1.730112607 1.552630670
#> [739] 1.665355817 1.539719959 1.722765733 1.732050808 1.680924200 1.670893354
#> [745] 1.621642749 1.546125959 1.681926517 1.524113512 1.625538768 1.529334517
#> 
#> $weights
#>   [1] 14.146010 16.496449 12.616524 15.947565  8.446605 15.689663  4.927510
#>   [8]  9.890944  6.662428  4.713069  7.188642  9.654327 13.804622  9.146965
#>  [15]  8.924426 13.994068  6.990013  5.043886  7.493630 11.679267  6.123463
#>  [22]  6.717512 15.369061  4.798894 13.872413  5.109337  4.104072  5.923615
#>  [29]  6.514581  5.582020  6.504645  5.353117  4.400122  5.934661  4.926648
#>  [36]  8.943093  6.054813  8.458441  7.231610 11.550482  5.409763  7.941222
#>  [43]  6.087276 11.193041 10.624186  9.531650 13.463551  7.479202  7.938331
#>  [50]  7.255812  7.555217  9.238210  9.281528  6.781595  7.119173  8.682272
#>  [57] 18.503201 13.909167  7.013383  8.922137  5.245151  5.720724  6.121093
#>  [64]  5.816029  4.712344  6.859793  6.287659  6.341032  6.573830  5.003089
#>  [71]  5.235635  5.922047  5.161816  4.952987  6.379268 14.146010 16.496449
#>  [78] 12.616524 15.947565  8.446605 15.689663  4.927510  9.890944  6.662428
#>  [85]  4.713069  7.188642  9.654327 13.804622  9.146965  8.924426 13.994068
#>  [92]  6.990013  5.043886  7.493630 11.679267  6.123463  6.717512 15.369061
#>  [99]  4.798894 13.872413  5.109337  4.104072  5.923615  6.514581  5.582020
#> [106]  6.504645  5.353117  4.400122  5.934661  4.926648  8.943093  6.054813
#> [113]  8.458441  7.231610 11.550482  5.409763  7.941222  6.087276 11.193041
#> [120] 10.624186  9.531650 13.463551  7.479202  7.938331  7.255812  7.555217
#> [127]  9.238210  9.281528  6.781595  7.119173  8.682272 18.503201 13.909167
#> [134]  7.013383  8.922137  5.245151  5.720724  6.121093  5.816029  4.712344
#> [141]  6.859793  6.287659  6.341032  6.573830  5.003089  5.235635  5.922047
#> [148]  5.161816  4.952987  6.379268 14.146010 16.496449 12.616524 15.947565
#> [155]  8.446605 15.689663  4.927510  9.890944  6.662428  4.713069  7.188642
#> [162]  9.654327 13.804622  9.146965  8.924426 13.994068  6.990013  5.043886
#> [169]  7.493630 11.679267  6.123463  6.717512 15.369061  4.798894 13.872413
#> [176]  5.109337  4.104072  5.923615  6.514581  5.582020  6.504645  5.353117
#> [183]  4.400122  5.934661  4.926648  8.943093  6.054813  8.458441  7.231610
#> [190] 11.550482  5.409763  7.941222  6.087276 11.193041 10.624186  9.531650
#> [197] 13.463551  7.479202  7.938331  7.255812  7.555217  9.238210  9.281528
#> [204]  6.781595  7.119173  8.682272 18.503201 13.909167  7.013383  8.922137
#> [211]  5.245151  5.720724  6.121093  5.816029  4.712344  6.859793  6.287659
#> [218]  6.341032  6.573830  5.003089  5.235635  5.922047  5.161816  4.952987
#> [225]  6.379268 14.146010 16.496449 12.616524 15.947565  8.446605 15.689663
#> [232]  4.927510  9.890944  6.662428  4.713069  7.188642  9.654327 13.804622
#> [239]  9.146965  8.924426 13.994068  6.990013  5.043886  7.493630 11.679267
#> [246]  6.123463  6.717512 15.369061  4.798894 13.872413  5.109337  4.104072
#> [253]  5.923615  6.514581  5.582020  6.504645  5.353117  4.400122  5.934661
#> [260]  4.926648  8.943093  6.054813  8.458441  7.231610 11.550482  5.409763
#> [267]  7.941222  6.087276 11.193041 10.624186  9.531650 13.463551  7.479202
#> [274]  7.938331  7.255812  7.555217  9.238210  9.281528  6.781595  7.119173
#> [281]  8.682272 18.503201 13.909167  7.013383  8.922137  5.245151  5.720724
#> [288]  6.121093  5.816029  4.712344  6.859793  6.287659  6.341032  6.573830
#> [295]  5.003089  5.235635  5.922047  5.161816  4.952987  6.379268 14.146010
#> [302] 16.496449 12.616524 15.947565  8.446605 15.689663  4.927510  9.890944
#> [309]  6.662428  4.713069  7.188642  9.654327 13.804622  9.146965  8.924426
#> [316] 13.994068  6.990013  5.043886  7.493630 11.679267  6.123463  6.717512
#> [323] 15.369061  4.798894 13.872413  5.109337  4.104072  5.923615  6.514581
#> [330]  5.582020  6.504645  5.353117  4.400122  5.934661  4.926648  8.943093
#> [337]  6.054813  8.458441  7.231610 11.550482  5.409763  7.941222  6.087276
#> [344] 11.193041 10.624186  9.531650 13.463551  7.479202  7.938331  7.255812
#> [351]  7.555217  9.238210  9.281528  6.781595  7.119173  8.682272 18.503201
#> [358] 13.909167  7.013383  8.922137  5.245151  5.720724  6.121093  5.816029
#> [365]  4.712344  6.859793  6.287659  6.341032  6.573830  5.003089  5.235635
#> [372]  5.922047  5.161816  4.952987  6.379268 14.146010 16.496449 12.616524
#> [379] 15.947565  8.446605 15.689663  4.927510  9.890944  6.662428  4.713069
#> [386]  7.188642  9.654327 13.804622  9.146965  8.924426 13.994068  6.990013
#> [393]  5.043886  7.493630 11.679267  6.123463  6.717512 15.369061  4.798894
#> [400] 13.872413  5.109337  4.104072  5.923615  6.514581  5.582020  6.504645
#> [407]  5.353117  4.400122  5.934661  4.926648  8.943093  6.054813  8.458441
#> [414]  7.231610 11.550482  5.409763  7.941222  6.087276 11.193041 10.624186
#> [421]  9.531650 13.463551  7.479202  7.938331  7.255812  7.555217  9.238210
#> [428]  9.281528  6.781595  7.119173  8.682272 18.503201 13.909167  7.013383
#> [435]  8.922137  5.245151  5.720724  6.121093  5.816029  4.712344  6.859793
#> [442]  6.287659  6.341032  6.573830  5.003089  5.235635  5.922047  5.161816
#> [449]  4.952987  6.379268 14.146010 16.496449 12.616524 15.947565  8.446605
#> [456] 15.689663  4.927510  9.890944  6.662428  4.713069  7.188642  9.654327
#> [463] 13.804622  9.146965  8.924426 13.994068  6.990013  5.043886  7.493630
#> [470] 11.679267  6.123463  6.717512 15.369061  4.798894 13.872413  5.109337
#> [477]  4.104072  5.923615  6.514581  5.582020  6.504645  5.353117  4.400122
#> [484]  5.934661  4.926648  8.943093  6.054813  8.458441  7.231610 11.550482
#> [491]  5.409763  7.941222  6.087276 11.193041 10.624186  9.531650 13.463551
#> [498]  7.479202  7.938331  7.255812  7.555217  9.238210  9.281528  6.781595
#> [505]  7.119173  8.682272 18.503201 13.909167  7.013383  8.922137  5.245151
#> [512]  5.720724  6.121093  5.816029  4.712344  6.859793  6.287659  6.341032
#> [519]  6.573830  5.003089  5.235635  5.922047  5.161816  4.952987  6.379268
#> [526] 14.146010 16.496449 12.616524 15.947565  8.446605 15.689663  4.927510
#> [533]  9.890944  6.662428  4.713069  7.188642  9.654327 13.804622  9.146965
#> [540]  8.924426 13.994068  6.990013  5.043886  7.493630 11.679267  6.123463
#> [547]  6.717512 15.369061  4.798894 13.872413  5.109337  4.104072  5.923615
#> [554]  6.514581  5.582020  6.504645  5.353117  4.400122  5.934661  4.926648
#> [561]  8.943093  6.054813  8.458441  7.231610 11.550482  5.409763  7.941222
#> [568]  6.087276 11.193041 10.624186  9.531650 13.463551  7.479202  7.938331
#> [575]  7.255812  7.555217  9.238210  9.281528  6.781595  7.119173  8.682272
#> [582] 18.503201 13.909167  7.013383  8.922137  5.245151  5.720724  6.121093
#> [589]  5.816029  4.712344  6.859793  6.287659  6.341032  6.573830  5.003089
#> [596]  5.235635  5.922047  5.161816  4.952987  6.379268 14.146010 16.496449
#> [603] 12.616524 15.947565  8.446605 15.689663  4.927510  9.890944  6.662428
#> [610]  4.713069  7.188642  9.654327 13.804622  9.146965  8.924426 13.994068
#> [617]  6.990013  5.043886  7.493630 11.679267  6.123463  6.717512 15.369061
#> [624]  4.798894 13.872413  5.109337  4.104072  5.923615  6.514581  5.582020
#> [631]  6.504645  5.353117  4.400122  5.934661  4.926648  8.943093  6.054813
#> [638]  8.458441  7.231610 11.550482  5.409763  7.941222  6.087276 11.193041
#> [645] 10.624186  9.531650 13.463551  7.479202  7.938331  7.255812  7.555217
#> [652]  9.238210  9.281528  6.781595  7.119173  8.682272 18.503201 13.909167
#> [659]  7.013383  8.922137  5.245151  5.720724  6.121093  5.816029  4.712344
#> [666]  6.859793  6.287659  6.341032  6.573830  5.003089  5.235635  5.922047
#> [673]  5.161816  4.952987  6.379268 14.146010 16.496449 12.616524 15.947565
#> [680]  8.446605 15.689663  4.927510  9.890944  6.662428  4.713069  7.188642
#> [687]  9.654327 13.804622  9.146965  8.924426 13.994068  6.990013  5.043886
#> [694]  7.493630 11.679267  6.123463  6.717512 15.369061  4.798894 13.872413
#> [701]  5.109337  4.104072  5.923615  6.514581  5.582020  6.504645  5.353117
#> [708]  4.400122  5.934661  4.926648  8.943093  6.054813  8.458441  7.231610
#> [715] 11.550482  5.409763  7.941222  6.087276 11.193041 10.624186  9.531650
#> [722] 13.463551  7.479202  7.938331  7.255812  7.555217  9.238210  9.281528
#> [729]  6.781595  7.119173  8.682272 18.503201 13.909167  7.013383  8.922137
#> [736]  5.245151  5.720724  6.121093  5.816029  4.712344  6.859793  6.287659
#> [743]  6.341032  6.573830  5.003089  5.235635  5.922047  5.161816  4.952987
#> [750]  6.379268
#>