Extract hexagonal bin centroids coordinates and the corresponding standardise counts.

Extract hexagonal bin centroids coordinates and the corresponding standardise counts.

Usage

extract_hexbin_centroids(centroids_df, counts_df)

Arguments

centroids_df

A tibble that contains all hexagonal bin centroid coordinates with hexagon IDs.

counts_df

A tibble that contains hexagon IDs with the standardise number of points within each hexagon.

Value

A tibble contains hexagon ID, centroid coordinates, and standardise counts.

Examples

all_centroids_df <- s_curve_obj$s_curve_umap_hb_obj$centroids
counts_df <- s_curve_obj$s_curve_umap_hb_obj$std_cts
extract_hexbin_centroids(centroids_df = all_centroids_df,
counts_df = counts_df)
#> # A tibble: 20 × 6
#>    hexID     c_x     c_y bin_counts std_counts drop_empty
#>    <int>   <dbl>   <dbl>      <int>      <dbl> <lgl>     
#>  1     1 -0.1    -0.0938         17     0.0302 FALSE     
#>  2     2  0.227  -0.0938        169     0.300  FALSE     
#>  3     3  0.554  -0.0938         34     0.0604 FALSE     
#>  4     4  0.880  -0.0938         NA    NA      TRUE      
#>  5     5  0.0634  0.189         431     0.766  FALSE     
#>  6     6  0.390   0.189         430     0.764  FALSE     
#>  7     7  0.717   0.189         407     0.723  FALSE     
#>  8     8  1.04    0.189          57     0.101  FALSE     
#>  9     9 -0.1     0.472          NA    NA      TRUE      
#> 10    10  0.227   0.472          NA    NA      TRUE      
#> 11    11  0.554   0.472          82     0.146  FALSE     
#> 12    12  0.880   0.472         367     0.652  FALSE     
#> 13    13  0.0634  0.755         442     0.785  FALSE     
#> 14    14  0.390   0.755         478     0.849  FALSE     
#> 15    15  0.717   0.755         563     1      FALSE     
#> 16    16  1.04    0.755          NA    NA      TRUE      
#> 17    17 -0.1     1.04           26     0.0462 FALSE     
#> 18    18  0.227   1.04          149     0.265  FALSE     
#> 19    19  0.554   1.04           98     0.174  FALSE     
#> 20    20  0.880   1.04           NA    NA      TRUE