Extract hexagonal bin mean coordinates and the corresponding standardize counts.

Extract hexagonal bin mean coordinates and the corresponding standardize counts.

Usage

extract_hexbin_mean(data_hb, counts_df)

Arguments

data_hb

A tibble with embedding components and hexagonal bin IDs.

counts_df

A tibble that contains hexagon IDs with the standardise number of points within each hexagon.

Value

A tibble contains hexagon ID, bin mean coordinates, and standardize counts.

Examples

range_umap2 <- diff(range(s_curve_noise_umap$UMAP2))
num_bins_x <- 3
hb_obj <- hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x,
s1 = -0.1, s2 = -0.1, r2 = range_umap2)
umap_with_hb_id <- hb_obj$data_hb_id
counts_df <- hb_obj$std_cts
extract_hexbin_mean(data_hb = umap_with_hb_id, counts_df = counts_df)
#> # A tibble: 9 × 4
#>   hexID   c_x    c_y std_counts
#>   <int> <dbl>  <dbl>      <dbl>
#> 1     1 0.137 0.0312     0.375 
#> 2     2 0.260 0.0805     0.188 
#> 3     4 0.170 0.266      1     
#> 4     5 0.598 0.585      0.312 
#> 5     8 0.590 0.681      0.375 
#> 6     9 0.775 0.916      0.75  
#> 7    11 0.796 1.44       0.75  
#> 8    14 0.698 1.62       0.0625
#> 9    15 0.888 1.64       0.875