Construct the 2D model and lift into high-D

This function fits a high-dimensional model using hexagonal bins and provides options to customize the modeling process, including the choice of bin centroids or bin means, removal of low-density hexagons, and averaging of high-dimensional data.

Usage

fit_highd_model(
  highd_data,
  nldr_data,
  bin1 = 4,
  r2,
  q = 0.1,
  is_bin_centroid = TRUE
)

Arguments

highd_data

A tibble that contains the training high-dimensional data.

nldr_data

A tibble that contains embedding with a unique identifier.

bin1

Number of bins along the x axis.

r2

The ratio of the ranges of the original embedding components.

q

The buffer amount as proportion of data range.

Value

A list containing the data frame with high-dimensional coordinates for 2D bin centroids (df_bin) and the data frame containing information about hexagonal bin centroids (df_bin_centroids) in 2D.

Examples

scurve_umap_scaled_obj <- s_curve_obj$s_curve_umap_scaled_obj
lim1 <- scurve_umap_scaled_obj$lim1
lim2 <- scurve_umap_scaled_obj$lim2
r2 <- diff(lim2)/diff(lim1)
fit_highd_model(highd_data = s_curve_noise_training,
nldr_data = s_curve_noise_umap_scaled, bin1 = 4, r2 = r2)
#> ✔ Model generated successfully! 🎉
#> $df_bin
#> # A tibble: 15 × 8
#>    hb_id      x1    x2       x3         x4         x5        x6          x7
#>    <int>   <dbl> <dbl>    <dbl>      <dbl>      <dbl>     <dbl>       <dbl>
#>  1     1  0.979  1.90   1.17    -0.000726   0.0000519 -0.0292   -0.000946  
#>  2     2  0.164  1.70   1.92    -0.00134    0.00123    0.00717  -0.000111  
#>  3     3 -0.736  1.88   1.65     0.000779   0.000445   0.000375  0.00120   
#>  4     5  0.760  0.878  1.55    -0.0000296 -0.000455  -0.00283  -0.000161  
#>  5     6 -0.442  0.732  1.77    -0.000256  -0.000205  -0.00290   0.000141  
#>  6     7 -0.885  1.25   0.736    0.000338   0.000525  -0.00174   0.000172  
#>  7     8 -0.299  1.81   0.0673  -0.00201   -0.000363  -0.00581   0.0000764 
#>  8    11 -0.894  0.188  0.678   -0.0000786 -0.000632  -0.00585   0.000328  
#>  9    12 -0.0121 0.886  0.00428 -0.000607   0.00105    0.00324   0.000302  
#> 10    13 -0.749  0.876 -1.55    -0.000230  -0.000361   0.000102 -0.0000497 
#> 11    14  0.440  0.761 -1.81    -0.000471   0.000461  -0.00361  -0.00000919
#> 12    15  0.867  0.957 -0.676    0.000797   0.000363  -0.00109   0.000304  
#> 13    17 -0.991  1.82  -1.11    -0.00117   -0.00109   -0.00551   0.000819  
#> 14    18 -0.284  1.74  -1.90     0.000209   0.00106    0.000850 -0.000549  
#> 15    19  0.876  1.70  -1.43    -0.00108   -0.00130    0.00220  -0.000737  
#> 
#> $df_bin_centroids
#> # A tibble: 15 × 6
#>    hexID     c_x     c_y bin_counts std_counts drop_empty
#>    <int>   <dbl>   <dbl>      <int>      <dbl> <lgl>     
#>  1     1 -0.1    -0.0938         17     0.0302 FALSE     
#>  2     2  0.227  -0.0938        169     0.300  FALSE     
#>  3     3  0.554  -0.0938         34     0.0604 FALSE     
#>  4     5  0.0634  0.189         431     0.766  FALSE     
#>  5     6  0.390   0.189         430     0.764  FALSE     
#>  6     7  0.717   0.189         407     0.723  FALSE     
#>  7     8  1.04    0.189          57     0.101  FALSE     
#>  8    11  0.554   0.472          82     0.146  FALSE     
#>  9    12  0.880   0.472         367     0.652  FALSE     
#> 10    13  0.0634  0.755         442     0.785  FALSE     
#> 11    14  0.390   0.755         478     0.849  FALSE     
#> 12    15  0.717   0.755         563     1      FALSE     
#> 13    17 -0.1     1.04           26     0.0462 FALSE     
#> 14    18  0.227   1.04          149     0.265  FALSE     
#> 15    19  0.554   1.04           98     0.174  FALSE     
#>