Construct the 2D model and lift into high-D

This function fits a high-dimensional model using hexagonal bins and provides options to customize the modeling process, including the choice of bin centroids or bin means, removal of low-density hexagons, and averaging of high-dimensional data.

Usage

fit_highd_model(
  highd_data,
  nldr_data,
  bin1 = 4,
  r2,
  q = 0.1,
  is_bin_centroid = TRUE
)

Arguments

highd_data

A tibble that contains the training high-dimensional data.

nldr_data

A tibble that contains embedding with a unique identifier.

bin1

Number of bins along the x axis.

r2

The ratio of the ranges of the original embedding components.

q

The buffer amount as proportion of data range.

Value

A list containing the data frame with high-dimensional coordinates for 2D bin centroids (df_bin) and the data frame containing information about hexagonal bin centroids (df_bin_centroids) in 2D.

Examples

scurve_umap_scaled_obj <- s_curve_obj$s_curve_umap_scaled_obj
lim1 <- scurve_umap_scaled_obj$lim1
lim2 <- scurve_umap_scaled_obj$lim2
r2 <- diff(lim2)/diff(lim1)
fit_highd_model(highd_data = s_curve_noise_training,
nldr_data = s_curve_noise_umap_scaled, bin1 = 4, r2 = r2)
#> ✔ Model generated successfully! 🎉
#> $df_bin
#> # A tibble: 15 × 8
#>    hb_id     x1    x2      x3         x4        x5         x6         x7
#>    <int>  <dbl> <dbl>   <dbl>      <dbl>     <dbl>      <dbl>      <dbl>
#>  1     2  0.856 0.201  1.43   -0.000535   0.000684 -0.000925   0.000230 
#>  2     3 -0.440 0.192  1.88    0.000338  -0.000838 -0.00858   -0.000574 
#>  3     5  0.945 1.10   1.28    0.000149  -0.000608  0.00579   -0.000326 
#>  4     6  0.157 0.974  1.89   -0.000520   0.000437 -0.00375    0.0000215
#>  5     7 -0.904 0.826  1.13    0.000804  -0.000838  0.00109    0.000116 
#>  6     8 -0.727 0.318  0.337  -0.00258    0.000865 -0.00693    0.000145 
#>  7    10 -0.163 0.973 -0.142  -0.000960  -0.000414 -0.00203   -0.000240 
#>  8    11 -0.494 1.82   1.15    0.000316   0.00126  -0.0000240  0.000738 
#>  9    12 -0.101 1.15   0.0409 -0.000504   0.000973  0.00120    0.000415 
#> 10    13 -0.848 1.36  -1.44   -0.000112  -0.000803 -0.00210   -0.0000712
#> 11    14  0.186 1.02  -1.89    0.0000167  0.000649 -0.000965  -0.000124 
#> 12    15  0.911 0.961 -1.05   -0.000322   0.000368 -0.00437    0.0000264
#> 13    16  0.713 0.324 -0.341   0.00182   -0.000960  0.00526   -0.000172 
#> 14    18 -0.212 1.82  -1.95   -0.000153   0.000124 -0.00374   -0.000188 
#> 15    19  0.761 0.761 -1.65    0.000556   0.000606  0.00905   -0.000400 
#> 
#> $df_bin_centroids
#> # A tibble: 15 × 6
#>    hexID    c_x     c_y bin_counts std_counts drop_empty
#>    <int>  <dbl>   <dbl>      <int>      <dbl> <lgl>     
#>  1     2 0.210  -0.0885         98     0.179  FALSE     
#>  2     3 0.520  -0.0885         36     0.0659 FALSE     
#>  3     5 0.0549  0.180         132     0.242  FALSE     
#>  4     6 0.365   0.180         463     0.848  FALSE     
#>  5     7 0.675   0.180         425     0.778  FALSE     
#>  6     8 0.985   0.180          77     0.141  FALSE     
#>  7    10 0.210   0.448         236     0.432  FALSE     
#>  8    11 0.520   0.448         181     0.332  FALSE     
#>  9    12 0.830   0.448         545     0.998  FALSE     
#> 10    13 0.0549  0.717         284     0.520  FALSE     
#> 11    14 0.365   0.717         519     0.951  FALSE     
#> 12    15 0.675   0.717         546     1      FALSE     
#> 13    16 0.985   0.717         130     0.238  FALSE     
#> 14    18 0.210   0.985          62     0.114  FALSE     
#> 15    19 0.520   0.985          16     0.0293 FALSE     
#>