Construct the 2D model and lift into high-D

This function fits a high-dimensional model using hexagonal bins and provides options to customize the modeling process, including the choice of bin centroids or bin means, removal of low-density hexagons, and averaging of high-dimensional data.

Usage

fit_highd_model(
  training_data,
  emb_df,
  bin1 = 2,
  s1 = -0.1,
  s2 = -0.1,
  r2,
  is_bin_centroid = TRUE,
  is_rm_lwd_hex = FALSE,
  benchmark_to_rm_lwd_hex,
  col_start_highd = "x"
)

Arguments

training_data

A tibble that contains the training high-dimensional data.

emb_df

A tibble that contains embedding with a unique identifier.

bin1

Number of bins along the x axis.

s1

The x-coordinate of the hexagonal grid starting point.

s2

The y-coordinate of the hexagonal grid starting point.

r2

The ratio of the ranges of the original embedding components.

is_bin_centroid

Logical, indicating whether to use bin centroids (default is TRUE).

is_rm_lwd_hex

Logical, indicating whether to remove low-density hexagons (default is FALSE).

benchmark_to_rm_lwd_hex

The benchmark value to remove low-density hexagons.

col_start_highd

The text prefix for columns in the high-dimensional data.

Value

A list containing the data frame with high-dimensional coordinates for 2D bin centroids (df_bin) and the data frame containing information about hexagonal bin centroids (df_bin_centroids) in 2D.

Examples

r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
fit_highd_model(training_data = s_curve_noise_training,
emb_df = s_curve_noise_umap_scaled, bin1 = 3, r2 = r2,
col_start_highd = "x")
#> $df_bin
#> # A tibble: 8 × 8
#>   hb_id      x1    x2      x3        x4        x5       x6         x7
#>   <int>   <dbl> <dbl>   <dbl>     <dbl>     <dbl>    <dbl>      <dbl>
#> 1     1 -0.909  1.46  -1.33    0.00547   0.0125   -0.0392   0.00177  
#> 2     2 -0.265  1.70  -1.88    0.00608  -0.00651   0.00410 -0.00271  
#> 3     4  0.0831 0.499 -1.82    0.00335  -0.000775 -0.0175  -0.000610 
#> 4     5  0.927  1.07  -0.926   0.00824  -0.00342   0.0608   0.00257  
#> 5     8  0.714  1.01  -0.335  -0.00495   0.00580   0.0234   0.000646 
#> 6     9 -0.127  1.11   0.0697  0.00389   0.00286   0.00384  0.0000872
#> 7    11 -0.841  1.66   0.647   0.00380  -0.00955  -0.00463  0.00422  
#> 8    15  0.291  1.05   1.71   -0.000141  0.00314  -0.00694 -0.00105  
#> 
#> $df_bin_centroids
#> # A tibble: 8 × 4
#>   hexID    c_x    c_y std_counts
#>   <int>  <dbl>  <dbl>      <dbl>
#> 1     1 -0.1   -0.1        0.115
#> 2     2  0.389 -0.1        0.231
#> 3     4  0.145  0.324      0.577
#> 4     5  0.634  0.324      0.115
#> 5     8  0.389  0.748      0.231
#> 6     9  0.879  0.748      0.462
#> 7    11  0.634  1.17       0.154
#> 8    15  0.879  1.60       1    
#>