Generate hexagonal polygon coordinates

This function generates the coordinates of hexagons after passing hexagonal centroids.

Usage

gen_hex_coord(centroids_df, hex_size = NA)

Arguments

centroids_df

The dataset with all hexbin ID and centroid coordinates.

hex_size

A numeric value that initializes the radius of the outer circle surrounding the hexagon.

Value

A list contains polygon id, x and y coordinates (hex_poly_id, x, and y respectively) of hexagons.

Examples

num_bins_list <- calc_bins(data = s_curve_noise_umap_scaled, x = "UMAP1",
y = "UMAP2", hex_size = NA, buffer_x = NA, buffer_y = NA)
#> Hex size is set to 0.2.
#> Buffer along the x-axis is set to 0.519615242270663.
#> Buffer along the y-axis is set to 0.45.
num_bins_x <- num_bins_list$num_x
num_bins_y <- num_bins_list$num_y
centroid_list <- gen_centroids(data = s_curve_noise_umap_scaled,
x = "UMAP1", y = "UMAP2", num_bins_x = num_bins_x,
num_bins_y = num_bins_y, x_start = NA, y_start = NA, buffer_x = NA,
buffer_y = NA, hex_size = NA)
#> Hex size is set to 0.2.
#> x_start is set to -0.173205080756888.
#> y_start is set to -0.15.
all_centroids_df <- as.data.frame(do.call(cbind, centroid_list))
gen_hex_coord(centroids_df = all_centroids_df, hex_size = NA)
#> Hex size is set to 0.2.
#> $hex_poly_id
#>   [1]  1  1  1  1  1  1  2  2  2  2  2  2  3  3  3  3  3  3  4  4  4  4  4  4  5
#>  [26]  5  5  5  5  5  6  6  6  6  6  6  7  7  7  7  7  7  8  8  8  8  8  8  9  9
#>  [51]  9  9  9  9 10 10 10 10 10 10 11 11 11 11 11 11 12 12 12 12 12 12 13 13 13
#>  [76] 13 13 13 14 14 14 14 14 14 15 15 15 15 15 15 16 16 16 16 16 16 17 17 17 17
#> [101] 17 17 18 18 18 18 18 18 19 19 19 19 19 19 20 20 20 20 20 20 21 21 21 21 21
#> [126] 21 22 22 22 22 22 22 23 23 23 23 23 23 24 24 24 24 24 24 25 25 25 25 25 25
#> [151] 26 26 26 26 26 26 27 27 27 27 27 27 28 28 28 28 28 28 29 29 29 29 29 29 30
#> [176] 30 30 30 30 30 31 31 31 31 31 31 32 32 32 32 32 32 33 33 33 33 33 33 34 34
#> [201] 34 34 34 34 35 35 35 35 35 35 36 36 36 36 36 36 37 37 37 37 37 37 38 38 38
#> [226] 38 38 38 39 39 39 39 39 39 40 40 40 40 40 40
#> 
#> $x
#>   [1]  0.0000000  0.0000000 -0.1732051 -0.3464102 -0.3464102 -0.1732051
#>   [7]  0.3464102  0.3464102  0.1732051  0.0000000  0.0000000  0.1732051
#>  [13]  0.6928203  0.6928203  0.5196152  0.3464102  0.3464102  0.5196152
#>  [19]  1.0392305  1.0392305  0.8660254  0.6928203  0.6928203  0.8660254
#>  [25]  1.3856406  1.3856406  1.2124356  1.0392305  1.0392305  1.2124356
#>  [31]  0.1732051  0.1732051  0.0000000 -0.1732051 -0.1732051  0.0000000
#>  [37]  0.5196152  0.5196152  0.3464102  0.1732051  0.1732051  0.3464102
#>  [43]  0.8660254  0.8660254  0.6928203  0.5196152  0.5196152  0.6928203
#>  [49]  1.2124356  1.2124356  1.0392305  0.8660254  0.8660254  1.0392305
#>  [55]  1.5588457  1.5588457  1.3856406  1.2124356  1.2124356  1.3856406
#>  [61]  0.0000000  0.0000000 -0.1732051 -0.3464102 -0.3464102 -0.1732051
#>  [67]  0.3464102  0.3464102  0.1732051  0.0000000  0.0000000  0.1732051
#>  [73]  0.6928203  0.6928203  0.5196152  0.3464102  0.3464102  0.5196152
#>  [79]  1.0392305  1.0392305  0.8660254  0.6928203  0.6928203  0.8660254
#>  [85]  1.3856406  1.3856406  1.2124356  1.0392305  1.0392305  1.2124356
#>  [91]  0.1732051  0.1732051  0.0000000 -0.1732051 -0.1732051  0.0000000
#>  [97]  0.5196152  0.5196152  0.3464102  0.1732051  0.1732051  0.3464102
#> [103]  0.8660254  0.8660254  0.6928203  0.5196152  0.5196152  0.6928203
#> [109]  1.2124356  1.2124356  1.0392305  0.8660254  0.8660254  1.0392305
#> [115]  1.5588457  1.5588457  1.3856406  1.2124356  1.2124356  1.3856406
#> [121]  0.0000000  0.0000000 -0.1732051 -0.3464102 -0.3464102 -0.1732051
#> [127]  0.3464102  0.3464102  0.1732051  0.0000000  0.0000000  0.1732051
#> [133]  0.6928203  0.6928203  0.5196152  0.3464102  0.3464102  0.5196152
#> [139]  1.0392305  1.0392305  0.8660254  0.6928203  0.6928203  0.8660254
#> [145]  1.3856406  1.3856406  1.2124356  1.0392305  1.0392305  1.2124356
#> [151]  0.1732051  0.1732051  0.0000000 -0.1732051 -0.1732051  0.0000000
#> [157]  0.5196152  0.5196152  0.3464102  0.1732051  0.1732051  0.3464102
#> [163]  0.8660254  0.8660254  0.6928203  0.5196152  0.5196152  0.6928203
#> [169]  1.2124356  1.2124356  1.0392305  0.8660254  0.8660254  1.0392305
#> [175]  1.5588457  1.5588457  1.3856406  1.2124356  1.2124356  1.3856406
#> [181]  0.0000000  0.0000000 -0.1732051 -0.3464102 -0.3464102 -0.1732051
#> [187]  0.3464102  0.3464102  0.1732051  0.0000000  0.0000000  0.1732051
#> [193]  0.6928203  0.6928203  0.5196152  0.3464102  0.3464102  0.5196152
#> [199]  1.0392305  1.0392305  0.8660254  0.6928203  0.6928203  0.8660254
#> [205]  1.3856406  1.3856406  1.2124356  1.0392305  1.0392305  1.2124356
#> [211]  0.1732051  0.1732051  0.0000000 -0.1732051 -0.1732051  0.0000000
#> [217]  0.5196152  0.5196152  0.3464102  0.1732051  0.1732051  0.3464102
#> [223]  0.8660254  0.8660254  0.6928203  0.5196152  0.5196152  0.6928203
#> [229]  1.2124356  1.2124356  1.0392305  0.8660254  0.8660254  1.0392305
#> [235]  1.5588457  1.5588457  1.3856406  1.2124356  1.2124356  1.3856406
#> 
#> $y
#>   [1] -0.05040708 -0.24959292 -0.34918584 -0.24959292 -0.05040708  0.04918584
#>   [7] -0.05040708 -0.24959292 -0.34918584 -0.24959292 -0.05040708  0.04918584
#>  [13] -0.05040708 -0.24959292 -0.34918584 -0.24959292 -0.05040708  0.04918584
#>  [19] -0.05040708 -0.24959292 -0.34918584 -0.24959292 -0.05040708  0.04918584
#>  [25] -0.05040708 -0.24959292 -0.34918584 -0.24959292 -0.05040708  0.04918584
#>  [31]  0.24959292  0.05040708 -0.04918584  0.05040708  0.24959292  0.34918584
#>  [37]  0.24959292  0.05040708 -0.04918584  0.05040708  0.24959292  0.34918584
#>  [43]  0.24959292  0.05040708 -0.04918584  0.05040708  0.24959292  0.34918584
#>  [49]  0.24959292  0.05040708 -0.04918584  0.05040708  0.24959292  0.34918584
#>  [55]  0.24959292  0.05040708 -0.04918584  0.05040708  0.24959292  0.34918584
#>  [61]  0.54959292  0.35040708  0.25081416  0.35040708  0.54959292  0.64918584
#>  [67]  0.54959292  0.35040708  0.25081416  0.35040708  0.54959292  0.64918584
#>  [73]  0.54959292  0.35040708  0.25081416  0.35040708  0.54959292  0.64918584
#>  [79]  0.54959292  0.35040708  0.25081416  0.35040708  0.54959292  0.64918584
#>  [85]  0.54959292  0.35040708  0.25081416  0.35040708  0.54959292  0.64918584
#>  [91]  0.84959292  0.65040708  0.55081416  0.65040708  0.84959292  0.94918584
#>  [97]  0.84959292  0.65040708  0.55081416  0.65040708  0.84959292  0.94918584
#> [103]  0.84959292  0.65040708  0.55081416  0.65040708  0.84959292  0.94918584
#> [109]  0.84959292  0.65040708  0.55081416  0.65040708  0.84959292  0.94918584
#> [115]  0.84959292  0.65040708  0.55081416  0.65040708  0.84959292  0.94918584
#> [121]  1.14959292  0.95040708  0.85081416  0.95040708  1.14959292  1.24918584
#> [127]  1.14959292  0.95040708  0.85081416  0.95040708  1.14959292  1.24918584
#> [133]  1.14959292  0.95040708  0.85081416  0.95040708  1.14959292  1.24918584
#> [139]  1.14959292  0.95040708  0.85081416  0.95040708  1.14959292  1.24918584
#> [145]  1.14959292  0.95040708  0.85081416  0.95040708  1.14959292  1.24918584
#> [151]  1.44959292  1.25040708  1.15081416  1.25040708  1.44959292  1.54918584
#> [157]  1.44959292  1.25040708  1.15081416  1.25040708  1.44959292  1.54918584
#> [163]  1.44959292  1.25040708  1.15081416  1.25040708  1.44959292  1.54918584
#> [169]  1.44959292  1.25040708  1.15081416  1.25040708  1.44959292  1.54918584
#> [175]  1.44959292  1.25040708  1.15081416  1.25040708  1.44959292  1.54918584
#> [181]  1.74959292  1.55040708  1.45081416  1.55040708  1.74959292  1.84918584
#> [187]  1.74959292  1.55040708  1.45081416  1.55040708  1.74959292  1.84918584
#> [193]  1.74959292  1.55040708  1.45081416  1.55040708  1.74959292  1.84918584
#> [199]  1.74959292  1.55040708  1.45081416  1.55040708  1.74959292  1.84918584
#> [205]  1.74959292  1.55040708  1.45081416  1.55040708  1.74959292  1.84918584
#> [211]  2.04959292  1.85040708  1.75081416  1.85040708  2.04959292  2.14918584
#> [217]  2.04959292  1.85040708  1.75081416  1.85040708  2.04959292  2.14918584
#> [223]  2.04959292  1.85040708  1.75081416  1.85040708  2.04959292  2.14918584
#> [229]  2.04959292  1.85040708  1.75081416  1.85040708  2.04959292  2.14918584
#> [235]  2.04959292  1.85040708  1.75081416  1.85040708  2.04959292  2.14918584
#>