Hexagonal binning

This function generates the hexagonal object.

Usage

hex_binning(nldr_obj, b1 = 4, q = 0.1)

Arguments

nldr_obj

A list of a tibble contains scaled first and second columns of NLDR data, and numeric vectors representing the limits of the original NLDR data.

b1

Number of bins along the x axis.

q

The buffer amount as proportion of data range.

Value

A object that contains numeric vector that contains binwidths (a1), vertical distance (a2), bins along the x and y axes respectively (bins), numeric vector that contains hexagonal starting point coordinates all hexagonal bin centroids (centroids), hexagonal coordinates of the full grid (hex_poly), embedding components with their corresponding hexagon IDs (data_hb_id), hex bins with their corresponding standardise counts (std_cts), total number of hex bins (b), number of non-empty hex bins (m) and points within each hexagon (pts_bins).

Examples

hex_binning(nldr_obj = scurve_model_obj$nldr_obj, b1 = 4, q = 0.1)
#> $a1
#> [1] 0.3122301
#> 
#> $a2
#> [1] 0.2703992
#> 
#> $bins
#> [1] 4 5
#> 
#> $start_point
#> [1] -0.10000000 -0.08923607
#> 
#> $centroids
#> # A tibble: 20 × 3
#>        h     c_x     c_y
#>    <int>   <dbl>   <dbl>
#>  1     1 -0.1    -0.0892
#>  2     2  0.212  -0.0892
#>  3     3  0.524  -0.0892
#>  4     4  0.837  -0.0892
#>  5     5  0.0561  0.181 
#>  6     6  0.368   0.181 
#>  7     7  0.681   0.181 
#>  8     8  0.993   0.181 
#>  9     9 -0.1     0.452 
#> 10    10  0.212   0.452 
#> 11    11  0.524   0.452 
#> 12    12  0.837   0.452 
#> 13    13  0.0561  0.722 
#> 14    14  0.368   0.722 
#> 15    15  0.681   0.722 
#> 16    16  0.993   0.722 
#> 17    17 -0.1     0.992 
#> 18    18  0.212   0.992 
#> 19    19  0.524   0.992 
#> 20    20  0.837   0.992 
#> 
#> $hex_poly
#>      h           x             y
#> 1    1 -0.10000000  0.0910300574
#> 2    1 -0.25611504  0.0008969943
#> 3    1 -0.25611504 -0.1793691320
#> 4    1 -0.10000000 -0.2695021951
#> 5    1  0.05611504 -0.1793691320
#> 6    1  0.05611504  0.0008969943
#> 7    2  0.21223009  0.0910300574
#> 8    2  0.05611504  0.0008969943
#> 9    2  0.05611504 -0.1793691320
#> 10   2  0.21223009 -0.2695021951
#> 11   2  0.36834513 -0.1793691320
#> 12   2  0.36834513  0.0008969943
#> 13   3  0.52446018  0.0910300574
#> 14   3  0.36834513  0.0008969943
#> 15   3  0.36834513 -0.1793691320
#> 16   3  0.52446018 -0.2695021951
#> 17   3  0.68057522 -0.1793691320
#> 18   3  0.68057522  0.0008969943
#> 19   4  0.83669027  0.0910300574
#> 20   4  0.68057522  0.0008969943
#> 21   4  0.68057522 -0.1793691320
#> 22   4  0.83669027 -0.2695021951
#> 23   4  0.99280531 -0.1793691320
#> 24   4  0.99280531  0.0008969943
#> 25   5  0.05611504  0.3614292467
#> 26   5 -0.10000000  0.2712961836
#> 27   5 -0.10000000  0.0910300574
#> 28   5  0.05611504  0.0008969943
#> 29   5  0.21223009  0.0910300574
#> 30   5  0.21223009  0.2712961836
#> 31   6  0.36834513  0.3614292467
#> 32   6  0.21223009  0.2712961836
#> 33   6  0.21223009  0.0910300574
#> 34   6  0.36834513  0.0008969943
#> 35   6  0.52446018  0.0910300574
#> 36   6  0.52446018  0.2712961836
#> 37   7  0.68057522  0.3614292467
#> 38   7  0.52446018  0.2712961836
#> 39   7  0.52446018  0.0910300574
#> 40   7  0.68057522  0.0008969943
#> 41   7  0.83669027  0.0910300574
#> 42   7  0.83669027  0.2712961836
#> 43   8  0.99280531  0.3614292467
#> 44   8  0.83669027  0.2712961836
#> 45   8  0.83669027  0.0910300574
#> 46   8  0.99280531  0.0008969943
#> 47   8  1.14892036  0.0910300574
#> 48   8  1.14892036  0.2712961836
#> 49   9 -0.10000000  0.6318284361
#> 50   9 -0.25611504  0.5416953729
#> 51   9 -0.25611504  0.3614292467
#> 52   9 -0.10000000  0.2712961836
#> 53   9  0.05611504  0.3614292467
#> 54   9  0.05611504  0.5416953729
#> 55  10  0.21223009  0.6318284361
#> 56  10  0.05611504  0.5416953729
#> 57  10  0.05611504  0.3614292467
#> 58  10  0.21223009  0.2712961836
#> 59  10  0.36834513  0.3614292467
#> 60  10  0.36834513  0.5416953729
#> 61  11  0.52446018  0.6318284361
#> 62  11  0.36834513  0.5416953729
#> 63  11  0.36834513  0.3614292467
#> 64  11  0.52446018  0.2712961836
#> 65  11  0.68057522  0.3614292467
#> 66  11  0.68057522  0.5416953729
#> 67  12  0.83669027  0.6318284361
#> 68  12  0.68057522  0.5416953729
#> 69  12  0.68057522  0.3614292467
#> 70  12  0.83669027  0.2712961836
#> 71  12  0.99280531  0.3614292467
#> 72  12  0.99280531  0.5416953729
#> 73  13  0.05611504  0.9022276254
#> 74  13 -0.10000000  0.8120945623
#> 75  13 -0.10000000  0.6318284361
#> 76  13  0.05611504  0.5416953729
#> 77  13  0.21223009  0.6318284361
#> 78  13  0.21223009  0.8120945623
#> 79  14  0.36834513  0.9022276254
#> 80  14  0.21223009  0.8120945623
#> 81  14  0.21223009  0.6318284361
#> 82  14  0.36834513  0.5416953729
#> 83  14  0.52446018  0.6318284361
#> 84  14  0.52446018  0.8120945623
#> 85  15  0.68057522  0.9022276254
#> 86  15  0.52446018  0.8120945623
#> 87  15  0.52446018  0.6318284361
#> 88  15  0.68057522  0.5416953729
#> 89  15  0.83669027  0.6318284361
#> 90  15  0.83669027  0.8120945623
#> 91  16  0.99280531  0.9022276254
#> 92  16  0.83669027  0.8120945623
#> 93  16  0.83669027  0.6318284361
#> 94  16  0.99280531  0.5416953729
#> 95  16  1.14892036  0.6318284361
#> 96  16  1.14892036  0.8120945623
#> 97  17 -0.10000000  1.1726268147
#> 98  17 -0.25611504  1.0824937516
#> 99  17 -0.25611504  0.9022276254
#> 100 17 -0.10000000  0.8120945623
#> 101 17  0.05611504  0.9022276254
#> 102 17  0.05611504  1.0824937516
#> 103 18  0.21223009  1.1726268147
#> 104 18  0.05611504  1.0824937516
#> 105 18  0.05611504  0.9022276254
#> 106 18  0.21223009  0.8120945623
#> 107 18  0.36834513  0.9022276254
#> 108 18  0.36834513  1.0824937516
#> 109 19  0.52446018  1.1726268147
#> 110 19  0.36834513  1.0824937516
#> 111 19  0.36834513  0.9022276254
#> 112 19  0.52446018  0.8120945623
#> 113 19  0.68057522  0.9022276254
#> 114 19  0.68057522  1.0824937516
#> 115 20  0.83669027  1.1726268147
#> 116 20  0.68057522  1.0824937516
#> 117 20  0.68057522  0.9022276254
#> 118 20  0.83669027  0.8120945623
#> 119 20  0.99280531  0.9022276254
#> 120 20  0.99280531  1.0824937516
#> 
#> $data_hb_id
#> # A tibble: 5,000 × 4
#>     emb1  emb2    ID     h
#>    <dbl> <dbl> <int> <int>
#>  1 0.707 0.839     1    15
#>  2 0.231 0.401     2    10
#>  3 0.232 0.215     3     6
#>  4 0.790 0.564     4    12
#>  5 0.761 0.551     5    12
#>  6 0.445 0.721     6    14
#>  7 0.900 0.137     7     8
#>  8 0.247 0.392     8    10
#>  9 0.325 0.542     9    10
#> 10 0.278 0.231    10     6
#> # ℹ 4,990 more rows
#> 
#> $std_cts
#> # A tibble: 16 × 3
#>        h   n_h    w_h
#>    <int> <int>  <dbl>
#>  1     3   116 0.0232
#>  2     4    82 0.0164
#>  3     5   193 0.0386
#>  4     6   674 0.135 
#>  5     7   529 0.106 
#>  6     8   243 0.0486
#>  7     9    87 0.0174
#>  8    10   601 0.120 
#>  9    11   244 0.0488
#> 10    12   355 0.071 
#> 11    13   183 0.0366
#> 12    14   627 0.125 
#> 13    15   669 0.134 
#> 14    16   318 0.0636
#> 15    19     2 0.0004
#> 16    20    77 0.0154
#> 
#> $b
#> [1] 20
#> 
#> $m
#> [1] 16
#> 
#> $pts_bins
#> # A tibble: 16 × 2
#>        h pts_list   
#>    <int> <list>     
#>  1     3 <int [116]>
#>  2     4 <int [82]> 
#>  3     5 <int [193]>
#>  4     6 <int [674]>
#>  5     7 <int [529]>
#>  6     8 <int [243]>
#>  7     9 <int [87]> 
#>  8    10 <int [601]>
#>  9    11 <int [244]>
#> 10    12 <int [355]>
#> 11    13 <int [183]>
#> 12    14 <int [627]>
#> 13    15 <int [669]>
#> 14    16 <int [318]>
#> 15    19 <int [2]>  
#> 16    20 <int [77]> 
#> 
#> attr(,"class")
#> [1] "hex_bin_obj"