Hexagonal binning

This function generates the hexagonal object.

Usage

hex_binning(data, bin1 = 2, s1 = -0.1, s2 = -0.1, r2)

Arguments

data

A tibble that contains embedding components.

bin1

Number of bins along the x axis.

s1

The x-coordinate of the hexagonal grid starting point.

s2

The y-coordinate of the hexagonal grid starting point.

r2

The range of the original second embedding component.

Value

A object that contains numeric vector that contains bins along the x and y axes respectively, all hexagonal bin centroids (centroids), hexagonal coordinates of the full grid(hex_poly), embedding components with their corresponding hexagon IDs (data_hb_id), hex bins with their corresponding standardise counts (std_cts), total number of hex bins(tot_bins), number of non-empty hex bins (non_bins) and points within each hexagon (pts_bins).

Examples

range_umap2 <- diff(range(s_curve_noise_umap$UMAP2))
num_bins_x <- 3
hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x, s1 = -0.1,
s2 = -0.1, r2 = range_umap2)
#> $bins
#> [1]  3 26
#> 
#> $centroids
#> # A tibble: 78 × 3
#>    hexID    c_x    c_y
#>    <int>  <dbl>  <dbl>
#>  1     1 -0.1   -0.1  
#>  2     2  0.439 -0.1  
#>  3     3  0.977 -0.1  
#>  4     4  0.169  0.367
#>  5     5  0.708  0.367
#>  6     6  1.25   0.367
#>  7     7 -0.1    0.833
#>  8     8  0.439  0.833
#>  9     9  0.977  0.833
#> 10    10  0.169  1.30 
#> # ℹ 68 more rows
#> 
#> $hex_poly
#> # A tibble: 468 × 3
#>    hex_poly_id      x       y
#>          <int>  <dbl>   <dbl>
#>  1           1 -0.1    0.211 
#>  2           1 -0.369  0.0555
#>  3           1 -0.369 -0.256 
#>  4           1 -0.1   -0.411 
#>  5           1  0.169 -0.256 
#>  6           1  0.169  0.0555
#>  7           2  0.439  0.211 
#>  8           2  0.169  0.0555
#>  9           2  0.169 -0.256 
#> 10           2  0.439 -0.411 
#> # ℹ 458 more rows
#> 
#> $data_hb_id
#> # A tibble: 75 × 4
#>     UMAP1  UMAP2    ID hb_id
#>     <dbl>  <dbl> <int> <int>
#>  1 0.0804 0.274      1     4
#>  2 0.739  0.860      2     9
#>  3 0.840  0.929      3     9
#>  4 0.167  0.0371     4     1
#>  5 0.263  0.342      6     4
#>  6 0.838  1.72       7    15
#>  7 0.734  0.834      8     9
#>  8 0.627  0.618      9     5
#>  9 0.810  0.868     11     9
#> 10 0.903  1.60      12    15
#> # ℹ 65 more rows
#> 
#> $std_cts
#> # A tibble: 9 × 3
#>   hb_id     n std_counts
#>   <int> <int>      <dbl>
#> 1     1     6     0.375 
#> 2     2     3     0.188 
#> 3     4    16     1     
#> 4     5     5     0.312 
#> 5     8     6     0.375 
#> 6     9    12     0.75  
#> 7    11    12     0.75  
#> 8    14     1     0.0625
#> 9    15    14     0.875 
#> 
#> $tot_bins
#> [1] 78
#> 
#> $non_bins
#> [1] 9
#> 
#> $pts_bins
#> # A tibble: 9 × 2
#>   hexID pts_list    
#>   <int> <named list>
#> 1     4 <int [75]>  
#> 2     9 <int [75]>  
#> 3     1 <int [75]>  
#> 4    15 <int [75]>  
#> 5     5 <int [75]>  
#> 6    11 <int [75]>  
#> 7     8 <int [75]>  
#> 8     2 <int [75]>  
#> 9    14 <int [75]>  
#> 
#> attr(,"class")
#> [1] "hex_bin_obj"