Hexagonal binning

This function generates the hexagonal object.

Usage

hex_binning(data, bin1 = 4, r2, q = 0.1)

Arguments

data

A tibble that contains embedding components.

bin1

Number of bins along the x axis.

r2

The ratio of the ranges of the original embedding components.

q

The buffer amount as proportion of data range.

Value

A object that contains numeric vector that contains binwidths (a1), vertical distance (a2), bins along the x and y axes respectively (bins), numeric vector that contains hexagonal starting point coordinates all hexagonal bin centroids (centroids), hexagonal coordinates of the full grid(hex_poly), embedding components with their corresponding hexagon IDs (data_hb_id), hex bins with their corresponding standardise counts (std_cts), total number of hex bins(tot_bins), number of non-empty hex bins (non_bins) and points within each hexagon (pts_bins).

Examples

scurve_umap_scaled_obj <- s_curve_obj$s_curve_umap_scaled_obj
lim1 <- scurve_umap_scaled_obj$lim1
lim2 <- scurve_umap_scaled_obj$lim2
r2 <- diff(lim2)/diff(lim1)
num_bins_x <- 4
hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x,
r2 = r2, q = 0.1)
#> $a1
#> [1] 0.3268102
#> 
#> $a2
#> [1] 0.2830259
#> 
#> $bins
#> [1] 4 5
#> 
#> $start_point
#> [1] -0.10000000 -0.09382762
#> 
#> $centroids
#> # A tibble: 20 × 3
#>    hexID     c_x     c_y
#>    <int>   <dbl>   <dbl>
#>  1     1 -0.1    -0.0938
#>  2     2  0.227  -0.0938
#>  3     3  0.554  -0.0938
#>  4     4  0.880  -0.0938
#>  5     5  0.0634  0.189 
#>  6     6  0.390   0.189 
#>  7     7  0.717   0.189 
#>  8     8  1.04    0.189 
#>  9     9 -0.1     0.472 
#> 10    10  0.227   0.472 
#> 11    11  0.554   0.472 
#> 12    12  0.880   0.472 
#> 13    13  0.0634  0.755 
#> 14    14  0.390   0.755 
#> 15    15  0.717   0.755 
#> 16    16  1.04    0.755 
#> 17    17 -0.1     1.04  
#> 18    18  0.227   1.04  
#> 19    19  0.554   1.04  
#> 20    20  0.880   1.04  
#> 
#> $hex_poly
#> # A tibble: 120 × 3
#>    hex_poly_id       x         y
#>          <int>   <dbl>     <dbl>
#>  1           1 -0.1     0.0949  
#>  2           1 -0.263   0.000514
#>  3           1 -0.263  -0.188   
#>  4           1 -0.1    -0.283   
#>  5           1  0.0634 -0.188   
#>  6           1  0.0634  0.000514
#>  7           2  0.227   0.0949  
#>  8           2  0.0634  0.000514
#>  9           2  0.0634 -0.188   
#> 10           2  0.227  -0.283   
#> # ℹ 110 more rows
#> 
#> $data_hb_id
#> # A tibble: 3,750 × 4
#>      emb1   emb2    ID hb_id
#>     <dbl>  <dbl> <int> <int>
#>  1 0.276  0.915      1    18
#>  2 0.927  0.347      2    12
#>  3 0.810  0.242      3     7
#>  4 0.137  0.657      5    13
#>  5 0.476  0.799      6    14
#>  6 0.0485 0.0657     7     5
#>  7 0.737  0.851      9    15
#>  8 0.833  0.224     10     7
#>  9 0.909  0.371     11    12
#> 10 0.0890 0.194     12     5
#> # ℹ 3,740 more rows
#> 
#> $std_cts
#> # A tibble: 15 × 3
#>    hb_id     n std_counts
#>    <int> <int>      <dbl>
#>  1     1    17     0.0302
#>  2     2   169     0.300 
#>  3     3    34     0.0604
#>  4     5   431     0.766 
#>  5     6   430     0.764 
#>  6     7   407     0.723 
#>  7     8    57     0.101 
#>  8    11    82     0.146 
#>  9    12   367     0.652 
#> 10    13   442     0.785 
#> 11    14   478     0.849 
#> 12    15   563     1     
#> 13    17    26     0.0462
#> 14    18   149     0.265 
#> 15    19    98     0.174 
#> 
#> $tot_bins
#> [1] 20
#> 
#> $non_bins
#> [1] 15
#> 
#> $pts_bins
#> # A tibble: 15 × 2
#>    hb_id pts_list   
#>    <int> <list>     
#>  1     1 <int [17]> 
#>  2     2 <int [169]>
#>  3     3 <int [34]> 
#>  4     5 <int [431]>
#>  5     6 <int [430]>
#>  6     7 <int [407]>
#>  7     8 <int [57]> 
#>  8    11 <int [82]> 
#>  9    12 <int [367]>
#> 10    13 <int [442]>
#> 11    14 <int [478]>
#> 12    15 <int [563]>
#> 13    17 <int [26]> 
#> 14    18 <int [149]>
#> 15    19 <int [98]> 
#> 
#> attr(,"class")
#> [1] "hex_bin_obj"