Hexagonal binning

This function generates a list which contains hexagonal binning information.

Usage

hex_binning(
  data,
  x,
  y,
  num_bins_x,
  num_bins_y,
  x_start,
  y_start,
  buffer_x,
  buffer_y,
  hex_size,
  col_start
)

Arguments

data

A tibble or data frame.

x

The name of the column that contains values along the x-axis.

y

The name of the column that contains values along the y-axis.

num_bins_x

Number of bins along the x-axis.

num_bins_y

Number of bins along the y-axis.

x_start

Starting point along the x-axis for hexagonal binning.

y_start

Starting point along the y-axis for hexagonal binning.

buffer_x

The buffer size along the x-axis.

buffer_y

The buffer size along the y-axis.

hex_size

A numeric value that initializes the radius of the outer circle surrounding the hexagon.

col_start

The text that begins the column name of x and y axes of data.

Value

A list contains all hexagonal bin centroids (centroids), hexagonal coordinates of the full grid(hex_poly), 2D embeddings with corresponding hexagon IDs (data_hb_id), hex bins with their corresponding standardise counts (std_cts), total number of hex bins(tot_bins), number of non-empty hex bins (non_bins) and points within each hexagon (pts_bins).

Examples

num_bins_list <- calc_bins(data = s_curve_noise_umap_scaled, x = "UMAP1", y = "UMAP2",
hex_size = 0.2, buffer_x = 0.346, buffer_y = 0.3)
num_bins_x <- num_bins_list$num_x
num_bins_y <- num_bins_list$num_y
hex_binning(data = s_curve_noise_umap_scaled,
x = "UMAP1", y = "UMAP2", num_bins_x = num_bins_x,
num_bins_y = num_bins_y, x_start = -0.1732051, y_start = -0.15, buffer_x = 0.346,
buffer_y = 0.3, hex_size = 0.2, col_start = "UMAP")
#> $centroids
#> # A tibble: 28 × 3
#>    hexID           c_x   c_y
#>    <int>         <dbl> <dbl>
#>  1     1 -0.173        -0.15
#>  2     2  0.173        -0.15
#>  3     3  0.520        -0.15
#>  4     4  0.866        -0.15
#>  5     5 -0.0000000192  0.15
#>  6     6  0.346         0.15
#>  7     7  0.693         0.15
#>  8     8  1.04          0.15
#>  9     9 -0.173         0.45
#> 10    10  0.173         0.45
#> # ℹ 18 more rows
#> 
#> $hex_poly
#> # A tibble: 168 × 3
#>    hex_poly_id             x       y
#>          <int>         <dbl>   <dbl>
#>  1           1 -0.0000000192 -0.0504
#>  2           1 -0.0000000192 -0.250 
#>  3           1 -0.173        -0.349 
#>  4           1 -0.346        -0.250 
#>  5           1 -0.346        -0.0504
#>  6           1 -0.173         0.0492
#>  7           2  0.346        -0.0504
#>  8           2  0.346        -0.250 
#>  9           2  0.173        -0.349 
#> 10           2 -0.0000000192 -0.250 
#> # ℹ 158 more rows
#> 
#> $data_hb_id
#> # A tibble: 75 × 4
#>     UMAP1  UMAP2    ID hb_id
#>     <dbl>  <dbl> <int> <int>
#>  1 0.0804 0.274      1     5
#>  2 0.739  0.860      2    15
#>  3 0.840  0.929      3    20
#>  4 0.167  0.0371     4     2
#>  5 0.263  0.342      6    10
#>  6 0.838  1.72       7    28
#>  7 0.734  0.834      8    15
#>  8 0.627  0.618      9    15
#>  9 0.810  0.868     11    15
#> 10 0.903  1.60      12    28
#> # ℹ 65 more rows
#> 
#> $std_cts
#> # A tibble: 10 × 3
#>    hb_id     n std_counts
#>    <int> <int>      <dbl>
#>  1     2     4      0.235
#>  2     5     9      0.529
#>  3     6     7      0.412
#>  4    10     3      0.176
#>  5    11     6      0.353
#>  6    15    12      0.706
#>  7    20     8      0.471
#>  8    23     5      0.294
#>  9    24     4      0.235
#> 10    28    17      1    
#> 
#> $tot_bins
#> [1] 28
#> 
#> $non_bins
#> [1] 10
#> 
#> $pts_bins
#> # A tibble: 10 × 2
#>    hexID pts_list    
#>    <int> <named list>
#>  1     5 <int [75]>  
#>  2    15 <int [75]>  
#>  3    20 <int [75]>  
#>  4     2 <int [75]>  
#>  5    10 <int [75]>  
#>  6    28 <int [75]>  
#>  7    23 <int [75]>  
#>  8    11 <int [75]>  
#>  9     6 <int [75]>  
#> 10    24 <int [75]>  
#>