Hexagonal binning

This function generates the hexagonal object.

Usage

hex_binning(data, bin1 = 2, r2)

Arguments

data

A tibble that contains embedding components.

bin1

Number of bins along the x axis.

r2

The ratio of the ranges of the original embedding components.

Value

A object that contains numeric vector that contains bins along the x and y axes respectively (bins), numeric vector that contains hexagonal starting point coordinates all hexagonal bin centroids (centroids), hexagonal coordinates of the full grid(hex_poly), embedding components with their corresponding hexagon IDs (data_hb_id), hex bins with their corresponding standardise counts (std_cts), total number of hex bins(tot_bins), number of non-empty hex bins (non_bins) and points within each hexagon (pts_bins).

Examples

r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
num_bins_x <- 3
hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x, r2 = r2)
#> $bins
#> [1] 3 6
#> 
#> $start_point
#> [1] -0.1000000 -0.2019414
#> 
#> $centroids
#> # A tibble: 18 × 3
#>    hexID    c_x    c_y
#>    <int>  <dbl>  <dbl>
#>  1     1 -0.1   -0.202
#>  2     2  0.413 -0.202
#>  3     3  0.926 -0.202
#>  4     4  0.157  0.242
#>  5     5  0.670  0.242
#>  6     6  1.18   0.242
#>  7     7 -0.1    0.687
#>  8     8  0.413  0.687
#>  9     9  0.926  0.687
#> 10    10  0.157  1.13 
#> 11    11  0.670  1.13 
#> 12    12  1.18   1.13 
#> 13    13 -0.1    1.58 
#> 14    14  0.413  1.58 
#> 15    15  0.926  1.58 
#> 16    16  0.157  2.02 
#> 17    17  0.670  2.02 
#> 18    18  1.18   2.02 
#> 
#> $hex_poly
#> # A tibble: 108 × 3
#>    hex_poly_id      x       y
#>          <int>  <dbl>   <dbl>
#>  1           1 -0.1    0.0942
#>  2           1 -0.357 -0.0539
#>  3           1 -0.357 -0.350 
#>  4           1 -0.1   -0.498 
#>  5           1  0.157 -0.350 
#>  6           1  0.157 -0.0539
#>  7           2  0.413  0.0942
#>  8           2  0.157 -0.0539
#>  9           2  0.157 -0.350 
#> 10           2  0.413 -0.498 
#> # ℹ 98 more rows
#> 
#> $data_hb_id
#> # A tibble: 75 × 4
#>     UMAP1  UMAP2    ID hb_id
#>     <dbl>  <dbl> <int> <int>
#>  1 0.0804 0.274      1     4
#>  2 0.739  0.860      2     9
#>  3 0.840  0.929      3     9
#>  4 0.167  0.0371     4     4
#>  5 0.263  0.342      6     4
#>  6 0.838  1.72       7    15
#>  7 0.734  0.834      8     9
#>  8 0.627  0.618      9     8
#>  9 0.810  0.868     11     9
#> 10 0.903  1.60      12    15
#> # ℹ 65 more rows
#> 
#> $std_cts
#> # A tibble: 5 × 3
#>   hb_id     n std_counts
#>   <int> <int>      <dbl>
#> 1     4    25      0.962
#> 2     8    10      0.385
#> 3     9     7      0.269
#> 4    11     7      0.269
#> 5    15    26      1    
#> 
#> $tot_bins
#> [1] 18
#> 
#> $non_bins
#> [1] 5
#> 
#> $pts_bins
#> # A tibble: 5 × 2
#>   hexID pts_list    
#>   <int> <named list>
#> 1     4 <int [75]>  
#> 2     9 <int [75]>  
#> 3    15 <int [75]>  
#> 4     8 <int [75]>  
#> 5    11 <int [75]>  
#> 
#> attr(,"class")
#> [1] "hex_bin_obj"