Triangulate bin centroids

This function triangulates the bin centroids using the x and y coordinates provided in the input data frame and returns the triangular object.

Usage

tri_bin_centroids(centroids_data)

Arguments

centroids_data

The tibble containing the all the bin centroids.

Value

A triangular object representing the triangulated bin centroids.

Examples

all_centroids_df <- scurve_model_obj$hb_obj$centroids
counts_data <- scurve_model_obj$hb_obj$std_cts
umap_with_hb_id <- scurve_model_obj$hb_obj$data_hb_id
df_bin_centroids <- extract_hexbin_mean(data_hb = umap_with_hb_id,
counts_data = counts_data, centroids_data = all_centroids_df)
tri_bin_centroids(centroids_data = df_bin_centroids)
#> $trimesh_object
#> Delaunay triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,2,9), [2,3,0]
#> 2: (10,9,2), [1,7,5]
#> 3: (8,1,9), [1,8,65]
#> 4: (19,9,20), [5,12,8]
#> 5: (9,10,20), [9,4,2]
#> 6: (11,10,3), [7,19,10]
#> 7: (3,10,2), [2,0,6]
#> 8: (19,8,9), [3,4,11]
#> 9: (10,21,20), [14,5,10]
#> 10: (11,21,10), [9,6,18]
#> 11: (18,8,19), [8,17,26]
#> 12: (31,19,20), [4,15,13]
#> 13: (31,30,19), [17,12,28]
#> 14: (20,21,32), [21,15,9]
#> 15: (20,32,31), [29,12,14]
#> 16: (11,12,22), [34,18,22]
#> 17: (30,18,19), [11,13,23]
#> 18: (22,21,11), [10,16,20]
#> 19: (4,11,3), [6,62,22]
#> 20: (33,21,22), [18,40,21]
#> 21: (33,32,21), [14,20,32]
#> 22: (11,4,12), [36,16,19]
#> 23: (29,18,30), [17,39,27]
#> 24: (43,31,44), [29,43,28]
#> 25: (8,17,27), [66,65,26]
#> 26: (17,8,18), [11,27,25]
#> 27: (17,18,29), [23,41,26]
#> 28: (43,30,31), [13,24,31]
#> 29: (44,31,32), [15,30,24]
#> 30: (32,45,44), [46,29,32]
#> 31: (42,30,43), [28,53,39]
#> 32: (32,33,45), [48,30,21]
#> 33: (46,33,34), [40,49,48]
#> 34: (22,12,23), [50,45,16]
#> 35: (13,12,5), [36,61,50]
#> 36: (5,12,4), [22,37,35]
#> 37: (4,6,5), [61,36,62]
#> 38: (41,29,42), [39,63,51]
#> 39: (42,29,30), [23,31,38]
#> 40: (34,33,22), [20,45,33]
#> 41: (29,28,17), [66,27,51]
#> 42: (55,43,56), [43,73,53]
#> 43: (56,43,44), [24,44,42]
#> 44: (44,57,56), [103,43,46]
#> 45: (22,23,34), [57,40,34]
#> 46: (44,45,57), [56,44,30]
#> 47: (58,45,46), [48,58,56]
#> 48: (46,45,33), [32,33,47]
#> 49: (46,34,47), [70,75,33]
#> 50: (13,23,12), [34,35,69]
#> 51: (41,28,29), [41,38,68]
#> 52: (54,42,55), [53,77,63]
#> 53: (55,42,43), [31,42,52]
#> 54: (56,79,68), [64,73,103]
#> 55: (80,57,58), [56,102,110]
#> 56: (58,57,45), [46,47,55]
#> 57: (34,23,35), [59,70,45]
#> 58: (46,59,58), [74,47,75]
#> 59: (24,35,23), [57,69,71]
#> 60: (14,13,6), [61,86,80]
#> 61: (6,13,5), [35,37,60]
#> 62: (4,3,6), [0,37,19]
#> 63: (54,41,42), [38,52,72]
#> 64: (79,78,68), [101,54,109]
#> 65: (8,27,1), [0,3,25]
#> 66: (27,17,28), [41,67,25]
#> 67: (40,27,28), [66,68,94]
#> 68: (40,28,41), [51,81,67]
#> 69: (24,23,13), [50,80,59]
#> 70: (34,35,47), [79,49,57]
#> 71: (36,35,24), [59,76,84]
#> 72: (41,54,53), [87,81,63]
#> 73: (68,55,56), [42,54,78]
#> 74: (58,59,69), [93,102,58]
#> 75: (46,47,59), [90,58,49]
#> 76: (24,25,36), [97,71,98]
#> 77: (67,54,55), [52,78,82]
#> 78: (67,55,68), [73,101,77]
#> 79: (48,47,35), [70,84,83]
#> 80: (14,24,13), [69,60,98]
#> 81: (41,53,40), [96,68,72]
#> 82: (54,67,66), [88,87,77]
#> 83: (60,47,48), [79,112,90]
#> 84: (48,35,36), [71,106,79]
#> 85: (15,14,7), [86,136,107]
#> 86: (7,14,6), [60,0,85]
#> 87: (54,66,53), [99,72,82]
#> 88: (67,77,66), [108,82,100]
#> 89: (59,60,70), [118,93,90]
#> 90: (60,59,47), [75,83,89]
#> 91: (61,48,49), [106,138,112]
#> 92: (69,89,80), [127,102,114]
#> 93: (59,70,69), [133,74,89]
#> 94: (39,27,40), [67,95,115]
#> 95: (52,39,40), [94,96,116]
#> 96: (52,40,53), [81,104,95]
#> 97: (37,36,25), [76,124,120]
#> 98: (24,14,25), [107,76,80]
#> 99: (53,66,65), [122,104,87]
#> 100: (67,78,77), [130,88,101]
#> 101: (67,68,78), [64,100,78]
#> 102: (80,58,69), [74,92,55]
#> 103: (56,57,79), [110,54,44]
#> 104: (53,65,52), [128,96,99]
#> 105: (79,88,87), [155,109,126]
#> 106: (49,48,36), [84,120,91]
#> 107: (15,25,14), [98,85,113]
#> 108: (66,77,76), [123,122,88]
#> 109: (87,78,79), [64,105,125]
#> 110: (80,79,57), [103,55,126]
#> 111: (71,60,61), [112,140,118]
#> 112: (61,60,48), [83,91,111]
#> 113: (25,15,26), [135,124,107]
#> 114: (90,89,69), [92,133,167]
#> 115: (51,27,39), [94,116,0]
#> 116: (51,39,52), [95,117,115]
#> 117: (64,51,52), [116,128,147]
#> 118: (71,70,60), [89,111,157]
#> 119: (37,50,49), [145,120,137]
#> 120: (36,37,49), [119,106,97]
#> 121: (75,64,65), [128,141,148]
#> 122: (66,76,65), [141,99,108]
#> 123: (85,76,77), [108,129,142]
#> 124: (37,25,26), [113,139,97]
#> 125: (86,78,87), [109,154,130]
#> 126: (79,80,88), [127,105,110]
#> 127: (80,89,88), [152,126,92]
#> 128: (64,52,65), [104,121,117]
#> 129: (85,77,86), [130,131,123]
#> 130: (77,78,86), [125,129,100]
#> 131: (86,97,85), [165,129,143]
#> 132: (99,98,87), [154,155,177]
#> 133: (90,69,70), [93,151,114]
#> 134: (26,16,38), [164,139,135]
#> 135: (16,26,15), [113,136,134]
#> 136: (16,15,7), [85,0,135]
#> 137: (37,38,50), [163,119,139]
#> 138: (49,62,61), [146,91,145]
#> 139: (26,38,37), [137,124,134]
#> 140: (61,72,71), [168,111,146]
#> 141: (75,65,76), [122,153,121]
#> 142: (84,76,85), [123,175,153]
#> 143: (98,97,86), [131,154,166]
#> 144: (70,92,91), [169,151,157]
#> 145: (49,50,62), [162,138,119]
#> 146: (61,62,72), [160,140,138]
#> 147: (74,51,64), [117,148,0]
#> 148: (74,64,75), [121,172,147]
#> 149: (83,75,84), [153,188,172]
#> 150: (100,99,88), [155,152,158]
#> 151: (70,91,90), [174,133,144]
#> 152: (100,88,89), [127,159,150]
#> 153: (84,75,76), [141,142,149]
#> 154: (98,86,87), [125,132,143]
#> 155: (99,87,88), [105,150,132]
#> 156: (102,101,90), [167,174,197]
#> 157: (70,71,92), [171,144,118]
#> 158: (111,99,100), [150,180,178]
#> 159: (89,101,100), [179,152,167]
#> 160: (62,73,72), [170,146,161]
#> 161: (62,63,73), [208,160,162]
#> 162: (63,62,50), [145,163,161]
#> 163: (63,50,38), [137,164,162]
#> 164: (63,38,16), [134,249,163]
#> 165: (96,85,97), [131,195,175]
#> 166: (98,109,97), [185,143,176]
#> 167: (89,90,101), [156,159,114]
#> 168: (71,72,81), [186,171,140]
#> 169: (103,91,92), [144,201,182]
#> 170: (72,73,82), [194,186,160]
#> 171: (71,81,92), [184,157,168]
#> 172: (83,74,75), [148,149,187]
#> 173: (95,84,96), [175,198,188]
#> 174: (102,90,91), [151,182,156]
#> 175: (96,84,85), [142,165,173]
#> 176: (98,110,109), [196,166,177]
#> 177: (110,98,99), [132,178,176]
#> 178: (111,110,99), [177,158,191]
#> 179: (112,100,101), [159,192,180]
#> 180: (112,111,100), [158,179,211]
#> 181: (114,102,103), [182,214,199]
#> 182: (103,102,91), [174,169,181]
#> 183: (92,93,104), [219,201,184]
#> 184: (93,92,81), [171,193,183]
#> 185: (108,97,109), [166,217,195]
#> 186: (72,82,81), [193,168,170]
#> 187: (95,74,83), [172,188,0]
#> 188: (95,83,84), [149,173,187]
#> 189: (107,96,108), [195,212,198]
#> 190: (111,123,122), [224,191,211]
#> 191: (122,110,111), [178,190,204]
#> 192: (101,113,112), [205,179,197]
#> 193: (81,82,93), [202,184,186]
#> 194: (73,94,82), [202,170,236]
#> 195: (108,96,97), [165,185,189]
#> 196: (121,109,110), [176,204,203]
#> 197: (101,102,113), [199,192,156]
#> 198: (107,95,96), [173,189,216]
#> 199: (114,113,102), [197,181,209]
#> 200: (115,103,104), [201,230,214]
#> 201: (92,104,103), [200,169,183]
#> 202: (93,82,94), [194,207,193]
#> 203: (120,109,121), [196,227,217]
#> 204: (121,110,122), [191,213,196]
#> 205: (124,112,113), [192,221,206]
#> 206: (124,123,112), [211,205,225]
#> 207: (105,93,94), [202,235,219]
#> 208: (73,63,106), [237,236,161]
#> 209: (113,114,125), [210,221,199]
#> 210: (114,126,125), [245,209,232]
#> 211: (111,112,123), [206,190,180]
#> 212: (108,119,107), [216,189,218]
#> 213: (131,121,122), [204,228,233]
#> 214: (115,114,103), [181,200,232]
#> 215: (116,104,105), [219,242,230]
#> 216: (119,95,107), [198,212,0]
#> 217: (120,108,109), [185,203,218]
#> 218: (108,120,119), [226,212,217]
#> 219: (105,104,93), [183,207,215]
#> 220: (134,124,125), [221,229,231]
#> 221: (113,125,124), [220,205,209]
#> 222: (127,115,116), [230,247,240]
#> 223: (123,133,132), [238,224,225]
#> 224: (123,132,122), [228,190,223]
#> 225: (124,133,123), [223,206,231]
#> 226: (130,119,120), [218,227,0]
#> 227: (121,130,120), [226,203,233]
#> 228: (131,122,132), [224,252,213]
#> 229: (125,135,134), [244,220,245]
#> 230: (116,115,104), [200,215,222]
#> 231: (134,133,124), [225,220,238]
#> 232: (114,115,126), [240,210,214]
#> 233: (131,130,121), [227,213,0]
#> 234: (117,105,106), [235,248,242]
#> 235: (106,105,94), [207,236,234]
#> 236: (73,106,94), [235,194,208]
#> 237: (63,118,106), [248,208,249]
#> 238: (134,132,133), [223,231,243]
#> 239: (126,127,136), [255,251,240]
#> 240: (127,126,115), [232,222,239]
#> 241: (128,116,117), [242,246,247]
#> 242: (117,116,105), [215,234,241]
#> 243: (134,140,132), [252,238,244]
#> 244: (134,135,140), [253,243,229]
#> 245: (125,126,135), [251,229,210]
#> 246: (117,129,128), [256,241,257]
#> 247: (128,127,116), [222,241,250]
#> 248: (118,117,106), [234,237,257]
#> 249: (63,16,118), [265,237,164]
#> 250: (137,127,128), [247,260,255]
#> 251: (126,136,135), [253,245,239]
#> 252: (132,140,131), [0,228,243]
#> 253: (140,135,136), [251,258,244]
#> 254: (136,137,141), [261,258,255]
#> 255: (137,136,127), [239,250,254]
#> 256: (138,128,129), [246,263,259]
#> 257: (117,118,129), [264,246,248]
#> 258: (136,141,140), [0,253,254]
#> 259: (142,128,138), [256,262,260]
#> 260: (142,137,128), [250,259,261]
#> 261: (142,141,137), [254,260,0]
#> 262: (139,142,138), [259,263,0]
#> 263: (139,138,129), [256,264,262]
#> 264: (139,129,118), [257,265,263]
#> 265: (139,118,16), [249,0,264]
#> boundary nodes:  1 2 3 6 7 16 139 142 141 140 131 130 119 95 74 51 27 
#> 
#> $n_h
#>   [1] 12 17 22 10  7 12  1 42 42 44 39 52 38 52 36 27  2 65 47 45 35 31 40 43 36
#>  [26] 43  3 39 67 56 38 42 42 30 41 33 38 10  1 62 62 68 50 32 44 34 40 36 43 40
#>  [51] 13 45 34 58 32 25 38 40 43 37 42 50 14 41 34 45 60  4  1  8 22 24 19 10 31
#>  [76] 36 39 54 34  7  3  5 32 32 52 48 48 49 36  1 28 46 53 42 47 35 37 41 46 42
#> [101] 46 56 35 31 44 26 28 34 43 49 43 56 75 49 38 42 50  4 17 42 43 45 49 63 58
#> [126] 45 42 41 39 11 39 36  8 17 44 47 47 41 11  3  8  1
#>