Triangulate bin centroids

This function triangulates the bin centroids using the x and y coordinates provided in the input data frame and returns the triangular object.

Usage

tri_bin_centroids(hex_df, x, y)

Arguments

hex_df

The tibble containing the bin centroids.

x

The name of the column that contains x coordinates of bin centroids.

y

The name of the column that contains y coordinates of bin centroids.

Value

A triangular object representing the triangulated bin centroids.

Examples

df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
tri_bin_centroids(hex_df = df_bin_centroids, x = "c_x", y = "c_y")
#> $trimesh_object
#> Delaunay triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,2,4), [2,4,0]
#> 2: (5,4,2), [1,3,6]
#> 3: (3,5,2), [2,0,8]
#> 4: (12,1,4), [1,5,34]
#> 5: (4,13,12), [10,4,6]
#> 6: (4,5,13), [9,5,2]
#> 7: (14,5,6), [8,14,9]
#> 8: (5,3,6), [17,7,3]
#> 9: (14,13,5), [6,7,12]
#> 10: (22,12,13), [5,11,18]
#> 11: (13,23,22), [20,10,12]
#> 12: (13,14,23), [16,11,9]
#> 13: (14,15,24), [24,16,14]
#> 14: (15,14,6), [7,15,13]
#> 15: (6,7,15), [26,14,17]
#> 16: (14,24,23), [21,12,13]
#> 17: (6,3,7), [33,15,8]
#> 18: (32,12,22), [10,19,35]
#> 19: (33,32,22), [18,20,27]
#> 20: (33,22,23), [11,22,19]
#> 21: (23,24,34), [30,22,16]
#> 22: (23,34,33), [37,20,21]
#> 23: (35,24,25), [24,31,30]
#> 24: (24,15,25), [25,23,13]
#> 25: (15,16,25), [40,24,26]
#> 26: (15,7,16), [32,25,15]
#> 27: (44,32,33), [19,28,36]
#> 28: (45,44,33), [27,37,55]
#> 29: (46,34,35), [30,42,38]
#> 30: (35,34,24), [21,23,29]
#> 31: (25,36,35), [45,23,39]
#> 32: (8,16,7), [26,33,44]
#> 33: (8,7,3), [17,47,32]
#> 34: (43,1,12), [4,35,143]
#> 35: (43,12,32), [18,36,34]
#> 36: (43,32,44), [27,54,35]
#> 37: (45,33,34), [22,38,28]
#> 38: (46,45,34), [37,29,56]
#> 39: (25,26,36), [49,31,40]
#> 40: (25,16,26), [41,39,25]
#> 41: (16,17,26), [43,40,44]
#> 42: (35,47,46), [58,29,45]
#> 43: (26,17,27), [50,48,41]
#> 44: (16,8,17), [46,41,32]
#> 45: (35,36,47), [51,42,31]
#> 46: (9,17,8), [44,47,52]
#> 47: (9,8,3), [33,63,46]
#> 48: (26,27,37), [66,49,43]
#> 49: (26,37,36), [53,39,48]
#> 50: (17,18,27), [60,43,52]
#> 51: (36,48,47), [86,45,53]
#> 52: (17,9,18), [62,50,46]
#> 53: (36,37,48), [64,51,49]
#> 54: (61,43,44), [36,55,70]
#> 55: (61,44,45), [28,57,54]
#> 56: (62,45,46), [38,59,57]
#> 57: (62,61,45), [55,56,72]
#> 58: (70,46,47), [42,67,59]
#> 59: (70,62,46), [56,58,78]
#> 60: (27,18,28), [74,69,50]
#> 61: (19,18,10), [62,76,74]
#> 62: (10,18,9), [52,63,61]
#> 63: (10,9,3), [47,77,62]
#> 64: (37,49,48), [73,53,65]
#> 65: (37,38,49), [80,64,66]
#> 66: (27,38,37), [65,48,69]
#> 67: (71,70,47), [58,86,93]
#> 68: (38,28,39), [81,79,69]
#> 69: (27,28,38), [68,66,60]
#> 70: (67,43,61), [54,71,91]
#> 71: (68,67,61), [70,72,87]
#> 72: (68,61,62), [57,75,71]
#> 73: (48,49,55), [85,83,64]
#> 74: (19,28,18), [60,61,82]
#> 75: (62,69,68), [88,72,78]
#> 76: (11,19,10), [61,77,84]
#> 77: (11,10,3), [63,0,76]
#> 78: (70,69,62), [75,59,95]
#> 79: (38,39,50), [98,80,68]
#> 80: (38,50,49), [85,65,79]
#> 81: (29,39,28), [68,82,100]
#> 82: (29,28,19), [74,94,81]
#> 83: (48,55,71), [102,86,73]
#> 84: (19,11,20), [121,94,76]
#> 85: (49,50,55), [97,73,80]
#> 86: (71,47,48), [51,83,67]
#> 87: (79,67,68), [71,89,92]
#> 88: (80,68,69), [75,90,89]
#> 89: (80,79,68), [87,88,110]
#> 90: (81,80,69), [88,95,116]
#> 91: (78,43,67), [70,92,114]
#> 92: (78,67,79), [87,109,91]
#> 93: (82,70,71), [67,118,101]
#> 94: (19,20,29), [103,82,84]
#> 95: (81,69,70), [78,101,90]
#> 96: (56,50,51), [98,105,97]
#> 97: (56,55,50), [85,96,104]
#> 98: (51,50,39), [79,99,96]
#> 99: (40,51,39), [98,100,125]
#> 100: (40,39,29), [81,107,99]
#> 101: (82,81,70), [95,93,122]
#> 102: (72,71,55), [83,104,119]
#> 103: (29,20,30), [108,107,94]
#> 104: (72,55,56), [97,130,102]
#> 105: (57,56,51), [96,120,123]
#> 106: (41,40,30), [107,142,133]
#> 107: (29,30,40), [106,100,103]
#> 108: (20,21,30), [124,103,121]
#> 109: (91,78,79), [92,111,115]
#> 110: (92,79,80), [89,112,111]
#> 111: (92,91,79), [109,110,135]
#> 112: (93,92,80), [110,116,136]
#> 113: (94,93,81), [116,122,146]
#> 114: (90,43,78), [91,115,144]
#> 115: (90,78,91), [109,139,114]
#> 116: (93,80,81), [90,113,112]
#> 117: (95,94,82), [122,129,153]
#> 118: (71,83,82), [129,93,119]
#> 119: (72,83,71), [118,102,127]
#> 120: (57,51,52), [125,132,105]
#> 121: (20,11,21), [0,108,84]
#> 122: (94,81,82), [101,117,113]
#> 123: (63,56,57), [105,154,128]
#> 124: (30,21,31), [0,142,108]
#> 125: (52,51,40), [99,133,120]
#> 126: (96,83,84), [127,161,148]
#> 127: (84,83,72), [119,150,126]
#> 128: (73,56,63), [123,151,130]
#> 129: (95,82,83), [118,148,117]
#> 130: (73,72,56), [104,128,141]
#> 131: (53,58,52), [132,155,185]
#> 132: (58,57,52), [120,131,152]
#> 133: (41,52,40), [125,106,155]
#> 134: (91,104,103), [167,139,135]
#> 135: (104,91,92), [111,137,134]
#> 136: (105,92,93), [112,138,137]
#> 137: (105,104,92), [135,136,168]
#> 138: (106,105,93), [136,146,169]
#> 139: (91,103,90), [145,115,134]
#> 140: (107,106,94), [146,153,173]
#> 141: (85,72,73), [130,175,150]
#> 142: (41,30,31), [124,157,106]
#> 143: (102,1,43), [34,144,0]
#> 144: (102,43,90), [114,145,143]
#> 145: (102,90,103), [139,166,144]
#> 146: (106,93,94), [113,140,138]
#> 147: (108,107,95), [153,165,182]
#> 148: (96,95,83), [129,126,165]
#> 149: (97,84,85), [150,162,161]
#> 150: (85,84,72), [127,141,149]
#> 151: (74,73,63), [128,156,158]
#> 152: (64,57,58), [132,176,154]
#> 153: (107,94,95), [117,147,140]
#> 154: (64,63,57), [123,152,156]
#> 155: (53,52,41), [133,164,131]
#> 156: (74,63,64), [154,159,151]
#> 157: (41,31,42), [0,164,142]
#> 158: (86,73,74), [151,187,175]
#> 159: (75,74,64), [156,181,184]
#> 160: (96,97,109), [195,179,161]
#> 161: (97,96,84), [126,149,160]
#> 162: (85,98,97), [180,149,174]
#> 163: (54,53,42), [164,0,188]
#> 164: (41,42,53), [163,155,157]
#> 165: (108,95,96), [148,179,147]
#> 166: (114,102,103), [145,167,0]
#> 167: (114,103,104), [134,170,166]
#> 168: (115,104,105), [137,171,170]
#> 169: (116,105,106), [138,177,171]
#> 170: (115,114,104), [167,168,190]
#> 171: (116,115,105), [168,169,191]
#> 172: (107,118,117), [202,173,182]
#> 173: (117,106,107), [140,172,177]
#> 174: (85,86,98), [199,162,175]
#> 175: (86,85,73), [141,158,174]
#> 176: (65,64,58), [152,186,181]
#> 177: (117,116,106), [169,173,192]
#> 178: (119,118,108), [182,183,198]
#> 179: (96,109,108), [183,165,160]
#> 180: (110,97,98), [162,209,195]
#> 181: (75,64,65), [176,206,159]
#> 182: (107,108,118), [178,172,147]
#> 183: (119,108,109), [179,204,178]
#> 184: (74,75,87), [200,187,159]
#> 185: (58,53,59), [188,186,131]
#> 186: (58,59,65), [207,176,185]
#> 187: (74,87,86), [189,158,184]
#> 188: (54,59,53), [185,163,212]
#> 189: (99,86,87), [187,213,199]
#> 190: (123,114,115), [170,191,0]
#> 191: (123,115,116), [171,193,190]
#> 192: (124,116,117), [177,197,193]
#> 193: (124,123,116), [191,192,0]
#> 194: (109,110,120), [218,204,195]
#> 195: (110,109,97), [160,180,194]
#> 196: (111,98,99), [199,220,209]
#> 197: (117,125,124), [0,192,202]
#> 198: (118,119,125), [203,202,178]
#> 199: (99,98,86), [174,189,196]
#> 200: (88,87,75), [184,210,211]
#> 201: (76,65,66), [207,215,206]
#> 202: (117,118,125), [198,197,172]
#> 203: (120,125,119), [198,204,217]
#> 204: (109,120,119), [203,183,194]
#> 205: (121,110,111), [209,226,218]
#> 206: (76,75,65), [181,201,210]
#> 207: (66,65,59), [186,208,201]
#> 208: (59,60,66), [223,207,212]
#> 209: (111,110,98), [180,196,205]
#> 210: (88,75,76), [206,222,200]
#> 211: (87,88,100), [221,213,200]
#> 212: (59,54,60), [0,208,188]
#> 213: (87,100,99), [216,189,211]
#> 214: (89,76,77), [215,0,222]
#> 215: (77,76,66), [201,223,214]
#> 216: (112,99,100), [213,228,220]
#> 217: (121,125,120), [203,218,225]
#> 218: (121,120,110), [194,205,217]
#> 219: (122,111,112), [220,229,226]
#> 220: (112,111,99), [196,216,219]
#> 221: (88,101,100), [227,211,224]
#> 222: (89,88,76), [210,214,224]
#> 223: (66,60,77), [0,215,208]
#> 224: (88,89,101), [0,221,222]
#> 225: (122,125,121), [217,226,0]
#> 226: (122,121,111), [205,219,225]
#> 227: (113,100,101), [221,0,228]
#> 228: (113,112,100), [216,227,229]
#> 229: (113,122,112), [219,228,0]
#> boundary nodes:  1 2 3 11 21 31 42 54 60 77 89 101 113 122 125 124 123 114 102 
#> 
#> $bin_counts
#>   [1] 0.078125 0.203125 0.171875 0.640625 0.593750 0.562500 0.609375 0.328125
#>   [9] 0.375000 0.296875 0.156250 0.203125 0.453125 0.531250 0.546875 0.500000
#>  [17] 0.593750 0.500000 0.468750 0.468750 0.203125 0.546875 0.390625 0.515625
#>  [25] 0.500000 0.515625 0.562500 0.484375 0.468750 0.531250 0.468750 0.375000
#>  [33] 0.484375 0.500000 0.578125 0.421875 0.593750 0.562500 0.640625 0.390625
#>  [41] 0.390625 0.140625 0.015625 0.546875 0.562500 0.562500 0.625000 0.609375
#>  [49] 0.562500 0.562500 0.750000 0.656250 0.515625 0.375000 0.343750 0.468750
#>  [57] 0.734375 0.531250 0.390625 0.234375 0.234375 0.171875 0.875000 0.671875
#>  [65] 0.421875 0.421875 0.640625 0.609375 0.562500 0.515625 0.109375 0.390625
#>  [73] 0.687500 0.640625 0.468750 0.437500 0.093750 0.484375 0.546875 0.437500
#>  [81] 0.437500 0.593750 0.296875 0.593750 0.562500 0.421875 0.562500 0.453125
#>  [89] 0.468750 0.328125 0.437500 0.484375 0.546875 0.468750 0.406250 0.734375
#>  [97] 0.609375 0.578125 0.593750 0.546875 0.203125 0.265625 0.609375 0.468750
#> [105] 0.578125 0.437500 0.687500 1.000000 0.734375 0.750000 0.406250 0.500000
#> [113] 0.203125 0.296875 0.562500 0.578125 0.578125 0.593750 0.390625 0.593750
#> [121] 0.578125 0.265625 0.015625 0.140625 0.015625
#>