Triangulate bin centroids

This function triangulates the bin centroids using the x and y coordinates provided in the input data frame and returns the triangular object.

Usage

tri_bin_centroids(centroids_data)

Arguments

centroids_data

The tibble containing the all the bin centroids.

Value

A triangular object representing the triangulated bin centroids.

Examples

all_centroids_df <- scurve_model_obj$hb_obj$centroids
counts_data <- scurve_model_obj$hb_obj$std_cts
umap_with_hb_id <- scurve_model_obj$hb_obj$data_hb_id
df_bin_centroids <- extract_hexbin_mean(data_hb = umap_with_hb_id,
counts_data = counts_data, centroids_data = all_centroids_df)
tri_bin_centroids(centroids_data = df_bin_centroids)
#> $trimesh_object
#> Delaunay triangulation, node and triangle indices:
#> triangle: nodes (a,b,c), neighbour triangles [i,j,k] 
#> 1: (1,2,8), [2,6,0]
#> 2: (9,8,2), [1,5,3]
#> 3: (18,8,9), [2,8,7]
#> 4: (9,3,10), [15,9,5]
#> 5: (3,9,2), [2,74,4]
#> 6: (17,1,8), [1,7,10]
#> 7: (17,8,18), [3,12,6]
#> 8: (9,19,18), [13,3,9]
#> 9: (9,10,19), [17,8,4]
#> 10: (27,1,17), [6,11,0]
#> 11: (28,27,17), [10,12,26]
#> 12: (28,17,18), [7,16,11]
#> 13: (18,19,29), [18,16,8]
#> 14: (4,5,11), [36,20,60]
#> 15: (4,10,3), [4,62,20]
#> 16: (18,29,28), [24,12,13]
#> 17: (19,10,20), [22,19,9]
#> 18: (19,30,29), [28,13,19]
#> 19: (19,20,30), [30,18,17]
#> 20: (4,11,10), [22,15,14]
#> 21: (20,11,21), [31,29,22]
#> 22: (11,20,10), [17,20,21]
#> 23: (28,39,38), [34,26,24]
#> 24: (39,28,29), [16,25,23]
#> 25: (29,40,39), [27,24,28]
#> 26: (28,38,27), [134,11,23]
#> 27: (39,40,48), [39,35,25]
#> 28: (29,30,40), [32,25,18]
#> 29: (20,21,31), [41,30,21]
#> 30: (20,31,30), [33,19,29]
#> 31: (11,12,21), [43,21,36]
#> 32: (30,41,40), [37,28,33]
#> 33: (30,31,41), [48,32,30]
#> 34: (39,47,38), [98,23,35]
#> 35: (39,48,47), [93,34,27]
#> 36: (11,5,12), [53,31,14]
#> 37: (40,41,49), [69,39,32]
#> 38: (77,48,49), [39,100,93]
#> 39: (40,49,48), [38,27,37]
#> 40: (33,31,32), [41,49,44]
#> 41: (21,32,31), [40,29,42]
#> 42: (21,22,32), [49,41,43]
#> 43: (22,21,12), [31,46,42]
#> 44: (42,31,33), [40,56,48]
#> 45: (23,22,13), [46,52,50]
#> 46: (13,22,12), [43,59,45]
#> 47: (6,14,13), [52,59,58]
#> 48: (42,41,31), [33,44,54]
#> 49: (22,33,32), [40,42,50]
#> 50: (23,33,22), [49,45,55]
#> 51: (24,23,14), [52,64,57]
#> 52: (14,23,13), [45,47,51]
#> 53: (12,5,6), [60,59,36]
#> 54: (41,42,54), [68,69,48]
#> 55: (33,23,34), [57,56,50]
#> 56: (33,34,42), [66,44,55]
#> 57: (24,34,23), [55,51,67]
#> 58: (14,6,15), [73,64,47]
#> 59: (6,13,12), [46,53,47]
#> 60: (6,5,4), [14,61,53]
#> 61: (7,6,4), [60,62,73]
#> 62: (7,4,3), [15,74,61]
#> 63: (25,35,24), [67,77,81]
#> 64: (14,15,24), [77,51,58]
#> 65: (42,43,50), [70,68,66]
#> 66: (43,42,34), [56,71,65]
#> 67: (34,24,35), [63,71,57]
#> 68: (42,50,54), [76,54,65]
#> 69: (41,54,49), [84,37,54]
#> 70: (51,50,43), [65,83,80]
#> 71: (34,35,43), [82,66,67]
#> 72: (16,15,7), [73,0,79]
#> 73: (7,15,6), [58,61,72]
#> 74: (3,2,7), [0,62,5]
#> 75: (64,54,59), [85,107,84]
#> 76: (50,55,54), [85,68,80]
#> 77: (25,24,15), [64,79,63]
#> 78: (26,25,16), [79,110,89]
#> 79: (16,25,15), [77,72,78]
#> 80: (51,55,50), [76,70,94]
#> 81: (36,35,25), [63,89,88]
#> 82: (43,35,44), [88,83,71]
#> 83: (43,44,51), [91,70,82]
#> 84: (64,49,54), [69,75,90]
#> 85: (59,54,55), [76,86,75]
#> 86: (60,59,55), [85,87,106]
#> 87: (60,55,56), [94,129,86]
#> 88: (36,44,35), [82,81,109]
#> 89: (26,36,25), [81,78,102]
#> 90: (69,49,64), [84,105,101]
#> 91: (52,51,44), [83,96,95]
#> 92: (47,77,76), [99,97,93]
#> 93: (77,47,48), [35,38,92]
#> 94: (56,55,51), [80,95,87]
#> 95: (52,56,51), [94,91,108]
#> 96: (45,52,44), [91,109,141]
#> 97: (47,76,75), [104,121,92]
#> 98: (47,85,38), [134,34,121]
#> 99: (76,77,88), [119,114,92]
#> 100: (78,77,49), [38,101,125]
#> 101: (69,78,49), [100,90,116]
#> 102: (36,26,37), [110,113,89]
#> 103: (85,75,86), [122,153,121]
#> 104: (76,87,75), [122,97,114]
#> 105: (70,69,64), [90,120,118]
#> 106: (65,59,60), [86,128,107]
#> 107: (65,64,59), [75,106,111]
#> 108: (57,56,52), [95,112,130]
#> 109: (45,44,36), [88,113,96]
#> 110: (37,26,16), [78,0,102]
#> 111: (64,65,71), [143,120,107]
#> 112: (52,53,57), [140,108,141]
#> 113: (36,37,45), [138,109,102]
#> 114: (76,88,87), [150,104,99]
#> 115: (78,90,89), [147,125,139]
#> 116: (79,78,69), [101,117,139]
#> 117: (80,79,69), [116,118,135]
#> 118: (80,69,70), [105,127,117]
#> 119: (77,89,88), [155,99,125]
#> 120: (64,71,70), [136,105,111]
#> 121: (47,75,85), [103,98,97]
#> 122: (86,75,87), [104,124,103]
#> 123: (99,98,87), [124,150,174]
#> 124: (98,86,87), [122,123,157]
#> 125: (77,78,89), [115,119,100]
#> 126: (91,90,79), [139,135,164]
#> 127: (70,81,80), [148,118,136]
#> 128: (66,65,60), [106,133,132]
#> 129: (56,61,60), [133,87,130]
#> 130: (56,57,61), [158,129,108]
#> 131: (53,45,46), [138,159,141]
#> 132: (72,65,66), [128,167,143]
#> 133: (61,66,60), [128,129,144]
#> 134: (85,27,38), [26,98,152]
#> 135: (80,91,79), [126,117,161]
#> 136: (70,71,81), [137,127,120]
#> 137: (71,82,81), [170,136,142]
#> 138: (45,37,46), [0,131,113]
#> 139: (78,79,90), [126,115,116]
#> 140: (58,57,53), [112,159,145]
#> 141: (52,45,53), [131,112,96]
#> 142: (71,72,82), [172,137,143]
#> 143: (72,71,65), [111,132,142]
#> 144: (67,66,61), [133,168,163]
#> 145: (62,57,58), [140,186,158]
#> 146: (90,102,101), [190,147,164]
#> 147: (101,89,90), [115,146,160]
#> 148: (80,81,92), [162,161,127]
#> 149: (97,96,85), [152,153,188]
#> 150: (99,87,88), [114,151,123]
#> 151: (100,99,88), [150,155,156]
#> 152: (96,27,85), [134,149,0]
#> 153: (97,85,86), [103,157,149]
#> 154: (99,110,109), [207,174,156]
#> 155: (100,88,89), [119,160,151]
#> 156: (110,99,100), [151,175,154]
#> 157: (97,86,98), [124,179,153]
#> 158: (62,61,57), [130,145,168]
#> 159: (58,53,46), [131,187,140]
#> 160: (101,100,89), [155,147,176]
#> 161: (80,92,91), [166,135,148]
#> 162: (81,93,92), [195,148,170]
#> 163: (67,73,66), [167,144,191]
#> 164: (90,91,102), [178,146,126]
#> 165: (101,112,111), [189,176,190]
#> 166: (91,92,103), [169,178,161]
#> 167: (73,72,66), [132,163,171]
#> 168: (62,67,61), [144,158,184]
#> 169: (104,103,92), [166,195,194]
#> 170: (81,82,93), [182,162,137]
#> 171: (83,72,73), [167,197,172]
#> 172: (83,82,72), [142,171,183]
#> 173: (62,63,68), [205,184,186]
#> 174: (99,109,98), [180,123,154]
#> 175: (100,111,110), [218,156,176]
#> 176: (100,101,111), [165,175,160]
#> 177: (113,112,102), [190,199,181]
#> 178: (91,103,102), [199,164,166]
#> 179: (108,97,98), [157,180,192]
#> 180: (108,98,109), [174,201,179]
#> 181: (123,112,113), [177,208,222]
#> 182: (82,94,93), [210,170,183]
#> 183: (82,83,94), [215,182,172]
#> 184: (62,68,67), [185,168,173]
#> 185: (67,68,74), [204,191,184]
#> 186: (63,62,58), [145,187,173]
#> 187: (63,58,46), [159,0,186]
#> 188: (97,107,96), [212,149,192]
#> 189: (112,122,111), [198,165,222]
#> 190: (101,102,112), [177,165,146]
#> 191: (67,74,73), [193,163,185]
#> 192: (97,108,107), [213,188,179]
#> 193: (73,74,84), [204,197,191]
#> 194: (114,103,104), [169,216,209]
#> 195: (104,92,93), [162,196,169]
#> 196: (93,105,104), [228,195,210]
#> 197: (73,84,83), [203,171,193]
#> 198: (121,111,122), [189,233,218]
#> 199: (113,102,103), [178,209,177]
#> 200: (108,119,118), [221,213,201]
#> 201: (119,108,109), [180,202,200]
#> 202: (109,120,119), [219,201,207]
#> 203: (95,83,84), [197,227,215]
#> 204: (84,74,68), [185,205,193]
#> 205: (84,68,63), [173,0,204]
#> 206: (120,110,121), [218,234,207]
#> 207: (109,110,120), [206,202,154]
#> 208: (113,124,123), [242,181,223]
#> 209: (114,113,103), [199,194,223]
#> 210: (93,94,105), [226,196,182]
#> 211: (106,94,95), [215,227,226]
#> 212: (117,96,107), [188,214,0]
#> 213: (107,108,118), [200,214,192]
#> 214: (107,118,117), [230,212,213]
#> 215: (95,94,83), [183,203,211]
#> 216: (115,114,104), [194,228,224]
#> 217: (129,128,120), [219,234,243]
#> 218: (121,110,111), [175,198,206]
#> 219: (128,119,120), [202,217,231]
#> 220: (130,129,121), [234,233,238]
#> 221: (119,127,118), [240,200,246]
#> 222: (123,122,112), [189,181,239]
#> 223: (113,114,124), [236,208,209]
#> 224: (125,114,115), [216,237,236]
#> 225: (116,105,106), [226,0,229]
#> 226: (106,105,94), [210,211,225]
#> 227: (106,95,84), [203,0,211]
#> 228: (115,104,105), [196,229,216]
#> 229: (116,115,105), [228,225,237]
#> 230: (126,117,118), [214,240,0]
#> 231: (119,128,134), [249,246,219]
#> 232: (137,130,122), [233,251,244]
#> 233: (130,121,122), [198,232,220]
#> 234: (129,120,121), [206,220,217]
#> 235: (132,124,125), [236,0,242]
#> 236: (125,124,114), [223,224,235]
#> 237: (125,115,116), [229,0,224]
#> 238: (136,129,130), [220,244,241]
#> 239: (131,122,123), [222,248,251]
#> 240: (118,127,126), [252,230,221]
#> 241: (135,129,136), [238,250,243]
#> 242: (123,124,132), [235,248,208]
#> 243: (135,128,129), [217,241,249]
#> 244: (137,136,130), [238,232,250]
#> 245: (127,134,133), [253,252,246]
#> 246: (119,134,127), [245,221,231]
#> 247: (131,132,137), [0,251,248]
#> 248: (123,132,131), [247,239,242]
#> 249: (135,134,128), [231,243,253]
#> 250: (137,135,136), [241,244,0]
#> 251: (137,122,131), [239,247,232]
#> 252: (127,133,126), [0,240,245]
#> 253: (133,134,135), [249,0,245]
#> boundary nodes:  1 2 7 16 37 46 63 84 106 116 125 132 137 135 133 126 117 96 27 
#> 
#> $n_h
#>   [1]  1 10  8  8  2  2 12 11  9  4 11 12  1  6 12 12  5 10 10 11  5 14 11  9  6
#>  [26]  6  1 10  5  8  8  1  9  5 11  7  9 11  5 10  8 10 12  2 13  1  9  3  2  6
#>  [51]  9 10  7  5 12  6  6  9  9  9  6  6  1 10  9  5  5  8  1 10  9  6 11  5  3
#>  [76]  5  3 11  3 12  6  5  8  9 12  9 10 11  9 12  7  8 11  5  7  4  8  5  8  8
#> [101]  4  9  7 10  7 15 12  6  5  5  7 14  3  3  8  7  4  9  9  5  8  8  7  5  4
#> [126]  4 13 14  6  7 10 10  3  1  2  1  1
#>